feat: remove support for xhr
- add response ok handler to fetch

BREAKING CHANGE
