+++ package.json
@@ -29,10 +29,6 @@
     "dev": "snowdev dev",
     "release": "snowdev release"
   },
-  "devDependencies": {
-    "@thi.ng/hdom": "^9.1.5",
-    "es-module-shims": "^1.4.7"
-  },
   "engines": {
     "node": ">=16.0.0",
     "npm": ">=7.0.0"
