--- web_modules/@thi.ng/hdom.js
@@ -1,5700 +1,0 @@
-var commonjsGlobal = typeof globalThis !== 'undefined' ? globalThis : typeof window !== 'undefined' ? window : typeof global !== 'undefined' ? global : typeof self !== 'undefined' ? self : {};
-
-function createCommonjsModule(fn, basedir, module) {
-	return module = {
-		path: basedir,
-		exports: {},
-		require: function (path, base) {
-			return commonjsRequire(path, (base === undefined || base === null) ? module.path : base);
-		}
-	}, fn(module, module.exports), module.exports;
-}
-
-function commonjsRequire () {
-	throw new Error('Dynamic requires are not currently supported by @rollup/plugin-commonjs');
-}
-
-var check = function (it) {
-  return it && it.Math == Math && it;
-};
-
-// https://github.com/zloirock/core-js/issues/86#issuecomment-115759028
-var global_1 =
-  // eslint-disable-next-line es/no-global-this -- safe
-  check(typeof globalThis == 'object' && globalThis) ||
-  check(typeof window == 'object' && window) ||
-  // eslint-disable-next-line no-restricted-globals -- safe
-  check(typeof self == 'object' && self) ||
-  check(typeof commonjsGlobal == 'object' && commonjsGlobal) ||
-  // eslint-disable-next-line no-new-func -- fallback
-  (function () { return this; })() || Function('return this')();
-
-// iterable DOM collections
-// flag - `iterable` interface - 'entries', 'keys', 'values', 'forEach' methods
-var domIterables = {
-  CSSRuleList: 0,
-  CSSStyleDeclaration: 0,
-  CSSValueList: 0,
-  ClientRectList: 0,
-  DOMRectList: 0,
-  DOMStringList: 0,
-  DOMTokenList: 1,
-  DataTransferItemList: 0,
-  FileList: 0,
-  HTMLAllCollection: 0,
-  HTMLCollection: 0,
-  HTMLFormElement: 0,
-  HTMLSelectElement: 0,
-  MediaList: 0,
-  MimeTypeArray: 0,
-  NamedNodeMap: 0,
-  NodeList: 1,
-  PaintRequestList: 0,
-  Plugin: 0,
-  PluginArray: 0,
-  SVGLengthList: 0,
-  SVGNumberList: 0,
-  SVGPathSegList: 0,
-  SVGPointList: 0,
-  SVGStringList: 0,
-  SVGTransformList: 0,
-  SourceBufferList: 0,
-  StyleSheetList: 0,
-  TextTrackCueList: 0,
-  TextTrackList: 0,
-  TouchList: 0
-};
-
-// `IsCallable` abstract operation
-// https://tc39.es/ecma262/#sec-iscallable
-var isCallable = function (argument) {
-  return typeof argument == 'function';
-};
-
-var isObject = function (it) {
-  return typeof it == 'object' ? it !== null : isCallable(it);
-};
-
-var document$1 = global_1.document;
-// typeof document.createElement is 'object' in old IE
-var EXISTS = isObject(document$1) && isObject(document$1.createElement);
-
-var documentCreateElement = function (it) {
-  return EXISTS ? document$1.createElement(it) : {};
-};
-
-// in old WebKit versions, `element.classList` is not an instance of global `DOMTokenList`
-
-
-var classList = documentCreateElement('span').classList;
-var DOMTokenListPrototype = classList && classList.constructor && classList.constructor.prototype;
-
-var domTokenListPrototype = DOMTokenListPrototype === Object.prototype ? undefined : DOMTokenListPrototype;
-
-var fails = function (exec) {
-  try {
-    return !!exec();
-  } catch (error) {
-    return true;
-  }
-};
-
-var functionBindNative = !fails(function () {
-  var test = (function () { /* empty */ }).bind();
-  // eslint-disable-next-line no-prototype-builtins -- safe
-  return typeof test != 'function' || test.hasOwnProperty('prototype');
-});
-
-var FunctionPrototype = Function.prototype;
-var bind = FunctionPrototype.bind;
-var call = FunctionPrototype.call;
-var uncurryThis = functionBindNative && bind.bind(call, call);
-
-var functionUncurryThis = functionBindNative ? function (fn) {
-  return fn && uncurryThis(fn);
-} : function (fn) {
-  return fn && function () {
-    return call.apply(fn, arguments);
-  };
-};
-
-var toString = functionUncurryThis({}.toString);
-var stringSlice = functionUncurryThis(''.slice);
-
-var classofRaw = function (it) {
-  return stringSlice(toString(it), 8, -1);
-};
-
-var Object$1 = global_1.Object;
-var split = functionUncurryThis(''.split);
-
-// fallback for non-array-like ES3 and non-enumerable old V8 strings
-var indexedObject = fails(function () {
-  // throws an error in rhino, see https://github.com/mozilla/rhino/issues/346
-  // eslint-disable-next-line no-prototype-builtins -- safe
-  return !Object$1('z').propertyIsEnumerable(0);
-}) ? function (it) {
-  return classofRaw(it) == 'String' ? split(it, '') : Object$1(it);
-} : Object$1;
-
-var TypeError$1 = global_1.TypeError;
-
-// `RequireObjectCoercible` abstract operation
-// https://tc39.es/ecma262/#sec-requireobjectcoercible
-var requireObjectCoercible = function (it) {
-  if (it == undefined) throw TypeError$1("Can't call method on " + it);
-  return it;
-};
-
-// toObject with fallback for non-array-like ES3 strings
-
-
-
-var toIndexedObject = function (it) {
-  return indexedObject(requireObjectCoercible(it));
-};
-
-// eslint-disable-next-line es/no-object-defineproperty -- safe
-var defineProperty = Object.defineProperty;
-
-var setGlobal = function (key, value) {
-  try {
-    defineProperty(global_1, key, { value: value, configurable: true, writable: true });
-  } catch (error) {
-    global_1[key] = value;
-  } return value;
-};
-
-var SHARED = '__core-js_shared__';
-var store = global_1[SHARED] || setGlobal(SHARED, {});
-
-var sharedStore = store;
-
-var shared = createCommonjsModule(function (module) {
-(module.exports = function (key, value) {
-  return sharedStore[key] || (sharedStore[key] = value !== undefined ? value : {});
-})('versions', []).push({
-  version: '3.21.1',
-  mode:  'global',
-  copyright: '© 2014-2022 Denis Pushkarev (zloirock.ru)',
-  license: 'https://github.com/zloirock/core-js/blob/v3.21.1/LICENSE',
-  source: 'https://github.com/zloirock/core-js'
-});
-});
-
-var Object$2 = global_1.Object;
-
-// `ToObject` abstract operation
-// https://tc39.es/ecma262/#sec-toobject
-var toObject = function (argument) {
-  return Object$2(requireObjectCoercible(argument));
-};
-
-var hasOwnProperty = functionUncurryThis({}.hasOwnProperty);
-
-// `HasOwnProperty` abstract operation
-// https://tc39.es/ecma262/#sec-hasownproperty
-var hasOwnProperty_1 = Object.hasOwn || function hasOwn(it, key) {
-  return hasOwnProperty(toObject(it), key);
-};
-
-var id = 0;
-var postfix = Math.random();
-var toString$1 = functionUncurryThis(1.0.toString);
-
-var uid = function (key) {
-  return 'Symbol(' + (key === undefined ? '' : key) + ')_' + toString$1(++id + postfix, 36);
-};
-
-var aFunction = function (argument) {
-  return isCallable(argument) ? argument : undefined;
-};
-
-var getBuiltIn = function (namespace, method) {
-  return arguments.length < 2 ? aFunction(global_1[namespace]) : global_1[namespace] && global_1[namespace][method];
-};
-
-var engineUserAgent = getBuiltIn('navigator', 'userAgent') || '';
-
-var process = global_1.process;
-var Deno = global_1.Deno;
-var versions = process && process.versions || Deno && Deno.version;
-var v8 = versions && versions.v8;
-var match, version;
-
-if (v8) {
-  match = v8.split('.');
-  // in old Chrome, versions of V8 isn't V8 = Chrome / 10
-  // but their correct versions are not interesting for us
-  version = match[0] > 0 && match[0] < 4 ? 1 : +(match[0] + match[1]);
-}
-
-// BrowserFS NodeJS `process` polyfill incorrectly set `.v8` to `0.0`
-// so check `userAgent` even if `.v8` exists, but 0
-if (!version && engineUserAgent) {
-  match = engineUserAgent.match(/Edge\/(\d+)/);
-  if (!match || match[1] >= 74) {
-    match = engineUserAgent.match(/Chrome\/(\d+)/);
-    if (match) version = +match[1];
-  }
-}
-
-var engineV8Version = version;
-
-/* eslint-disable es/no-symbol -- required for testing */
-
-
-
-// eslint-disable-next-line es/no-object-getownpropertysymbols -- required for testing
-var nativeSymbol = !!Object.getOwnPropertySymbols && !fails(function () {
-  var symbol = Symbol();
-  // Chrome 38 Symbol has incorrect toString conversion
-  // `get-own-property-symbols` polyfill symbols converted to object are not Symbol instances
-  return !String(symbol) || !(Object(symbol) instanceof Symbol) ||
-    // Chrome 38-40 symbols are not inherited from DOM collections prototypes to instances
-    !Symbol.sham && engineV8Version && engineV8Version < 41;
-});
-
-/* eslint-disable es/no-symbol -- required for testing */
-
-
-var useSymbolAsUid = nativeSymbol
-  && !Symbol.sham
-  && typeof Symbol.iterator == 'symbol';
-
-var WellKnownSymbolsStore = shared('wks');
-var Symbol$1 = global_1.Symbol;
-var symbolFor = Symbol$1 && Symbol$1['for'];
-var createWellKnownSymbol = useSymbolAsUid ? Symbol$1 : Symbol$1 && Symbol$1.withoutSetter || uid;
-
-var wellKnownSymbol = function (name) {
-  if (!hasOwnProperty_1(WellKnownSymbolsStore, name) || !(nativeSymbol || typeof WellKnownSymbolsStore[name] == 'string')) {
-    var description = 'Symbol.' + name;
-    if (nativeSymbol && hasOwnProperty_1(Symbol$1, name)) {
-      WellKnownSymbolsStore[name] = Symbol$1[name];
-    } else if (useSymbolAsUid && symbolFor) {
-      WellKnownSymbolsStore[name] = symbolFor(description);
-    } else {
-      WellKnownSymbolsStore[name] = createWellKnownSymbol(description);
-    }
-  } return WellKnownSymbolsStore[name];
-};
-
-var String$1 = global_1.String;
-var TypeError$2 = global_1.TypeError;
-
-// `Assert: Type(argument) is Object`
-var anObject = function (argument) {
-  if (isObject(argument)) return argument;
-  throw TypeError$2(String$1(argument) + ' is not an object');
-};
-
-// Detect IE8's incomplete defineProperty implementation
-var descriptors = !fails(function () {
-  // eslint-disable-next-line es/no-object-defineproperty -- required for testing
-  return Object.defineProperty({}, 1, { get: function () { return 7; } })[1] != 7;
-});
-
-// V8 ~ Chrome 36-
-// https://bugs.chromium.org/p/v8/issues/detail?id=3334
-var v8PrototypeDefineBug = descriptors && fails(function () {
-  // eslint-disable-next-line es/no-object-defineproperty -- required for testing
-  return Object.defineProperty(function () { /* empty */ }, 'prototype', {
-    value: 42,
-    writable: false
-  }).prototype != 42;
-});
-
-// Thanks to IE8 for its funny defineProperty
-var ie8DomDefine = !descriptors && !fails(function () {
-  // eslint-disable-next-line es/no-object-defineproperty -- required for testing
-  return Object.defineProperty(documentCreateElement('div'), 'a', {
-    get: function () { return 7; }
-  }).a != 7;
-});
-
-var call$1 = Function.prototype.call;
-
-var functionCall = functionBindNative ? call$1.bind(call$1) : function () {
-  return call$1.apply(call$1, arguments);
-};
-
-var objectIsPrototypeOf = functionUncurryThis({}.isPrototypeOf);
-
-var Object$3 = global_1.Object;
-
-var isSymbol = useSymbolAsUid ? function (it) {
-  return typeof it == 'symbol';
-} : function (it) {
-  var $Symbol = getBuiltIn('Symbol');
-  return isCallable($Symbol) && objectIsPrototypeOf($Symbol.prototype, Object$3(it));
-};
-
-var String$2 = global_1.String;
-
-var tryToString = function (argument) {
-  try {
-    return String$2(argument);
-  } catch (error) {
-    return 'Object';
-  }
-};
-
-var TypeError$3 = global_1.TypeError;
-
-// `Assert: IsCallable(argument) is true`
-var aCallable = function (argument) {
-  if (isCallable(argument)) return argument;
-  throw TypeError$3(tryToString(argument) + ' is not a function');
-};
-
-// `GetMethod` abstract operation
-// https://tc39.es/ecma262/#sec-getmethod
-var getMethod = function (V, P) {
-  var func = V[P];
-  return func == null ? undefined : aCallable(func);
-};
-
-var TypeError$4 = global_1.TypeError;
-
-// `OrdinaryToPrimitive` abstract operation
-// https://tc39.es/ecma262/#sec-ordinarytoprimitive
-var ordinaryToPrimitive = function (input, pref) {
-  var fn, val;
-  if (pref === 'string' && isCallable(fn = input.toString) && !isObject(val = functionCall(fn, input))) return val;
-  if (isCallable(fn = input.valueOf) && !isObject(val = functionCall(fn, input))) return val;
-  if (pref !== 'string' && isCallable(fn = input.toString) && !isObject(val = functionCall(fn, input))) return val;
-  throw TypeError$4("Can't convert object to primitive value");
-};
-
-var TypeError$5 = global_1.TypeError;
-var TO_PRIMITIVE = wellKnownSymbol('toPrimitive');
-
-// `ToPrimitive` abstract operation
-// https://tc39.es/ecma262/#sec-toprimitive
-var toPrimitive = function (input, pref) {
-  if (!isObject(input) || isSymbol(input)) return input;
-  var exoticToPrim = getMethod(input, TO_PRIMITIVE);
-  var result;
-  if (exoticToPrim) {
-    if (pref === undefined) pref = 'default';
-    result = functionCall(exoticToPrim, input, pref);
-    if (!isObject(result) || isSymbol(result)) return result;
-    throw TypeError$5("Can't convert object to primitive value");
-  }
-  if (pref === undefined) pref = 'number';
-  return ordinaryToPrimitive(input, pref);
-};
-
-// `ToPropertyKey` abstract operation
-// https://tc39.es/ecma262/#sec-topropertykey
-var toPropertyKey = function (argument) {
-  var key = toPrimitive(argument, 'string');
-  return isSymbol(key) ? key : key + '';
-};
-
-var TypeError$6 = global_1.TypeError;
-// eslint-disable-next-line es/no-object-defineproperty -- safe
-var $defineProperty = Object.defineProperty;
-// eslint-disable-next-line es/no-object-getownpropertydescriptor -- safe
-var $getOwnPropertyDescriptor = Object.getOwnPropertyDescriptor;
-var ENUMERABLE = 'enumerable';
-var CONFIGURABLE = 'configurable';
-var WRITABLE = 'writable';
-
-// `Object.defineProperty` method
-// https://tc39.es/ecma262/#sec-object.defineproperty
-var f = descriptors ? v8PrototypeDefineBug ? function defineProperty(O, P, Attributes) {
-  anObject(O);
-  P = toPropertyKey(P);
-  anObject(Attributes);
-  if (typeof O === 'function' && P === 'prototype' && 'value' in Attributes && WRITABLE in Attributes && !Attributes[WRITABLE]) {
-    var current = $getOwnPropertyDescriptor(O, P);
-    if (current && current[WRITABLE]) {
-      O[P] = Attributes.value;
-      Attributes = {
-        configurable: CONFIGURABLE in Attributes ? Attributes[CONFIGURABLE] : current[CONFIGURABLE],
-        enumerable: ENUMERABLE in Attributes ? Attributes[ENUMERABLE] : current[ENUMERABLE],
-        writable: false
-      };
-    }
-  } return $defineProperty(O, P, Attributes);
-} : $defineProperty : function defineProperty(O, P, Attributes) {
-  anObject(O);
-  P = toPropertyKey(P);
-  anObject(Attributes);
-  if (ie8DomDefine) try {
-    return $defineProperty(O, P, Attributes);
-  } catch (error) { /* empty */ }
-  if ('get' in Attributes || 'set' in Attributes) throw TypeError$6('Accessors not supported');
-  if ('value' in Attributes) O[P] = Attributes.value;
-  return O;
-};
-
-var objectDefineProperty = {
-	f: f
-};
-
-var ceil = Math.ceil;
-var floor = Math.floor;
-
-// `ToIntegerOrInfinity` abstract operation
-// https://tc39.es/ecma262/#sec-tointegerorinfinity
-var toIntegerOrInfinity = function (argument) {
-  var number = +argument;
-  // eslint-disable-next-line no-self-compare -- safe
-  return number !== number || number === 0 ? 0 : (number > 0 ? floor : ceil)(number);
-};
-
-var max = Math.max;
-var min = Math.min;
-
-// Helper for a popular repeating case of the spec:
-// Let integer be ? ToInteger(index).
-// If integer < 0, let result be max((length + integer), 0); else let result be min(integer, length).
-var toAbsoluteIndex = function (index, length) {
-  var integer = toIntegerOrInfinity(index);
-  return integer < 0 ? max(integer + length, 0) : min(integer, length);
-};
-
-var min$1 = Math.min;
-
-// `ToLength` abstract operation
-// https://tc39.es/ecma262/#sec-tolength
-var toLength = function (argument) {
-  return argument > 0 ? min$1(toIntegerOrInfinity(argument), 0x1FFFFFFFFFFFFF) : 0; // 2 ** 53 - 1 == 9007199254740991
-};
-
-// `LengthOfArrayLike` abstract operation
-// https://tc39.es/ecma262/#sec-lengthofarraylike
-var lengthOfArrayLike = function (obj) {
-  return toLength(obj.length);
-};
-
-// `Array.prototype.{ indexOf, includes }` methods implementation
-var createMethod = function (IS_INCLUDES) {
-  return function ($this, el, fromIndex) {
-    var O = toIndexedObject($this);
-    var length = lengthOfArrayLike(O);
-    var index = toAbsoluteIndex(fromIndex, length);
-    var value;
-    // Array#includes uses SameValueZero equality algorithm
-    // eslint-disable-next-line no-self-compare -- NaN check
-    if (IS_INCLUDES && el != el) while (length > index) {
-      value = O[index++];
-      // eslint-disable-next-line no-self-compare -- NaN check
-      if (value != value) return true;
-    // Array#indexOf ignores holes, Array#includes - not
-    } else for (;length > index; index++) {
-      if ((IS_INCLUDES || index in O) && O[index] === el) return IS_INCLUDES || index || 0;
-    } return !IS_INCLUDES && -1;
-  };
-};
-
-var arrayIncludes = {
-  // `Array.prototype.includes` method
-  // https://tc39.es/ecma262/#sec-array.prototype.includes
-  includes: createMethod(true),
-  // `Array.prototype.indexOf` method
-  // https://tc39.es/ecma262/#sec-array.prototype.indexof
-  indexOf: createMethod(false)
-};
-
-var hiddenKeys = {};
-
-var indexOf = arrayIncludes.indexOf;
-
-
-var push = functionUncurryThis([].push);
-
-var objectKeysInternal = function (object, names) {
-  var O = toIndexedObject(object);
-  var i = 0;
-  var result = [];
-  var key;
-  for (key in O) !hasOwnProperty_1(hiddenKeys, key) && hasOwnProperty_1(O, key) && push(result, key);
-  // Don't enum bug & hidden keys
-  while (names.length > i) if (hasOwnProperty_1(O, key = names[i++])) {
-    ~indexOf(result, key) || push(result, key);
-  }
-  return result;
-};
-
-// IE8- don't enum bug keys
-var enumBugKeys = [
-  'constructor',
-  'hasOwnProperty',
-  'isPrototypeOf',
-  'propertyIsEnumerable',
-  'toLocaleString',
-  'toString',
-  'valueOf'
-];
-
-// `Object.keys` method
-// https://tc39.es/ecma262/#sec-object.keys
-// eslint-disable-next-line es/no-object-keys -- safe
-var objectKeys = Object.keys || function keys(O) {
-  return objectKeysInternal(O, enumBugKeys);
-};
-
-// `Object.defineProperties` method
-// https://tc39.es/ecma262/#sec-object.defineproperties
-// eslint-disable-next-line es/no-object-defineproperties -- safe
-var f$1 = descriptors && !v8PrototypeDefineBug ? Object.defineProperties : function defineProperties(O, Properties) {
-  anObject(O);
-  var props = toIndexedObject(Properties);
-  var keys = objectKeys(Properties);
-  var length = keys.length;
-  var index = 0;
-  var key;
-  while (length > index) objectDefineProperty.f(O, key = keys[index++], props[key]);
-  return O;
-};
-
-var objectDefineProperties = {
-	f: f$1
-};
-
-var html = getBuiltIn('document', 'documentElement');
-
-var keys = shared('keys');
-
-var sharedKey = function (key) {
-  return keys[key] || (keys[key] = uid(key));
-};
-
-/* global ActiveXObject -- old IE, WSH */
-
-
-
-
-
-
-
-
-var GT = '>';
-var LT = '<';
-var PROTOTYPE = 'prototype';
-var SCRIPT = 'script';
-var IE_PROTO = sharedKey('IE_PROTO');
-
-var EmptyConstructor = function () { /* empty */ };
-
-var scriptTag = function (content) {
-  return LT + SCRIPT + GT + content + LT + '/' + SCRIPT + GT;
-};
-
-// Create object with fake `null` prototype: use ActiveX Object with cleared prototype
-var NullProtoObjectViaActiveX = function (activeXDocument) {
-  activeXDocument.write(scriptTag(''));
-  activeXDocument.close();
-  var temp = activeXDocument.parentWindow.Object;
-  activeXDocument = null; // avoid memory leak
-  return temp;
-};
-
-// Create object with fake `null` prototype: use iframe Object with cleared prototype
-var NullProtoObjectViaIFrame = function () {
-  // Thrash, waste and sodomy: IE GC bug
-  var iframe = documentCreateElement('iframe');
-  var JS = 'java' + SCRIPT + ':';
-  var iframeDocument;
-  iframe.style.display = 'none';
-  html.appendChild(iframe);
-  // https://github.com/zloirock/core-js/issues/475
-  iframe.src = String(JS);
-  iframeDocument = iframe.contentWindow.document;
-  iframeDocument.open();
-  iframeDocument.write(scriptTag('document.F=Object'));
-  iframeDocument.close();
-  return iframeDocument.F;
-};
-
-// Check for document.domain and active x support
-// No need to use active x approach when document.domain is not set
-// see https://github.com/es-shims/es5-shim/issues/150
-// variation of https://github.com/kitcambridge/es5-shim/commit/4f738ac066346
-// avoid IE GC bug
-var activeXDocument;
-var NullProtoObject = function () {
-  try {
-    activeXDocument = new ActiveXObject('htmlfile');
-  } catch (error) { /* ignore */ }
-  NullProtoObject = typeof document != 'undefined'
-    ? document.domain && activeXDocument
-      ? NullProtoObjectViaActiveX(activeXDocument) // old IE
-      : NullProtoObjectViaIFrame()
-    : NullProtoObjectViaActiveX(activeXDocument); // WSH
-  var length = enumBugKeys.length;
-  while (length--) delete NullProtoObject[PROTOTYPE][enumBugKeys[length]];
-  return NullProtoObject();
-};
-
-hiddenKeys[IE_PROTO] = true;
-
-// `Object.create` method
-// https://tc39.es/ecma262/#sec-object.create
-var objectCreate = Object.create || function create(O, Properties) {
-  var result;
-  if (O !== null) {
-    EmptyConstructor[PROTOTYPE] = anObject(O);
-    result = new EmptyConstructor();
-    EmptyConstructor[PROTOTYPE] = null;
-    // add "__proto__" for Object.getPrototypeOf polyfill
-    result[IE_PROTO] = O;
-  } else result = NullProtoObject();
-  return Properties === undefined ? result : objectDefineProperties.f(result, Properties);
-};
-
-var UNSCOPABLES = wellKnownSymbol('unscopables');
-var ArrayPrototype = Array.prototype;
-
-// Array.prototype[@@unscopables]
-// https://tc39.es/ecma262/#sec-array.prototype-@@unscopables
-if (ArrayPrototype[UNSCOPABLES] == undefined) {
-  objectDefineProperty.f(ArrayPrototype, UNSCOPABLES, {
-    configurable: true,
-    value: objectCreate(null)
-  });
-}
-
-// add a key to Array.prototype[@@unscopables]
-var addToUnscopables = function (key) {
-  ArrayPrototype[UNSCOPABLES][key] = true;
-};
-
-var iterators = {};
-
-var functionToString = functionUncurryThis(Function.toString);
-
-// this helper broken in `core-js@3.4.1-3.4.4`, so we can't use `shared` helper
-if (!isCallable(sharedStore.inspectSource)) {
-  sharedStore.inspectSource = function (it) {
-    return functionToString(it);
-  };
-}
-
-var inspectSource = sharedStore.inspectSource;
-
-var WeakMap = global_1.WeakMap;
-
-var nativeWeakMap = isCallable(WeakMap) && /native code/.test(inspectSource(WeakMap));
-
-var createPropertyDescriptor = function (bitmap, value) {
-  return {
-    enumerable: !(bitmap & 1),
-    configurable: !(bitmap & 2),
-    writable: !(bitmap & 4),
-    value: value
-  };
-};
-
-var createNonEnumerableProperty = descriptors ? function (object, key, value) {
-  return objectDefineProperty.f(object, key, createPropertyDescriptor(1, value));
-} : function (object, key, value) {
-  object[key] = value;
-  return object;
-};
-
-var OBJECT_ALREADY_INITIALIZED = 'Object already initialized';
-var TypeError$7 = global_1.TypeError;
-var WeakMap$1 = global_1.WeakMap;
-var set, get, has;
-
-var enforce = function (it) {
-  return has(it) ? get(it) : set(it, {});
-};
-
-var getterFor = function (TYPE) {
-  return function (it) {
-    var state;
-    if (!isObject(it) || (state = get(it)).type !== TYPE) {
-      throw TypeError$7('Incompatible receiver, ' + TYPE + ' required');
-    } return state;
-  };
-};
-
-if (nativeWeakMap || sharedStore.state) {
-  var store$1 = sharedStore.state || (sharedStore.state = new WeakMap$1());
-  var wmget = functionUncurryThis(store$1.get);
-  var wmhas = functionUncurryThis(store$1.has);
-  var wmset = functionUncurryThis(store$1.set);
-  set = function (it, metadata) {
-    if (wmhas(store$1, it)) throw new TypeError$7(OBJECT_ALREADY_INITIALIZED);
-    metadata.facade = it;
-    wmset(store$1, it, metadata);
-    return metadata;
-  };
-  get = function (it) {
-    return wmget(store$1, it) || {};
-  };
-  has = function (it) {
-    return wmhas(store$1, it);
-  };
-} else {
-  var STATE = sharedKey('state');
-  hiddenKeys[STATE] = true;
-  set = function (it, metadata) {
-    if (hasOwnProperty_1(it, STATE)) throw new TypeError$7(OBJECT_ALREADY_INITIALIZED);
-    metadata.facade = it;
-    createNonEnumerableProperty(it, STATE, metadata);
-    return metadata;
-  };
-  get = function (it) {
-    return hasOwnProperty_1(it, STATE) ? it[STATE] : {};
-  };
-  has = function (it) {
-    return hasOwnProperty_1(it, STATE);
-  };
-}
-
-var internalState = {
-  set: set,
-  get: get,
-  has: has,
-  enforce: enforce,
-  getterFor: getterFor
-};
-
-var $propertyIsEnumerable = {}.propertyIsEnumerable;
-// eslint-disable-next-line es/no-object-getownpropertydescriptor -- safe
-var getOwnPropertyDescriptor = Object.getOwnPropertyDescriptor;
-
-// Nashorn ~ JDK8 bug
-var NASHORN_BUG = getOwnPropertyDescriptor && !$propertyIsEnumerable.call({ 1: 2 }, 1);
-
-// `Object.prototype.propertyIsEnumerable` method implementation
-// https://tc39.es/ecma262/#sec-object.prototype.propertyisenumerable
-var f$2 = NASHORN_BUG ? function propertyIsEnumerable(V) {
-  var descriptor = getOwnPropertyDescriptor(this, V);
-  return !!descriptor && descriptor.enumerable;
-} : $propertyIsEnumerable;
-
-var objectPropertyIsEnumerable = {
-	f: f$2
-};
-
-// eslint-disable-next-line es/no-object-getownpropertydescriptor -- safe
-var $getOwnPropertyDescriptor$1 = Object.getOwnPropertyDescriptor;
-
-// `Object.getOwnPropertyDescriptor` method
-// https://tc39.es/ecma262/#sec-object.getownpropertydescriptor
-var f$3 = descriptors ? $getOwnPropertyDescriptor$1 : function getOwnPropertyDescriptor(O, P) {
-  O = toIndexedObject(O);
-  P = toPropertyKey(P);
-  if (ie8DomDefine) try {
-    return $getOwnPropertyDescriptor$1(O, P);
-  } catch (error) { /* empty */ }
-  if (hasOwnProperty_1(O, P)) return createPropertyDescriptor(!functionCall(objectPropertyIsEnumerable.f, O, P), O[P]);
-};
-
-var objectGetOwnPropertyDescriptor = {
-	f: f$3
-};
-
-var FunctionPrototype$1 = Function.prototype;
-// eslint-disable-next-line es/no-object-getownpropertydescriptor -- safe
-var getDescriptor = descriptors && Object.getOwnPropertyDescriptor;
-
-var EXISTS$1 = hasOwnProperty_1(FunctionPrototype$1, 'name');
-// additional protection from minified / mangled / dropped function names
-var PROPER = EXISTS$1 && (function something() { /* empty */ }).name === 'something';
-var CONFIGURABLE$1 = EXISTS$1 && (!descriptors || (descriptors && getDescriptor(FunctionPrototype$1, 'name').configurable));
-
-var functionName = {
-  EXISTS: EXISTS$1,
-  PROPER: PROPER,
-  CONFIGURABLE: CONFIGURABLE$1
-};
-
-var redefine = createCommonjsModule(function (module) {
-var CONFIGURABLE_FUNCTION_NAME = functionName.CONFIGURABLE;
-
-var getInternalState = internalState.get;
-var enforceInternalState = internalState.enforce;
-var TEMPLATE = String(String).split('String');
-
-(module.exports = function (O, key, value, options) {
-  var unsafe = options ? !!options.unsafe : false;
-  var simple = options ? !!options.enumerable : false;
-  var noTargetGet = options ? !!options.noTargetGet : false;
-  var name = options && options.name !== undefined ? options.name : key;
-  var state;
-  if (isCallable(value)) {
-    if (String(name).slice(0, 7) === 'Symbol(') {
-      name = '[' + String(name).replace(/^Symbol\(([^)]*)\)/, '$1') + ']';
-    }
-    if (!hasOwnProperty_1(value, 'name') || (CONFIGURABLE_FUNCTION_NAME && value.name !== name)) {
-      createNonEnumerableProperty(value, 'name', name);
-    }
-    state = enforceInternalState(value);
-    if (!state.source) {
-      state.source = TEMPLATE.join(typeof name == 'string' ? name : '');
-    }
-  }
-  if (O === global_1) {
-    if (simple) O[key] = value;
-    else setGlobal(key, value);
-    return;
-  } else if (!unsafe) {
-    delete O[key];
-  } else if (!noTargetGet && O[key]) {
-    simple = true;
-  }
-  if (simple) O[key] = value;
-  else createNonEnumerableProperty(O, key, value);
-// add fake Function#toString for correct work wrapped methods / constructors with methods like LoDash isNative
-})(Function.prototype, 'toString', function toString() {
-  return isCallable(this) && getInternalState(this).source || inspectSource(this);
-});
-});
-
-var hiddenKeys$1 = enumBugKeys.concat('length', 'prototype');
-
-// `Object.getOwnPropertyNames` method
-// https://tc39.es/ecma262/#sec-object.getownpropertynames
-// eslint-disable-next-line es/no-object-getownpropertynames -- safe
-var f$4 = Object.getOwnPropertyNames || function getOwnPropertyNames(O) {
-  return objectKeysInternal(O, hiddenKeys$1);
-};
-
-var objectGetOwnPropertyNames = {
-	f: f$4
-};
-
-// eslint-disable-next-line es/no-object-getownpropertysymbols -- safe
-var f$5 = Object.getOwnPropertySymbols;
-
-var objectGetOwnPropertySymbols = {
-	f: f$5
-};
-
-var concat = functionUncurryThis([].concat);
-
-// all object keys, includes non-enumerable and symbols
-var ownKeys = getBuiltIn('Reflect', 'ownKeys') || function ownKeys(it) {
-  var keys = objectGetOwnPropertyNames.f(anObject(it));
-  var getOwnPropertySymbols = objectGetOwnPropertySymbols.f;
-  return getOwnPropertySymbols ? concat(keys, getOwnPropertySymbols(it)) : keys;
-};
-
-var copyConstructorProperties = function (target, source, exceptions) {
-  var keys = ownKeys(source);
-  var defineProperty = objectDefineProperty.f;
-  var getOwnPropertyDescriptor = objectGetOwnPropertyDescriptor.f;
-  for (var i = 0; i < keys.length; i++) {
-    var key = keys[i];
-    if (!hasOwnProperty_1(target, key) && !(exceptions && hasOwnProperty_1(exceptions, key))) {
-      defineProperty(target, key, getOwnPropertyDescriptor(source, key));
-    }
-  }
-};
-
-var replacement = /#|\.prototype\./;
-
-var isForced = function (feature, detection) {
-  var value = data[normalize(feature)];
-  return value == POLYFILL ? true
-    : value == NATIVE ? false
-    : isCallable(detection) ? fails(detection)
-    : !!detection;
-};
-
-var normalize = isForced.normalize = function (string) {
-  return String(string).replace(replacement, '.').toLowerCase();
-};
-
-var data = isForced.data = {};
-var NATIVE = isForced.NATIVE = 'N';
-var POLYFILL = isForced.POLYFILL = 'P';
-
-var isForced_1 = isForced;
-
-var getOwnPropertyDescriptor$1 = objectGetOwnPropertyDescriptor.f;
-
-
-
-
-
-
-/*
-  options.target      - name of the target object
-  options.global      - target is the global object
-  options.stat        - export as static methods of target
-  options.proto       - export as prototype methods of target
-  options.real        - real prototype method for the `pure` version
-  options.forced      - export even if the native feature is available
-  options.bind        - bind methods to the target, required for the `pure` version
-  options.wrap        - wrap constructors to preventing global pollution, required for the `pure` version
-  options.unsafe      - use the simple assignment of property instead of delete + defineProperty
-  options.sham        - add a flag to not completely full polyfills
-  options.enumerable  - export as enumerable property
-  options.noTargetGet - prevent calling a getter on target
-  options.name        - the .name of the function if it does not match the key
-*/
-var _export = function (options, source) {
-  var TARGET = options.target;
-  var GLOBAL = options.global;
-  var STATIC = options.stat;
-  var FORCED, target, key, targetProperty, sourceProperty, descriptor;
-  if (GLOBAL) {
-    target = global_1;
-  } else if (STATIC) {
-    target = global_1[TARGET] || setGlobal(TARGET, {});
-  } else {
-    target = (global_1[TARGET] || {}).prototype;
-  }
-  if (target) for (key in source) {
-    sourceProperty = source[key];
-    if (options.noTargetGet) {
-      descriptor = getOwnPropertyDescriptor$1(target, key);
-      targetProperty = descriptor && descriptor.value;
-    } else targetProperty = target[key];
-    FORCED = isForced_1(GLOBAL ? key : TARGET + (STATIC ? '.' : '#') + key, options.forced);
-    // contained in target
-    if (!FORCED && targetProperty !== undefined) {
-      if (typeof sourceProperty == typeof targetProperty) continue;
-      copyConstructorProperties(sourceProperty, targetProperty);
-    }
-    // add a flag to not completely full polyfills
-    if (options.sham || (targetProperty && targetProperty.sham)) {
-      createNonEnumerableProperty(sourceProperty, 'sham', true);
-    }
-    // extend global
-    redefine(target, key, sourceProperty, options);
-  }
-};
-
-var correctPrototypeGetter = !fails(function () {
-  function F() { /* empty */ }
-  F.prototype.constructor = null;
-  // eslint-disable-next-line es/no-object-getprototypeof -- required for testing
-  return Object.getPrototypeOf(new F()) !== F.prototype;
-});
-
-var IE_PROTO$1 = sharedKey('IE_PROTO');
-var Object$4 = global_1.Object;
-var ObjectPrototype = Object$4.prototype;
-
-// `Object.getPrototypeOf` method
-// https://tc39.es/ecma262/#sec-object.getprototypeof
-var objectGetPrototypeOf = correctPrototypeGetter ? Object$4.getPrototypeOf : function (O) {
-  var object = toObject(O);
-  if (hasOwnProperty_1(object, IE_PROTO$1)) return object[IE_PROTO$1];
-  var constructor = object.constructor;
-  if (isCallable(constructor) && object instanceof constructor) {
-    return constructor.prototype;
-  } return object instanceof Object$4 ? ObjectPrototype : null;
-};
-
-var ITERATOR = wellKnownSymbol('iterator');
-var BUGGY_SAFARI_ITERATORS = false;
-
-// `%IteratorPrototype%` object
-// https://tc39.es/ecma262/#sec-%iteratorprototype%-object
-var IteratorPrototype, PrototypeOfArrayIteratorPrototype, arrayIterator;
-
-/* eslint-disable es/no-array-prototype-keys -- safe */
-if ([].keys) {
-  arrayIterator = [].keys();
-  // Safari 8 has buggy iterators w/o `next`
-  if (!('next' in arrayIterator)) BUGGY_SAFARI_ITERATORS = true;
-  else {
-    PrototypeOfArrayIteratorPrototype = objectGetPrototypeOf(objectGetPrototypeOf(arrayIterator));
-    if (PrototypeOfArrayIteratorPrototype !== Object.prototype) IteratorPrototype = PrototypeOfArrayIteratorPrototype;
-  }
-}
-
-var NEW_ITERATOR_PROTOTYPE = IteratorPrototype == undefined || fails(function () {
-  var test = {};
-  // FF44- legacy iterators case
-  return IteratorPrototype[ITERATOR].call(test) !== test;
-});
-
-if (NEW_ITERATOR_PROTOTYPE) IteratorPrototype = {};
-
-// `%IteratorPrototype%[@@iterator]()` method
-// https://tc39.es/ecma262/#sec-%iteratorprototype%-@@iterator
-if (!isCallable(IteratorPrototype[ITERATOR])) {
-  redefine(IteratorPrototype, ITERATOR, function () {
-    return this;
-  });
-}
-
-var iteratorsCore = {
-  IteratorPrototype: IteratorPrototype,
-  BUGGY_SAFARI_ITERATORS: BUGGY_SAFARI_ITERATORS
-};
-
-var defineProperty$1 = objectDefineProperty.f;
-
-
-
-var TO_STRING_TAG = wellKnownSymbol('toStringTag');
-
-var setToStringTag = function (target, TAG, STATIC) {
-  if (target && !STATIC) target = target.prototype;
-  if (target && !hasOwnProperty_1(target, TO_STRING_TAG)) {
-    defineProperty$1(target, TO_STRING_TAG, { configurable: true, value: TAG });
-  }
-};
-
-var IteratorPrototype$1 = iteratorsCore.IteratorPrototype;
-
-
-
-
-
-var returnThis = function () { return this; };
-
-var createIteratorConstructor = function (IteratorConstructor, NAME, next, ENUMERABLE_NEXT) {
-  var TO_STRING_TAG = NAME + ' Iterator';
-  IteratorConstructor.prototype = objectCreate(IteratorPrototype$1, { next: createPropertyDescriptor(+!ENUMERABLE_NEXT, next) });
-  setToStringTag(IteratorConstructor, TO_STRING_TAG, false);
-  iterators[TO_STRING_TAG] = returnThis;
-  return IteratorConstructor;
-};
-
-var String$3 = global_1.String;
-var TypeError$8 = global_1.TypeError;
-
-var aPossiblePrototype = function (argument) {
-  if (typeof argument == 'object' || isCallable(argument)) return argument;
-  throw TypeError$8("Can't set " + String$3(argument) + ' as a prototype');
-};
-
-/* eslint-disable no-proto -- safe */
-
-
-
-
-// `Object.setPrototypeOf` method
-// https://tc39.es/ecma262/#sec-object.setprototypeof
-// Works with __proto__ only. Old v8 can't work with null proto objects.
-// eslint-disable-next-line es/no-object-setprototypeof -- safe
-var objectSetPrototypeOf = Object.setPrototypeOf || ('__proto__' in {} ? function () {
-  var CORRECT_SETTER = false;
-  var test = {};
-  var setter;
-  try {
-    // eslint-disable-next-line es/no-object-getownpropertydescriptor -- safe
-    setter = functionUncurryThis(Object.getOwnPropertyDescriptor(Object.prototype, '__proto__').set);
-    setter(test, []);
-    CORRECT_SETTER = test instanceof Array;
-  } catch (error) { /* empty */ }
-  return function setPrototypeOf(O, proto) {
-    anObject(O);
-    aPossiblePrototype(proto);
-    if (CORRECT_SETTER) setter(O, proto);
-    else O.__proto__ = proto;
-    return O;
-  };
-}() : undefined);
-
-var PROPER_FUNCTION_NAME = functionName.PROPER;
-var CONFIGURABLE_FUNCTION_NAME = functionName.CONFIGURABLE;
-var IteratorPrototype$2 = iteratorsCore.IteratorPrototype;
-var BUGGY_SAFARI_ITERATORS$1 = iteratorsCore.BUGGY_SAFARI_ITERATORS;
-var ITERATOR$1 = wellKnownSymbol('iterator');
-var KEYS = 'keys';
-var VALUES = 'values';
-var ENTRIES = 'entries';
-
-var returnThis$1 = function () { return this; };
-
-var defineIterator = function (Iterable, NAME, IteratorConstructor, next, DEFAULT, IS_SET, FORCED) {
-  createIteratorConstructor(IteratorConstructor, NAME, next);
-
-  var getIterationMethod = function (KIND) {
-    if (KIND === DEFAULT && defaultIterator) return defaultIterator;
-    if (!BUGGY_SAFARI_ITERATORS$1 && KIND in IterablePrototype) return IterablePrototype[KIND];
-    switch (KIND) {
-      case KEYS: return function keys() { return new IteratorConstructor(this, KIND); };
-      case VALUES: return function values() { return new IteratorConstructor(this, KIND); };
-      case ENTRIES: return function entries() { return new IteratorConstructor(this, KIND); };
-    } return function () { return new IteratorConstructor(this); };
-  };
-
-  var TO_STRING_TAG = NAME + ' Iterator';
-  var INCORRECT_VALUES_NAME = false;
-  var IterablePrototype = Iterable.prototype;
-  var nativeIterator = IterablePrototype[ITERATOR$1]
-    || IterablePrototype['@@iterator']
-    || DEFAULT && IterablePrototype[DEFAULT];
-  var defaultIterator = !BUGGY_SAFARI_ITERATORS$1 && nativeIterator || getIterationMethod(DEFAULT);
-  var anyNativeIterator = NAME == 'Array' ? IterablePrototype.entries || nativeIterator : nativeIterator;
-  var CurrentIteratorPrototype, methods, KEY;
-
-  // fix native
-  if (anyNativeIterator) {
-    CurrentIteratorPrototype = objectGetPrototypeOf(anyNativeIterator.call(new Iterable()));
-    if (CurrentIteratorPrototype !== Object.prototype && CurrentIteratorPrototype.next) {
-      if ( objectGetPrototypeOf(CurrentIteratorPrototype) !== IteratorPrototype$2) {
-        if (objectSetPrototypeOf) {
-          objectSetPrototypeOf(CurrentIteratorPrototype, IteratorPrototype$2);
-        } else if (!isCallable(CurrentIteratorPrototype[ITERATOR$1])) {
-          redefine(CurrentIteratorPrototype, ITERATOR$1, returnThis$1);
-        }
-      }
-      // Set @@toStringTag to native iterators
-      setToStringTag(CurrentIteratorPrototype, TO_STRING_TAG, true);
-    }
-  }
-
-  // fix Array.prototype.{ values, @@iterator }.name in V8 / FF
-  if (PROPER_FUNCTION_NAME && DEFAULT == VALUES && nativeIterator && nativeIterator.name !== VALUES) {
-    if ( CONFIGURABLE_FUNCTION_NAME) {
-      createNonEnumerableProperty(IterablePrototype, 'name', VALUES);
-    } else {
-      INCORRECT_VALUES_NAME = true;
-      defaultIterator = function values() { return functionCall(nativeIterator, this); };
-    }
-  }
-
-  // export additional methods
-  if (DEFAULT) {
-    methods = {
-      values: getIterationMethod(VALUES),
-      keys: IS_SET ? defaultIterator : getIterationMethod(KEYS),
-      entries: getIterationMethod(ENTRIES)
-    };
-    if (FORCED) for (KEY in methods) {
-      if (BUGGY_SAFARI_ITERATORS$1 || INCORRECT_VALUES_NAME || !(KEY in IterablePrototype)) {
-        redefine(IterablePrototype, KEY, methods[KEY]);
-      }
-    } else _export({ target: NAME, proto: true, forced: BUGGY_SAFARI_ITERATORS$1 || INCORRECT_VALUES_NAME }, methods);
-  }
-
-  // define iterator
-  if ( IterablePrototype[ITERATOR$1] !== defaultIterator) {
-    redefine(IterablePrototype, ITERATOR$1, defaultIterator, { name: DEFAULT });
-  }
-  iterators[NAME] = defaultIterator;
-
-  return methods;
-};
-
-var defineProperty$2 = objectDefineProperty.f;
-
-
-
-
-var ARRAY_ITERATOR = 'Array Iterator';
-var setInternalState = internalState.set;
-var getInternalState = internalState.getterFor(ARRAY_ITERATOR);
-
-// `Array.prototype.entries` method
-// https://tc39.es/ecma262/#sec-array.prototype.entries
-// `Array.prototype.keys` method
-// https://tc39.es/ecma262/#sec-array.prototype.keys
-// `Array.prototype.values` method
-// https://tc39.es/ecma262/#sec-array.prototype.values
-// `Array.prototype[@@iterator]` method
-// https://tc39.es/ecma262/#sec-array.prototype-@@iterator
-// `CreateArrayIterator` internal method
-// https://tc39.es/ecma262/#sec-createarrayiterator
-var es_array_iterator = defineIterator(Array, 'Array', function (iterated, kind) {
-  setInternalState(this, {
-    type: ARRAY_ITERATOR,
-    target: toIndexedObject(iterated), // target
-    index: 0,                          // next index
-    kind: kind                         // kind
-  });
-// `%ArrayIteratorPrototype%.next` method
-// https://tc39.es/ecma262/#sec-%arrayiteratorprototype%.next
-}, function () {
-  var state = getInternalState(this);
-  var target = state.target;
-  var kind = state.kind;
-  var index = state.index++;
-  if (!target || index >= target.length) {
-    state.target = undefined;
-    return { value: undefined, done: true };
-  }
-  if (kind == 'keys') return { value: index, done: false };
-  if (kind == 'values') return { value: target[index], done: false };
-  return { value: [index, target[index]], done: false };
-}, 'values');
-
-// argumentsList[@@iterator] is %ArrayProto_values%
-// https://tc39.es/ecma262/#sec-createunmappedargumentsobject
-// https://tc39.es/ecma262/#sec-createmappedargumentsobject
-var values = iterators.Arguments = iterators.Array;
-
-// https://tc39.es/ecma262/#sec-array.prototype-@@unscopables
-addToUnscopables('keys');
-addToUnscopables('values');
-addToUnscopables('entries');
-
-// V8 ~ Chrome 45- bug
-if ( descriptors && values.name !== 'values') try {
-  defineProperty$2(values, 'name', { value: 'values' });
-} catch (error) { /* empty */ }
-
-var ITERATOR$2 = wellKnownSymbol('iterator');
-var TO_STRING_TAG$1 = wellKnownSymbol('toStringTag');
-var ArrayValues = es_array_iterator.values;
-
-var handlePrototype = function (CollectionPrototype, COLLECTION_NAME) {
-  if (CollectionPrototype) {
-    // some Chrome versions have non-configurable methods on DOMTokenList
-    if (CollectionPrototype[ITERATOR$2] !== ArrayValues) try {
-      createNonEnumerableProperty(CollectionPrototype, ITERATOR$2, ArrayValues);
-    } catch (error) {
-      CollectionPrototype[ITERATOR$2] = ArrayValues;
-    }
-    if (!CollectionPrototype[TO_STRING_TAG$1]) {
-      createNonEnumerableProperty(CollectionPrototype, TO_STRING_TAG$1, COLLECTION_NAME);
-    }
-    if (domIterables[COLLECTION_NAME]) for (var METHOD_NAME in es_array_iterator) {
-      // some Chrome versions have non-configurable methods on DOMTokenList
-      if (CollectionPrototype[METHOD_NAME] !== es_array_iterator[METHOD_NAME]) try {
-        createNonEnumerableProperty(CollectionPrototype, METHOD_NAME, es_array_iterator[METHOD_NAME]);
-      } catch (error) {
-        CollectionPrototype[METHOD_NAME] = es_array_iterator[METHOD_NAME];
-      }
-    }
-  }
-};
-
-for (var COLLECTION_NAME in domIterables) {
-  handlePrototype(global_1[COLLECTION_NAME] && global_1[COLLECTION_NAME].prototype, COLLECTION_NAME);
-}
-
-handlePrototype(domTokenListPrototype, 'DOMTokenList');
-
-// https://github.com/tc39/collection-methods
-var collectionAddAll = function addAll(/* ...elements */) {
-  var set = anObject(this);
-  var adder = aCallable(set.add);
-  for (var k = 0, len = arguments.length; k < len; k++) {
-    functionCall(adder, set, arguments[k]);
-  }
-  return set;
-};
-
-// `Set.prototype.addAll` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  addAll: collectionAddAll
-});
-
-// https://github.com/tc39/collection-methods
-var collectionDeleteAll = function deleteAll(/* ...elements */) {
-  var collection = anObject(this);
-  var remover = aCallable(collection['delete']);
-  var allDeleted = true;
-  var wasDeleted;
-  for (var k = 0, len = arguments.length; k < len; k++) {
-    wasDeleted = functionCall(remover, collection, arguments[k]);
-    allDeleted = allDeleted && wasDeleted;
-  }
-  return !!allDeleted;
-};
-
-// `Set.prototype.deleteAll` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  deleteAll: collectionDeleteAll
-});
-
-var TO_STRING_TAG$2 = wellKnownSymbol('toStringTag');
-var test = {};
-
-test[TO_STRING_TAG$2] = 'z';
-
-var toStringTagSupport = String(test) === '[object z]';
-
-var TO_STRING_TAG$3 = wellKnownSymbol('toStringTag');
-var Object$5 = global_1.Object;
-
-// ES3 wrong here
-var CORRECT_ARGUMENTS = classofRaw(function () { return arguments; }()) == 'Arguments';
-
-// fallback for IE11 Script Access Denied error
-var tryGet = function (it, key) {
-  try {
-    return it[key];
-  } catch (error) { /* empty */ }
-};
-
-// getting tag from ES6+ `Object.prototype.toString`
-var classof = toStringTagSupport ? classofRaw : function (it) {
-  var O, tag, result;
-  return it === undefined ? 'Undefined' : it === null ? 'Null'
-    // @@toStringTag case
-    : typeof (tag = tryGet(O = Object$5(it), TO_STRING_TAG$3)) == 'string' ? tag
-    // builtinTag case
-    : CORRECT_ARGUMENTS ? classofRaw(O)
-    // ES3 arguments fallback
-    : (result = classofRaw(O)) == 'Object' && isCallable(O.callee) ? 'Arguments' : result;
-};
-
-var noop = function () { /* empty */ };
-var empty = [];
-var construct = getBuiltIn('Reflect', 'construct');
-var constructorRegExp = /^\s*(?:class|function)\b/;
-var exec = functionUncurryThis(constructorRegExp.exec);
-var INCORRECT_TO_STRING = !constructorRegExp.exec(noop);
-
-var isConstructorModern = function isConstructor(argument) {
-  if (!isCallable(argument)) return false;
-  try {
-    construct(noop, empty, argument);
-    return true;
-  } catch (error) {
-    return false;
-  }
-};
-
-var isConstructorLegacy = function isConstructor(argument) {
-  if (!isCallable(argument)) return false;
-  switch (classof(argument)) {
-    case 'AsyncFunction':
-    case 'GeneratorFunction':
-    case 'AsyncGeneratorFunction': return false;
-  }
-  try {
-    // we can't check .prototype since constructors produced by .bind haven't it
-    // `Function#toString` throws on some built-it function in some legacy engines
-    // (for example, `DOMQuad` and similar in FF41-)
-    return INCORRECT_TO_STRING || !!exec(constructorRegExp, inspectSource(argument));
-  } catch (error) {
-    return true;
-  }
-};
-
-isConstructorLegacy.sham = true;
-
-// `IsConstructor` abstract operation
-// https://tc39.es/ecma262/#sec-isconstructor
-var isConstructor = !construct || fails(function () {
-  var called;
-  return isConstructorModern(isConstructorModern.call)
-    || !isConstructorModern(Object)
-    || !isConstructorModern(function () { called = true; })
-    || called;
-}) ? isConstructorLegacy : isConstructorModern;
-
-var TypeError$9 = global_1.TypeError;
-
-// `Assert: IsConstructor(argument) is true`
-var aConstructor = function (argument) {
-  if (isConstructor(argument)) return argument;
-  throw TypeError$9(tryToString(argument) + ' is not a constructor');
-};
-
-var SPECIES = wellKnownSymbol('species');
-
-// `SpeciesConstructor` abstract operation
-// https://tc39.es/ecma262/#sec-speciesconstructor
-var speciesConstructor = function (O, defaultConstructor) {
-  var C = anObject(O).constructor;
-  var S;
-  return C === undefined || (S = anObject(C)[SPECIES]) == undefined ? defaultConstructor : aConstructor(S);
-};
-
-var bind$1 = functionUncurryThis(functionUncurryThis.bind);
-
-// optional / simple context binding
-var functionBindContext = function (fn, that) {
-  aCallable(fn);
-  return that === undefined ? fn : functionBindNative ? bind$1(fn, that) : function (/* ...args */) {
-    return fn.apply(that, arguments);
-  };
-};
-
-var ITERATOR$3 = wellKnownSymbol('iterator');
-var ArrayPrototype$1 = Array.prototype;
-
-// check on default Array iterator
-var isArrayIteratorMethod = function (it) {
-  return it !== undefined && (iterators.Array === it || ArrayPrototype$1[ITERATOR$3] === it);
-};
-
-var ITERATOR$4 = wellKnownSymbol('iterator');
-
-var getIteratorMethod = function (it) {
-  if (it != undefined) return getMethod(it, ITERATOR$4)
-    || getMethod(it, '@@iterator')
-    || iterators[classof(it)];
-};
-
-var TypeError$a = global_1.TypeError;
-
-var getIterator = function (argument, usingIterator) {
-  var iteratorMethod = arguments.length < 2 ? getIteratorMethod(argument) : usingIterator;
-  if (aCallable(iteratorMethod)) return anObject(functionCall(iteratorMethod, argument));
-  throw TypeError$a(tryToString(argument) + ' is not iterable');
-};
-
-var iteratorClose = function (iterator, kind, value) {
-  var innerResult, innerError;
-  anObject(iterator);
-  try {
-    innerResult = getMethod(iterator, 'return');
-    if (!innerResult) {
-      if (kind === 'throw') throw value;
-      return value;
-    }
-    innerResult = functionCall(innerResult, iterator);
-  } catch (error) {
-    innerError = true;
-    innerResult = error;
-  }
-  if (kind === 'throw') throw value;
-  if (innerError) throw innerResult;
-  anObject(innerResult);
-  return value;
-};
-
-var TypeError$b = global_1.TypeError;
-
-var Result = function (stopped, result) {
-  this.stopped = stopped;
-  this.result = result;
-};
-
-var ResultPrototype = Result.prototype;
-
-var iterate = function (iterable, unboundFunction, options) {
-  var that = options && options.that;
-  var AS_ENTRIES = !!(options && options.AS_ENTRIES);
-  var IS_ITERATOR = !!(options && options.IS_ITERATOR);
-  var INTERRUPTED = !!(options && options.INTERRUPTED);
-  var fn = functionBindContext(unboundFunction, that);
-  var iterator, iterFn, index, length, result, next, step;
-
-  var stop = function (condition) {
-    if (iterator) iteratorClose(iterator, 'normal', condition);
-    return new Result(true, condition);
-  };
-
-  var callFn = function (value) {
-    if (AS_ENTRIES) {
-      anObject(value);
-      return INTERRUPTED ? fn(value[0], value[1], stop) : fn(value[0], value[1]);
-    } return INTERRUPTED ? fn(value, stop) : fn(value);
-  };
-
-  if (IS_ITERATOR) {
-    iterator = iterable;
-  } else {
-    iterFn = getIteratorMethod(iterable);
-    if (!iterFn) throw TypeError$b(tryToString(iterable) + ' is not iterable');
-    // optimisation for array iterators
-    if (isArrayIteratorMethod(iterFn)) {
-      for (index = 0, length = lengthOfArrayLike(iterable); length > index; index++) {
-        result = callFn(iterable[index]);
-        if (result && objectIsPrototypeOf(ResultPrototype, result)) return result;
-      } return new Result(false);
-    }
-    iterator = getIterator(iterable, iterFn);
-  }
-
-  next = iterator.next;
-  while (!(step = functionCall(next, iterator)).done) {
-    try {
-      result = callFn(step.value);
-    } catch (error) {
-      iteratorClose(iterator, 'throw', error);
-    }
-    if (typeof result == 'object' && result && objectIsPrototypeOf(ResultPrototype, result)) return result;
-  } return new Result(false);
-};
-
-// `Set.prototype.difference` method
-// https://github.com/tc39/proposal-set-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  difference: function difference(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))(set);
-    var remover = aCallable(newSet['delete']);
-    iterate(iterable, function (value) {
-      functionCall(remover, newSet, value);
-    });
-    return newSet;
-  }
-});
-
-var getSetIterator = function (it) {
-  // eslint-disable-next-line es/no-set -- safe
-  return functionCall(Set.prototype.values, it);
-};
-
-// `Set.prototype.every` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  every: function every(callbackfn /* , thisArg */) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return !iterate(iterator, function (value, stop) {
-      if (!boundFunction(value, value, set)) return stop();
-    }, { IS_ITERATOR: true, INTERRUPTED: true }).stopped;
-  }
-});
-
-// `Set.prototype.filter` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  filter: function filter(callbackfn /* , thisArg */) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))();
-    var adder = aCallable(newSet.add);
-    iterate(iterator, function (value) {
-      if (boundFunction(value, value, set)) functionCall(adder, newSet, value);
-    }, { IS_ITERATOR: true });
-    return newSet;
-  }
-});
-
-// `Set.prototype.find` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  find: function find(callbackfn /* , thisArg */) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return iterate(iterator, function (value, stop) {
-      if (boundFunction(value, value, set)) return stop(value);
-    }, { IS_ITERATOR: true, INTERRUPTED: true }).result;
-  }
-});
-
-// `Set.prototype.intersection` method
-// https://github.com/tc39/proposal-set-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  intersection: function intersection(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))();
-    var hasCheck = aCallable(set.has);
-    var adder = aCallable(newSet.add);
-    iterate(iterable, function (value) {
-      if (functionCall(hasCheck, set, value)) functionCall(adder, newSet, value);
-    });
-    return newSet;
-  }
-});
-
-// `Set.prototype.isDisjointFrom` method
-// https://tc39.github.io/proposal-set-methods/#Set.prototype.isDisjointFrom
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  isDisjointFrom: function isDisjointFrom(iterable) {
-    var set = anObject(this);
-    var hasCheck = aCallable(set.has);
-    return !iterate(iterable, function (value, stop) {
-      if (functionCall(hasCheck, set, value) === true) return stop();
-    }, { INTERRUPTED: true }).stopped;
-  }
-});
-
-// `Set.prototype.isSubsetOf` method
-// https://tc39.github.io/proposal-set-methods/#Set.prototype.isSubsetOf
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  isSubsetOf: function isSubsetOf(iterable) {
-    var iterator = getIterator(this);
-    var otherSet = anObject(iterable);
-    var hasCheck = otherSet.has;
-    if (!isCallable(hasCheck)) {
-      otherSet = new (getBuiltIn('Set'))(iterable);
-      hasCheck = aCallable(otherSet.has);
-    }
-    return !iterate(iterator, function (value, stop) {
-      if (functionCall(hasCheck, otherSet, value) === false) return stop();
-    }, { IS_ITERATOR: true, INTERRUPTED: true }).stopped;
-  }
-});
-
-// `Set.prototype.isSupersetOf` method
-// https://tc39.github.io/proposal-set-methods/#Set.prototype.isSupersetOf
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  isSupersetOf: function isSupersetOf(iterable) {
-    var set = anObject(this);
-    var hasCheck = aCallable(set.has);
-    return !iterate(iterable, function (value, stop) {
-      if (functionCall(hasCheck, set, value) === false) return stop();
-    }, { INTERRUPTED: true }).stopped;
-  }
-});
-
-var String$4 = global_1.String;
-
-var toString_1 = function (argument) {
-  if (classof(argument) === 'Symbol') throw TypeError('Cannot convert a Symbol value to a string');
-  return String$4(argument);
-};
-
-var arrayJoin = functionUncurryThis([].join);
-var push$1 = [].push;
-
-// `Set.prototype.join` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  join: function join(separator) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var sep = separator === undefined ? ',' : toString_1(separator);
-    var result = [];
-    iterate(iterator, push$1, { that: result, IS_ITERATOR: true });
-    return arrayJoin(result, sep);
-  }
-});
-
-// `Set.prototype.map` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  map: function map(callbackfn /* , thisArg */) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))();
-    var adder = aCallable(newSet.add);
-    iterate(iterator, function (value) {
-      functionCall(adder, newSet, boundFunction(value, value, set));
-    }, { IS_ITERATOR: true });
-    return newSet;
-  }
-});
-
-var TypeError$c = global_1.TypeError;
-
-// `Set.prototype.reduce` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  reduce: function reduce(callbackfn /* , initialValue */) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var noInitial = arguments.length < 2;
-    var accumulator = noInitial ? undefined : arguments[1];
-    aCallable(callbackfn);
-    iterate(iterator, function (value) {
-      if (noInitial) {
-        noInitial = false;
-        accumulator = value;
-      } else {
-        accumulator = callbackfn(accumulator, value, value, set);
-      }
-    }, { IS_ITERATOR: true });
-    if (noInitial) throw TypeError$c('Reduce of empty set with no initial value');
-    return accumulator;
-  }
-});
-
-// `Set.prototype.some` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  some: function some(callbackfn /* , thisArg */) {
-    var set = anObject(this);
-    var iterator = getSetIterator(set);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return iterate(iterator, function (value, stop) {
-      if (boundFunction(value, value, set)) return stop();
-    }, { IS_ITERATOR: true, INTERRUPTED: true }).stopped;
-  }
-});
-
-// `Set.prototype.symmetricDifference` method
-// https://github.com/tc39/proposal-set-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  symmetricDifference: function symmetricDifference(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))(set);
-    var remover = aCallable(newSet['delete']);
-    var adder = aCallable(newSet.add);
-    iterate(iterable, function (value) {
-      functionCall(remover, newSet, value) || functionCall(adder, newSet, value);
-    });
-    return newSet;
-  }
-});
-
-// `Set.prototype.union` method
-// https://github.com/tc39/proposal-set-methods
-_export({ target: 'Set', proto: true, real: true, forced: true }, {
-  union: function union(iterable) {
-    var set = anObject(this);
-    var newSet = new (speciesConstructor(set, getBuiltIn('Set')))(set);
-    iterate(iterable, aCallable(newSet.add), { that: newSet });
-    return newSet;
-  }
-});
-
-// `Map.prototype.deleteAll` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  deleteAll: collectionDeleteAll
-});
-
-// `Map.prototype.emplace` method
-// https://github.com/thumbsupep/proposal-upsert
-var mapEmplace = function emplace(key, handler) {
-  var map = anObject(this);
-  var get = aCallable(map.get);
-  var has = aCallable(map.has);
-  var set = aCallable(map.set);
-  var value = (functionCall(has, map, key) && 'update' in handler)
-    ? handler.update(functionCall(get, map, key), key, map)
-    : handler.insert(key, map);
-  functionCall(set, map, key, value);
-  return value;
-};
-
-// `Map.prototype.emplace` method
-// https://github.com/thumbsupep/proposal-upsert
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  emplace: mapEmplace
-});
-
-var getMapIterator = function (it) {
-  // eslint-disable-next-line es/no-map -- safe
-  return functionCall(Map.prototype.entries, it);
-};
-
-// `Map.prototype.every` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  every: function every(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return !iterate(iterator, function (key, value, stop) {
-      if (!boundFunction(value, key, map)) return stop();
-    }, { AS_ENTRIES: true, IS_ITERATOR: true, INTERRUPTED: true }).stopped;
-  }
-});
-
-// `Map.prototype.filter` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  filter: function filter(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    var newMap = new (speciesConstructor(map, getBuiltIn('Map')))();
-    var setter = aCallable(newMap.set);
-    iterate(iterator, function (key, value) {
-      if (boundFunction(value, key, map)) functionCall(setter, newMap, key, value);
-    }, { AS_ENTRIES: true, IS_ITERATOR: true });
-    return newMap;
-  }
-});
-
-// `Map.prototype.find` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  find: function find(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return iterate(iterator, function (key, value, stop) {
-      if (boundFunction(value, key, map)) return stop(value);
-    }, { AS_ENTRIES: true, IS_ITERATOR: true, INTERRUPTED: true }).result;
-  }
-});
-
-// `Map.prototype.findKey` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  findKey: function findKey(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return iterate(iterator, function (key, value, stop) {
-      if (boundFunction(value, key, map)) return stop(key);
-    }, { AS_ENTRIES: true, IS_ITERATOR: true, INTERRUPTED: true }).result;
-  }
-});
-
-// `SameValueZero` abstract operation
-// https://tc39.es/ecma262/#sec-samevaluezero
-var sameValueZero = function (x, y) {
-  // eslint-disable-next-line no-self-compare -- NaN check
-  return x === y || x != x && y != y;
-};
-
-// `Map.prototype.includes` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  includes: function includes(searchElement) {
-    return iterate(getMapIterator(anObject(this)), function (key, value, stop) {
-      if (sameValueZero(value, searchElement)) return stop();
-    }, { AS_ENTRIES: true, IS_ITERATOR: true, INTERRUPTED: true }).stopped;
-  }
-});
-
-// `Map.prototype.keyOf` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  keyOf: function keyOf(searchElement) {
-    return iterate(getMapIterator(anObject(this)), function (key, value, stop) {
-      if (value === searchElement) return stop(key);
-    }, { AS_ENTRIES: true, IS_ITERATOR: true, INTERRUPTED: true }).result;
-  }
-});
-
-// `Map.prototype.mapKeys` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  mapKeys: function mapKeys(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    var newMap = new (speciesConstructor(map, getBuiltIn('Map')))();
-    var setter = aCallable(newMap.set);
-    iterate(iterator, function (key, value) {
-      functionCall(setter, newMap, boundFunction(value, key, map), value);
-    }, { AS_ENTRIES: true, IS_ITERATOR: true });
-    return newMap;
-  }
-});
-
-// `Map.prototype.mapValues` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  mapValues: function mapValues(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    var newMap = new (speciesConstructor(map, getBuiltIn('Map')))();
-    var setter = aCallable(newMap.set);
-    iterate(iterator, function (key, value) {
-      functionCall(setter, newMap, key, boundFunction(value, key, map));
-    }, { AS_ENTRIES: true, IS_ITERATOR: true });
-    return newMap;
-  }
-});
-
-// `Map.prototype.merge` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  // eslint-disable-next-line no-unused-vars -- required for `.length`
-  merge: function merge(iterable /* ...iterables */) {
-    var map = anObject(this);
-    var setter = aCallable(map.set);
-    var argumentsLength = arguments.length;
-    var i = 0;
-    while (i < argumentsLength) {
-      iterate(arguments[i++], setter, { that: map, AS_ENTRIES: true });
-    }
-    return map;
-  }
-});
-
-var TypeError$d = global_1.TypeError;
-
-// `Map.prototype.reduce` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  reduce: function reduce(callbackfn /* , initialValue */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var noInitial = arguments.length < 2;
-    var accumulator = noInitial ? undefined : arguments[1];
-    aCallable(callbackfn);
-    iterate(iterator, function (key, value) {
-      if (noInitial) {
-        noInitial = false;
-        accumulator = value;
-      } else {
-        accumulator = callbackfn(accumulator, value, key, map);
-      }
-    }, { AS_ENTRIES: true, IS_ITERATOR: true });
-    if (noInitial) throw TypeError$d('Reduce of empty map with no initial value');
-    return accumulator;
-  }
-});
-
-// `Set.prototype.some` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  some: function some(callbackfn /* , thisArg */) {
-    var map = anObject(this);
-    var iterator = getMapIterator(map);
-    var boundFunction = functionBindContext(callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-    return iterate(iterator, function (key, value, stop) {
-      if (boundFunction(value, key, map)) return stop();
-    }, { AS_ENTRIES: true, IS_ITERATOR: true, INTERRUPTED: true }).stopped;
-  }
-});
-
-var TypeError$e = global_1.TypeError;
-
-// `Set.prototype.update` method
-// https://github.com/tc39/proposal-collection-methods
-_export({ target: 'Map', proto: true, real: true, forced: true }, {
-  update: function update(key, callback /* , thunk */) {
-    var map = anObject(this);
-    var get = aCallable(map.get);
-    var has = aCallable(map.has);
-    var set = aCallable(map.set);
-    var length = arguments.length;
-    aCallable(callback);
-    var isPresentInMap = functionCall(has, map, key);
-    if (!isPresentInMap && length < 3) {
-      throw TypeError$e('Updating absent value');
-    }
-    var value = isPresentInMap ? functionCall(get, map, key) : aCallable(length > 2 ? arguments[2] : undefined)(key, map);
-    functionCall(set, map, key, callback(value, key, map));
-    return map;
-  }
-});
-
-/**
- * Internal use only. **Do NOT use in user land code!**
- *
- * @internal
- */
-
-const SEMAPHORE = Symbol();
-
-var ITERATOR$5 = wellKnownSymbol('iterator');
-var SAFE_CLOSING = false;
-
-try {
-  var called = 0;
-  var iteratorWithReturn = {
-    next: function () {
-      return { done: !!called++ };
-    },
-    'return': function () {
-      SAFE_CLOSING = true;
-    }
-  };
-  iteratorWithReturn[ITERATOR$5] = function () {
-    return this;
-  };
-  // eslint-disable-next-line es/no-array-from, no-throw-literal -- required for testing
-  Array.from(iteratorWithReturn, function () { throw 2; });
-} catch (error) { /* empty */ }
-
-var checkCorrectnessOfIteration = function (exec, SKIP_CLOSING) {
-  if (!SKIP_CLOSING && !SAFE_CLOSING) return false;
-  var ITERATION_SUPPORT = false;
-  try {
-    var object = {};
-    object[ITERATOR$5] = function () {
-      return {
-        next: function () {
-          return { done: ITERATION_SUPPORT = true };
-        }
-      };
-    };
-    exec(object);
-  } catch (error) { /* empty */ }
-  return ITERATION_SUPPORT;
-};
-
-// eslint-disable-next-line es/no-typed-arrays -- safe
-var arrayBufferNative = typeof ArrayBuffer != 'undefined' && typeof DataView != 'undefined';
-
-var defineProperty$3 = objectDefineProperty.f;
-
-
-
-
-
-
-var Int8Array = global_1.Int8Array;
-var Int8ArrayPrototype = Int8Array && Int8Array.prototype;
-var Uint8ClampedArray$1 = global_1.Uint8ClampedArray;
-var Uint8ClampedArrayPrototype = Uint8ClampedArray$1 && Uint8ClampedArray$1.prototype;
-var TypedArray = Int8Array && objectGetPrototypeOf(Int8Array);
-var TypedArrayPrototype = Int8ArrayPrototype && objectGetPrototypeOf(Int8ArrayPrototype);
-var ObjectPrototype$1 = Object.prototype;
-var TypeError$f = global_1.TypeError;
-
-var TO_STRING_TAG$4 = wellKnownSymbol('toStringTag');
-var TYPED_ARRAY_TAG = uid('TYPED_ARRAY_TAG');
-var TYPED_ARRAY_CONSTRUCTOR = uid('TYPED_ARRAY_CONSTRUCTOR');
-// Fixing native typed arrays in Opera Presto crashes the browser, see #595
-var NATIVE_ARRAY_BUFFER_VIEWS = arrayBufferNative && !!objectSetPrototypeOf && classof(global_1.opera) !== 'Opera';
-var TYPED_ARRAY_TAG_REQUIRED = false;
-var NAME, Constructor, Prototype;
-
-var TypedArrayConstructorsList = {
-  Int8Array: 1,
-  Uint8Array: 1,
-  Uint8ClampedArray: 1,
-  Int16Array: 2,
-  Uint16Array: 2,
-  Int32Array: 4,
-  Uint32Array: 4,
-  Float32Array: 4,
-  Float64Array: 8
-};
-
-var BigIntArrayConstructorsList = {
-  BigInt64Array: 8,
-  BigUint64Array: 8
-};
-
-var isView = function isView(it) {
-  if (!isObject(it)) return false;
-  var klass = classof(it);
-  return klass === 'DataView'
-    || hasOwnProperty_1(TypedArrayConstructorsList, klass)
-    || hasOwnProperty_1(BigIntArrayConstructorsList, klass);
-};
-
-var isTypedArray = function (it) {
-  if (!isObject(it)) return false;
-  var klass = classof(it);
-  return hasOwnProperty_1(TypedArrayConstructorsList, klass)
-    || hasOwnProperty_1(BigIntArrayConstructorsList, klass);
-};
-
-var aTypedArray = function (it) {
-  if (isTypedArray(it)) return it;
-  throw TypeError$f('Target is not a typed array');
-};
-
-var aTypedArrayConstructor = function (C) {
-  if (isCallable(C) && (!objectSetPrototypeOf || objectIsPrototypeOf(TypedArray, C))) return C;
-  throw TypeError$f(tryToString(C) + ' is not a typed array constructor');
-};
-
-var exportTypedArrayMethod = function (KEY, property, forced, options) {
-  if (!descriptors) return;
-  if (forced) for (var ARRAY in TypedArrayConstructorsList) {
-    var TypedArrayConstructor = global_1[ARRAY];
-    if (TypedArrayConstructor && hasOwnProperty_1(TypedArrayConstructor.prototype, KEY)) try {
-      delete TypedArrayConstructor.prototype[KEY];
-    } catch (error) {
-      // old WebKit bug - some methods are non-configurable
-      try {
-        TypedArrayConstructor.prototype[KEY] = property;
-      } catch (error2) { /* empty */ }
-    }
-  }
-  if (!TypedArrayPrototype[KEY] || forced) {
-    redefine(TypedArrayPrototype, KEY, forced ? property
-      : NATIVE_ARRAY_BUFFER_VIEWS && Int8ArrayPrototype[KEY] || property, options);
-  }
-};
-
-var exportTypedArrayStaticMethod = function (KEY, property, forced) {
-  var ARRAY, TypedArrayConstructor;
-  if (!descriptors) return;
-  if (objectSetPrototypeOf) {
-    if (forced) for (ARRAY in TypedArrayConstructorsList) {
-      TypedArrayConstructor = global_1[ARRAY];
-      if (TypedArrayConstructor && hasOwnProperty_1(TypedArrayConstructor, KEY)) try {
-        delete TypedArrayConstructor[KEY];
-      } catch (error) { /* empty */ }
-    }
-    if (!TypedArray[KEY] || forced) {
-      // V8 ~ Chrome 49-50 `%TypedArray%` methods are non-writable non-configurable
-      try {
-        return redefine(TypedArray, KEY, forced ? property : NATIVE_ARRAY_BUFFER_VIEWS && TypedArray[KEY] || property);
-      } catch (error) { /* empty */ }
-    } else return;
-  }
-  for (ARRAY in TypedArrayConstructorsList) {
-    TypedArrayConstructor = global_1[ARRAY];
-    if (TypedArrayConstructor && (!TypedArrayConstructor[KEY] || forced)) {
-      redefine(TypedArrayConstructor, KEY, property);
-    }
-  }
-};
-
-for (NAME in TypedArrayConstructorsList) {
-  Constructor = global_1[NAME];
-  Prototype = Constructor && Constructor.prototype;
-  if (Prototype) createNonEnumerableProperty(Prototype, TYPED_ARRAY_CONSTRUCTOR, Constructor);
-  else NATIVE_ARRAY_BUFFER_VIEWS = false;
-}
-
-for (NAME in BigIntArrayConstructorsList) {
-  Constructor = global_1[NAME];
-  Prototype = Constructor && Constructor.prototype;
-  if (Prototype) createNonEnumerableProperty(Prototype, TYPED_ARRAY_CONSTRUCTOR, Constructor);
-}
-
-// WebKit bug - typed arrays constructors prototype is Object.prototype
-if (!NATIVE_ARRAY_BUFFER_VIEWS || !isCallable(TypedArray) || TypedArray === Function.prototype) {
-  // eslint-disable-next-line no-shadow -- safe
-  TypedArray = function TypedArray() {
-    throw TypeError$f('Incorrect invocation');
-  };
-  if (NATIVE_ARRAY_BUFFER_VIEWS) for (NAME in TypedArrayConstructorsList) {
-    if (global_1[NAME]) objectSetPrototypeOf(global_1[NAME], TypedArray);
-  }
-}
-
-if (!NATIVE_ARRAY_BUFFER_VIEWS || !TypedArrayPrototype || TypedArrayPrototype === ObjectPrototype$1) {
-  TypedArrayPrototype = TypedArray.prototype;
-  if (NATIVE_ARRAY_BUFFER_VIEWS) for (NAME in TypedArrayConstructorsList) {
-    if (global_1[NAME]) objectSetPrototypeOf(global_1[NAME].prototype, TypedArrayPrototype);
-  }
-}
-
-// WebKit bug - one more object in Uint8ClampedArray prototype chain
-if (NATIVE_ARRAY_BUFFER_VIEWS && objectGetPrototypeOf(Uint8ClampedArrayPrototype) !== TypedArrayPrototype) {
-  objectSetPrototypeOf(Uint8ClampedArrayPrototype, TypedArrayPrototype);
-}
-
-if (descriptors && !hasOwnProperty_1(TypedArrayPrototype, TO_STRING_TAG$4)) {
-  TYPED_ARRAY_TAG_REQUIRED = true;
-  defineProperty$3(TypedArrayPrototype, TO_STRING_TAG$4, { get: function () {
-    return isObject(this) ? this[TYPED_ARRAY_TAG] : undefined;
-  } });
-  for (NAME in TypedArrayConstructorsList) if (global_1[NAME]) {
-    createNonEnumerableProperty(global_1[NAME], TYPED_ARRAY_TAG, NAME);
-  }
-}
-
-var arrayBufferViewCore = {
-  NATIVE_ARRAY_BUFFER_VIEWS: NATIVE_ARRAY_BUFFER_VIEWS,
-  TYPED_ARRAY_CONSTRUCTOR: TYPED_ARRAY_CONSTRUCTOR,
-  TYPED_ARRAY_TAG: TYPED_ARRAY_TAG_REQUIRED && TYPED_ARRAY_TAG,
-  aTypedArray: aTypedArray,
-  aTypedArrayConstructor: aTypedArrayConstructor,
-  exportTypedArrayMethod: exportTypedArrayMethod,
-  exportTypedArrayStaticMethod: exportTypedArrayStaticMethod,
-  isView: isView,
-  isTypedArray: isTypedArray,
-  TypedArray: TypedArray,
-  TypedArrayPrototype: TypedArrayPrototype
-};
-
-/* eslint-disable no-new -- required for testing */
-
-
-
-var NATIVE_ARRAY_BUFFER_VIEWS$1 = arrayBufferViewCore.NATIVE_ARRAY_BUFFER_VIEWS;
-
-var ArrayBuffer$1 = global_1.ArrayBuffer;
-var Int8Array$1 = global_1.Int8Array;
-
-var typedArrayConstructorsRequireWrappers = !NATIVE_ARRAY_BUFFER_VIEWS$1 || !fails(function () {
-  Int8Array$1(1);
-}) || !fails(function () {
-  new Int8Array$1(-1);
-}) || !checkCorrectnessOfIteration(function (iterable) {
-  new Int8Array$1();
-  new Int8Array$1(null);
-  new Int8Array$1(1.5);
-  new Int8Array$1(iterable);
-}, true) || fails(function () {
-  // Safari (11+) bug - a reason why even Safari 13 should load a typed array polyfill
-  return new Int8Array$1(new ArrayBuffer$1(2), 1, undefined).length !== 1;
-});
-
-var redefineAll = function (target, src, options) {
-  for (var key in src) redefine(target, key, src[key], options);
-  return target;
-};
-
-var TypeError$g = global_1.TypeError;
-
-var anInstance = function (it, Prototype) {
-  if (objectIsPrototypeOf(Prototype, it)) return it;
-  throw TypeError$g('Incorrect invocation');
-};
-
-var RangeError = global_1.RangeError;
-
-// `ToIndex` abstract operation
-// https://tc39.es/ecma262/#sec-toindex
-var toIndex = function (it) {
-  if (it === undefined) return 0;
-  var number = toIntegerOrInfinity(it);
-  var length = toLength(number);
-  if (number !== length) throw RangeError('Wrong length or index');
-  return length;
-};
-
-// IEEE754 conversions based on https://github.com/feross/ieee754
-
-
-var Array$1 = global_1.Array;
-var abs = Math.abs;
-var pow = Math.pow;
-var floor$1 = Math.floor;
-var log = Math.log;
-var LN2 = Math.LN2;
-
-var pack = function (number, mantissaLength, bytes) {
-  var buffer = Array$1(bytes);
-  var exponentLength = bytes * 8 - mantissaLength - 1;
-  var eMax = (1 << exponentLength) - 1;
-  var eBias = eMax >> 1;
-  var rt = mantissaLength === 23 ? pow(2, -24) - pow(2, -77) : 0;
-  var sign = number < 0 || number === 0 && 1 / number < 0 ? 1 : 0;
-  var index = 0;
-  var exponent, mantissa, c;
-  number = abs(number);
-  // eslint-disable-next-line no-self-compare -- NaN check
-  if (number != number || number === Infinity) {
-    // eslint-disable-next-line no-self-compare -- NaN check
-    mantissa = number != number ? 1 : 0;
-    exponent = eMax;
-  } else {
-    exponent = floor$1(log(number) / LN2);
-    c = pow(2, -exponent);
-    if (number * c < 1) {
-      exponent--;
-      c *= 2;
-    }
-    if (exponent + eBias >= 1) {
-      number += rt / c;
-    } else {
-      number += rt * pow(2, 1 - eBias);
-    }
-    if (number * c >= 2) {
-      exponent++;
-      c /= 2;
-    }
-    if (exponent + eBias >= eMax) {
-      mantissa = 0;
-      exponent = eMax;
-    } else if (exponent + eBias >= 1) {
-      mantissa = (number * c - 1) * pow(2, mantissaLength);
-      exponent = exponent + eBias;
-    } else {
-      mantissa = number * pow(2, eBias - 1) * pow(2, mantissaLength);
-      exponent = 0;
-    }
-  }
-  while (mantissaLength >= 8) {
-    buffer[index++] = mantissa & 255;
-    mantissa /= 256;
-    mantissaLength -= 8;
-  }
-  exponent = exponent << mantissaLength | mantissa;
-  exponentLength += mantissaLength;
-  while (exponentLength > 0) {
-    buffer[index++] = exponent & 255;
-    exponent /= 256;
-    exponentLength -= 8;
-  }
-  buffer[--index] |= sign * 128;
-  return buffer;
-};
-
-var unpack = function (buffer, mantissaLength) {
-  var bytes = buffer.length;
-  var exponentLength = bytes * 8 - mantissaLength - 1;
-  var eMax = (1 << exponentLength) - 1;
-  var eBias = eMax >> 1;
-  var nBits = exponentLength - 7;
-  var index = bytes - 1;
-  var sign = buffer[index--];
-  var exponent = sign & 127;
-  var mantissa;
-  sign >>= 7;
-  while (nBits > 0) {
-    exponent = exponent * 256 + buffer[index--];
-    nBits -= 8;
-  }
-  mantissa = exponent & (1 << -nBits) - 1;
-  exponent >>= -nBits;
-  nBits += mantissaLength;
-  while (nBits > 0) {
-    mantissa = mantissa * 256 + buffer[index--];
-    nBits -= 8;
-  }
-  if (exponent === 0) {
-    exponent = 1 - eBias;
-  } else if (exponent === eMax) {
-    return mantissa ? NaN : sign ? -Infinity : Infinity;
-  } else {
-    mantissa = mantissa + pow(2, mantissaLength);
-    exponent = exponent - eBias;
-  } return (sign ? -1 : 1) * mantissa * pow(2, exponent - mantissaLength);
-};
-
-var ieee754 = {
-  pack: pack,
-  unpack: unpack
-};
-
-// `Array.prototype.fill` method implementation
-// https://tc39.es/ecma262/#sec-array.prototype.fill
-var arrayFill = function fill(value /* , start = 0, end = @length */) {
-  var O = toObject(this);
-  var length = lengthOfArrayLike(O);
-  var argumentsLength = arguments.length;
-  var index = toAbsoluteIndex(argumentsLength > 1 ? arguments[1] : undefined, length);
-  var end = argumentsLength > 2 ? arguments[2] : undefined;
-  var endPos = end === undefined ? length : toAbsoluteIndex(end, length);
-  while (endPos > index) O[index++] = value;
-  return O;
-};
-
-var createProperty = function (object, key, value) {
-  var propertyKey = toPropertyKey(key);
-  if (propertyKey in object) objectDefineProperty.f(object, propertyKey, createPropertyDescriptor(0, value));
-  else object[propertyKey] = value;
-};
-
-var Array$2 = global_1.Array;
-var max$1 = Math.max;
-
-var arraySliceSimple = function (O, start, end) {
-  var length = lengthOfArrayLike(O);
-  var k = toAbsoluteIndex(start, length);
-  var fin = toAbsoluteIndex(end === undefined ? length : end, length);
-  var result = Array$2(max$1(fin - k, 0));
-  for (var n = 0; k < fin; k++, n++) createProperty(result, n, O[k]);
-  result.length = n;
-  return result;
-};
-
-var getOwnPropertyNames = objectGetOwnPropertyNames.f;
-var defineProperty$4 = objectDefineProperty.f;
-
-
-
-
-
-var PROPER_FUNCTION_NAME$1 = functionName.PROPER;
-var CONFIGURABLE_FUNCTION_NAME$1 = functionName.CONFIGURABLE;
-var getInternalState$1 = internalState.get;
-var setInternalState$1 = internalState.set;
-var ARRAY_BUFFER = 'ArrayBuffer';
-var DATA_VIEW = 'DataView';
-var PROTOTYPE$1 = 'prototype';
-var WRONG_LENGTH = 'Wrong length';
-var WRONG_INDEX = 'Wrong index';
-var NativeArrayBuffer = global_1[ARRAY_BUFFER];
-var $ArrayBuffer = NativeArrayBuffer;
-var ArrayBufferPrototype = $ArrayBuffer && $ArrayBuffer[PROTOTYPE$1];
-var $DataView = global_1[DATA_VIEW];
-var DataViewPrototype = $DataView && $DataView[PROTOTYPE$1];
-var ObjectPrototype$2 = Object.prototype;
-var Array$3 = global_1.Array;
-var RangeError$1 = global_1.RangeError;
-var fill = functionUncurryThis(arrayFill);
-var reverse = functionUncurryThis([].reverse);
-
-var packIEEE754 = ieee754.pack;
-var unpackIEEE754 = ieee754.unpack;
-
-var packInt8 = function (number) {
-  return [number & 0xFF];
-};
-
-var packInt16 = function (number) {
-  return [number & 0xFF, number >> 8 & 0xFF];
-};
-
-var packInt32 = function (number) {
-  return [number & 0xFF, number >> 8 & 0xFF, number >> 16 & 0xFF, number >> 24 & 0xFF];
-};
-
-var unpackInt32 = function (buffer) {
-  return buffer[3] << 24 | buffer[2] << 16 | buffer[1] << 8 | buffer[0];
-};
-
-var packFloat32 = function (number) {
-  return packIEEE754(number, 23, 4);
-};
-
-var packFloat64 = function (number) {
-  return packIEEE754(number, 52, 8);
-};
-
-var addGetter = function (Constructor, key) {
-  defineProperty$4(Constructor[PROTOTYPE$1], key, { get: function () { return getInternalState$1(this)[key]; } });
-};
-
-var get$1 = function (view, count, index, isLittleEndian) {
-  var intIndex = toIndex(index);
-  var store = getInternalState$1(view);
-  if (intIndex + count > store.byteLength) throw RangeError$1(WRONG_INDEX);
-  var bytes = getInternalState$1(store.buffer).bytes;
-  var start = intIndex + store.byteOffset;
-  var pack = arraySliceSimple(bytes, start, start + count);
-  return isLittleEndian ? pack : reverse(pack);
-};
-
-var set$1 = function (view, count, index, conversion, value, isLittleEndian) {
-  var intIndex = toIndex(index);
-  var store = getInternalState$1(view);
-  if (intIndex + count > store.byteLength) throw RangeError$1(WRONG_INDEX);
-  var bytes = getInternalState$1(store.buffer).bytes;
-  var start = intIndex + store.byteOffset;
-  var pack = conversion(+value);
-  for (var i = 0; i < count; i++) bytes[start + i] = pack[isLittleEndian ? i : count - i - 1];
-};
-
-if (!arrayBufferNative) {
-  $ArrayBuffer = function ArrayBuffer(length) {
-    anInstance(this, ArrayBufferPrototype);
-    var byteLength = toIndex(length);
-    setInternalState$1(this, {
-      bytes: fill(Array$3(byteLength), 0),
-      byteLength: byteLength
-    });
-    if (!descriptors) this.byteLength = byteLength;
-  };
-
-  ArrayBufferPrototype = $ArrayBuffer[PROTOTYPE$1];
-
-  $DataView = function DataView(buffer, byteOffset, byteLength) {
-    anInstance(this, DataViewPrototype);
-    anInstance(buffer, ArrayBufferPrototype);
-    var bufferLength = getInternalState$1(buffer).byteLength;
-    var offset = toIntegerOrInfinity(byteOffset);
-    if (offset < 0 || offset > bufferLength) throw RangeError$1('Wrong offset');
-    byteLength = byteLength === undefined ? bufferLength - offset : toLength(byteLength);
-    if (offset + byteLength > bufferLength) throw RangeError$1(WRONG_LENGTH);
-    setInternalState$1(this, {
-      buffer: buffer,
-      byteLength: byteLength,
-      byteOffset: offset
-    });
-    if (!descriptors) {
-      this.buffer = buffer;
-      this.byteLength = byteLength;
-      this.byteOffset = offset;
-    }
-  };
-
-  DataViewPrototype = $DataView[PROTOTYPE$1];
-
-  if (descriptors) {
-    addGetter($ArrayBuffer, 'byteLength');
-    addGetter($DataView, 'buffer');
-    addGetter($DataView, 'byteLength');
-    addGetter($DataView, 'byteOffset');
-  }
-
-  redefineAll(DataViewPrototype, {
-    getInt8: function getInt8(byteOffset) {
-      return get$1(this, 1, byteOffset)[0] << 24 >> 24;
-    },
-    getUint8: function getUint8(byteOffset) {
-      return get$1(this, 1, byteOffset)[0];
-    },
-    getInt16: function getInt16(byteOffset /* , littleEndian */) {
-      var bytes = get$1(this, 2, byteOffset, arguments.length > 1 ? arguments[1] : undefined);
-      return (bytes[1] << 8 | bytes[0]) << 16 >> 16;
-    },
-    getUint16: function getUint16(byteOffset /* , littleEndian */) {
-      var bytes = get$1(this, 2, byteOffset, arguments.length > 1 ? arguments[1] : undefined);
-      return bytes[1] << 8 | bytes[0];
-    },
-    getInt32: function getInt32(byteOffset /* , littleEndian */) {
-      return unpackInt32(get$1(this, 4, byteOffset, arguments.length > 1 ? arguments[1] : undefined));
-    },
-    getUint32: function getUint32(byteOffset /* , littleEndian */) {
-      return unpackInt32(get$1(this, 4, byteOffset, arguments.length > 1 ? arguments[1] : undefined)) >>> 0;
-    },
-    getFloat32: function getFloat32(byteOffset /* , littleEndian */) {
-      return unpackIEEE754(get$1(this, 4, byteOffset, arguments.length > 1 ? arguments[1] : undefined), 23);
-    },
-    getFloat64: function getFloat64(byteOffset /* , littleEndian */) {
-      return unpackIEEE754(get$1(this, 8, byteOffset, arguments.length > 1 ? arguments[1] : undefined), 52);
-    },
-    setInt8: function setInt8(byteOffset, value) {
-      set$1(this, 1, byteOffset, packInt8, value);
-    },
-    setUint8: function setUint8(byteOffset, value) {
-      set$1(this, 1, byteOffset, packInt8, value);
-    },
-    setInt16: function setInt16(byteOffset, value /* , littleEndian */) {
-      set$1(this, 2, byteOffset, packInt16, value, arguments.length > 2 ? arguments[2] : undefined);
-    },
-    setUint16: function setUint16(byteOffset, value /* , littleEndian */) {
-      set$1(this, 2, byteOffset, packInt16, value, arguments.length > 2 ? arguments[2] : undefined);
-    },
-    setInt32: function setInt32(byteOffset, value /* , littleEndian */) {
-      set$1(this, 4, byteOffset, packInt32, value, arguments.length > 2 ? arguments[2] : undefined);
-    },
-    setUint32: function setUint32(byteOffset, value /* , littleEndian */) {
-      set$1(this, 4, byteOffset, packInt32, value, arguments.length > 2 ? arguments[2] : undefined);
-    },
-    setFloat32: function setFloat32(byteOffset, value /* , littleEndian */) {
-      set$1(this, 4, byteOffset, packFloat32, value, arguments.length > 2 ? arguments[2] : undefined);
-    },
-    setFloat64: function setFloat64(byteOffset, value /* , littleEndian */) {
-      set$1(this, 8, byteOffset, packFloat64, value, arguments.length > 2 ? arguments[2] : undefined);
-    }
-  });
-} else {
-  var INCORRECT_ARRAY_BUFFER_NAME = PROPER_FUNCTION_NAME$1 && NativeArrayBuffer.name !== ARRAY_BUFFER;
-  /* eslint-disable no-new -- required for testing */
-  if (!fails(function () {
-    NativeArrayBuffer(1);
-  }) || !fails(function () {
-    new NativeArrayBuffer(-1);
-  }) || fails(function () {
-    new NativeArrayBuffer();
-    new NativeArrayBuffer(1.5);
-    new NativeArrayBuffer(NaN);
-    return INCORRECT_ARRAY_BUFFER_NAME && !CONFIGURABLE_FUNCTION_NAME$1;
-  })) {
-  /* eslint-enable no-new -- required for testing */
-    $ArrayBuffer = function ArrayBuffer(length) {
-      anInstance(this, ArrayBufferPrototype);
-      return new NativeArrayBuffer(toIndex(length));
-    };
-
-    $ArrayBuffer[PROTOTYPE$1] = ArrayBufferPrototype;
-
-    for (var keys$1 = getOwnPropertyNames(NativeArrayBuffer), j = 0, key; keys$1.length > j;) {
-      if (!((key = keys$1[j++]) in $ArrayBuffer)) {
-        createNonEnumerableProperty($ArrayBuffer, key, NativeArrayBuffer[key]);
-      }
-    }
-
-    ArrayBufferPrototype.constructor = $ArrayBuffer;
-  } else if (INCORRECT_ARRAY_BUFFER_NAME && CONFIGURABLE_FUNCTION_NAME$1) {
-    createNonEnumerableProperty(NativeArrayBuffer, 'name', ARRAY_BUFFER);
-  }
-
-  // WebKit bug - the same parent prototype for typed arrays and data view
-  if (objectSetPrototypeOf && objectGetPrototypeOf(DataViewPrototype) !== ObjectPrototype$2) {
-    objectSetPrototypeOf(DataViewPrototype, ObjectPrototype$2);
-  }
-
-  // iOS Safari 7.x bug
-  var testView = new $DataView(new $ArrayBuffer(2));
-  var $setInt8 = functionUncurryThis(DataViewPrototype.setInt8);
-  testView.setInt8(0, 2147483648);
-  testView.setInt8(1, 2147483649);
-  if (testView.getInt8(0) || !testView.getInt8(1)) redefineAll(DataViewPrototype, {
-    setInt8: function setInt8(byteOffset, value) {
-      $setInt8(this, byteOffset, value << 24 >> 24);
-    },
-    setUint8: function setUint8(byteOffset, value) {
-      $setInt8(this, byteOffset, value << 24 >> 24);
-    }
-  }, { unsafe: true });
-}
-
-setToStringTag($ArrayBuffer, ARRAY_BUFFER);
-setToStringTag($DataView, DATA_VIEW);
-
-var arrayBuffer = {
-  ArrayBuffer: $ArrayBuffer,
-  DataView: $DataView
-};
-
-var floor$2 = Math.floor;
-
-// `IsIntegralNumber` abstract operation
-// https://tc39.es/ecma262/#sec-isintegralnumber
-// eslint-disable-next-line es/no-number-isinteger -- safe
-var isIntegralNumber = Number.isInteger || function isInteger(it) {
-  return !isObject(it) && isFinite(it) && floor$2(it) === it;
-};
-
-var RangeError$2 = global_1.RangeError;
-
-var toPositiveInteger = function (it) {
-  var result = toIntegerOrInfinity(it);
-  if (result < 0) throw RangeError$2("The argument can't be less than 0");
-  return result;
-};
-
-var RangeError$3 = global_1.RangeError;
-
-var toOffset = function (it, BYTES) {
-  var offset = toPositiveInteger(it);
-  if (offset % BYTES) throw RangeError$3('Wrong offset');
-  return offset;
-};
-
-var aTypedArrayConstructor$1 = arrayBufferViewCore.aTypedArrayConstructor;
-
-var typedArrayFrom = function from(source /* , mapfn, thisArg */) {
-  var C = aConstructor(this);
-  var O = toObject(source);
-  var argumentsLength = arguments.length;
-  var mapfn = argumentsLength > 1 ? arguments[1] : undefined;
-  var mapping = mapfn !== undefined;
-  var iteratorMethod = getIteratorMethod(O);
-  var i, length, result, step, iterator, next;
-  if (iteratorMethod && !isArrayIteratorMethod(iteratorMethod)) {
-    iterator = getIterator(O, iteratorMethod);
-    next = iterator.next;
-    O = [];
-    while (!(step = functionCall(next, iterator)).done) {
-      O.push(step.value);
-    }
-  }
-  if (mapping && argumentsLength > 2) {
-    mapfn = functionBindContext(mapfn, arguments[2]);
-  }
-  length = lengthOfArrayLike(O);
-  result = new (aTypedArrayConstructor$1(C))(length);
-  for (i = 0; length > i; i++) {
-    result[i] = mapping ? mapfn(O[i], i) : O[i];
-  }
-  return result;
-};
-
-// `IsArray` abstract operation
-// https://tc39.es/ecma262/#sec-isarray
-// eslint-disable-next-line es/no-array-isarray -- safe
-var isArray = Array.isArray || function isArray(argument) {
-  return classofRaw(argument) == 'Array';
-};
-
-var SPECIES$1 = wellKnownSymbol('species');
-var Array$4 = global_1.Array;
-
-// a part of `ArraySpeciesCreate` abstract operation
-// https://tc39.es/ecma262/#sec-arrayspeciescreate
-var arraySpeciesConstructor = function (originalArray) {
-  var C;
-  if (isArray(originalArray)) {
-    C = originalArray.constructor;
-    // cross-realm fallback
-    if (isConstructor(C) && (C === Array$4 || isArray(C.prototype))) C = undefined;
-    else if (isObject(C)) {
-      C = C[SPECIES$1];
-      if (C === null) C = undefined;
-    }
-  } return C === undefined ? Array$4 : C;
-};
-
-// `ArraySpeciesCreate` abstract operation
-// https://tc39.es/ecma262/#sec-arrayspeciescreate
-var arraySpeciesCreate = function (originalArray, length) {
-  return new (arraySpeciesConstructor(originalArray))(length === 0 ? 0 : length);
-};
-
-var push$2 = functionUncurryThis([].push);
-
-// `Array.prototype.{ forEach, map, filter, some, every, find, findIndex, filterReject }` methods implementation
-var createMethod$1 = function (TYPE) {
-  var IS_MAP = TYPE == 1;
-  var IS_FILTER = TYPE == 2;
-  var IS_SOME = TYPE == 3;
-  var IS_EVERY = TYPE == 4;
-  var IS_FIND_INDEX = TYPE == 6;
-  var IS_FILTER_REJECT = TYPE == 7;
-  var NO_HOLES = TYPE == 5 || IS_FIND_INDEX;
-  return function ($this, callbackfn, that, specificCreate) {
-    var O = toObject($this);
-    var self = indexedObject(O);
-    var boundFunction = functionBindContext(callbackfn, that);
-    var length = lengthOfArrayLike(self);
-    var index = 0;
-    var create = specificCreate || arraySpeciesCreate;
-    var target = IS_MAP ? create($this, length) : IS_FILTER || IS_FILTER_REJECT ? create($this, 0) : undefined;
-    var value, result;
-    for (;length > index; index++) if (NO_HOLES || index in self) {
-      value = self[index];
-      result = boundFunction(value, index, O);
-      if (TYPE) {
-        if (IS_MAP) target[index] = result; // map
-        else if (result) switch (TYPE) {
-          case 3: return true;              // some
-          case 5: return value;             // find
-          case 6: return index;             // findIndex
-          case 2: push$2(target, value);      // filter
-        } else switch (TYPE) {
-          case 4: return false;             // every
-          case 7: push$2(target, value);      // filterReject
-        }
-      }
-    }
-    return IS_FIND_INDEX ? -1 : IS_SOME || IS_EVERY ? IS_EVERY : target;
-  };
-};
-
-var arrayIteration = {
-  // `Array.prototype.forEach` method
-  // https://tc39.es/ecma262/#sec-array.prototype.foreach
-  forEach: createMethod$1(0),
-  // `Array.prototype.map` method
-  // https://tc39.es/ecma262/#sec-array.prototype.map
-  map: createMethod$1(1),
-  // `Array.prototype.filter` method
-  // https://tc39.es/ecma262/#sec-array.prototype.filter
-  filter: createMethod$1(2),
-  // `Array.prototype.some` method
-  // https://tc39.es/ecma262/#sec-array.prototype.some
-  some: createMethod$1(3),
-  // `Array.prototype.every` method
-  // https://tc39.es/ecma262/#sec-array.prototype.every
-  every: createMethod$1(4),
-  // `Array.prototype.find` method
-  // https://tc39.es/ecma262/#sec-array.prototype.find
-  find: createMethod$1(5),
-  // `Array.prototype.findIndex` method
-  // https://tc39.es/ecma262/#sec-array.prototype.findIndex
-  findIndex: createMethod$1(6),
-  // `Array.prototype.filterReject` method
-  // https://github.com/tc39/proposal-array-filtering
-  filterReject: createMethod$1(7)
-};
-
-var SPECIES$2 = wellKnownSymbol('species');
-
-var setSpecies = function (CONSTRUCTOR_NAME) {
-  var Constructor = getBuiltIn(CONSTRUCTOR_NAME);
-  var defineProperty = objectDefineProperty.f;
-
-  if (descriptors && Constructor && !Constructor[SPECIES$2]) {
-    defineProperty(Constructor, SPECIES$2, {
-      configurable: true,
-      get: function () { return this; }
-    });
-  }
-};
-
-// makes subclassing work correct for wrapped built-ins
-var inheritIfRequired = function ($this, dummy, Wrapper) {
-  var NewTarget, NewTargetPrototype;
-  if (
-    // it can work only with native `setPrototypeOf`
-    objectSetPrototypeOf &&
-    // we haven't completely correct pre-ES6 way for getting `new.target`, so use this
-    isCallable(NewTarget = dummy.constructor) &&
-    NewTarget !== Wrapper &&
-    isObject(NewTargetPrototype = NewTarget.prototype) &&
-    NewTargetPrototype !== Wrapper.prototype
-  ) objectSetPrototypeOf($this, NewTargetPrototype);
-  return $this;
-};
-
-var typedArrayConstructor = createCommonjsModule(function (module) {
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-var getOwnPropertyNames = objectGetOwnPropertyNames.f;
-
-var forEach = arrayIteration.forEach;
-
-
-
-
-
-
-var getInternalState = internalState.get;
-var setInternalState = internalState.set;
-var nativeDefineProperty = objectDefineProperty.f;
-var nativeGetOwnPropertyDescriptor = objectGetOwnPropertyDescriptor.f;
-var round = Math.round;
-var RangeError = global_1.RangeError;
-var ArrayBuffer = arrayBuffer.ArrayBuffer;
-var ArrayBufferPrototype = ArrayBuffer.prototype;
-var DataView = arrayBuffer.DataView;
-var NATIVE_ARRAY_BUFFER_VIEWS = arrayBufferViewCore.NATIVE_ARRAY_BUFFER_VIEWS;
-var TYPED_ARRAY_CONSTRUCTOR = arrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
-var TYPED_ARRAY_TAG = arrayBufferViewCore.TYPED_ARRAY_TAG;
-var TypedArray = arrayBufferViewCore.TypedArray;
-var TypedArrayPrototype = arrayBufferViewCore.TypedArrayPrototype;
-var aTypedArrayConstructor = arrayBufferViewCore.aTypedArrayConstructor;
-var isTypedArray = arrayBufferViewCore.isTypedArray;
-var BYTES_PER_ELEMENT = 'BYTES_PER_ELEMENT';
-var WRONG_LENGTH = 'Wrong length';
-
-var fromList = function (C, list) {
-  aTypedArrayConstructor(C);
-  var index = 0;
-  var length = list.length;
-  var result = new C(length);
-  while (length > index) result[index] = list[index++];
-  return result;
-};
-
-var addGetter = function (it, key) {
-  nativeDefineProperty(it, key, { get: function () {
-    return getInternalState(this)[key];
-  } });
-};
-
-var isArrayBuffer = function (it) {
-  var klass;
-  return objectIsPrototypeOf(ArrayBufferPrototype, it) || (klass = classof(it)) == 'ArrayBuffer' || klass == 'SharedArrayBuffer';
-};
-
-var isTypedArrayIndex = function (target, key) {
-  return isTypedArray(target)
-    && !isSymbol(key)
-    && key in target
-    && isIntegralNumber(+key)
-    && key >= 0;
-};
-
-var wrappedGetOwnPropertyDescriptor = function getOwnPropertyDescriptor(target, key) {
-  key = toPropertyKey(key);
-  return isTypedArrayIndex(target, key)
-    ? createPropertyDescriptor(2, target[key])
-    : nativeGetOwnPropertyDescriptor(target, key);
-};
-
-var wrappedDefineProperty = function defineProperty(target, key, descriptor) {
-  key = toPropertyKey(key);
-  if (isTypedArrayIndex(target, key)
-    && isObject(descriptor)
-    && hasOwnProperty_1(descriptor, 'value')
-    && !hasOwnProperty_1(descriptor, 'get')
-    && !hasOwnProperty_1(descriptor, 'set')
-    // TODO: add validation descriptor w/o calling accessors
-    && !descriptor.configurable
-    && (!hasOwnProperty_1(descriptor, 'writable') || descriptor.writable)
-    && (!hasOwnProperty_1(descriptor, 'enumerable') || descriptor.enumerable)
-  ) {
-    target[key] = descriptor.value;
-    return target;
-  } return nativeDefineProperty(target, key, descriptor);
-};
-
-if (descriptors) {
-  if (!NATIVE_ARRAY_BUFFER_VIEWS) {
-    objectGetOwnPropertyDescriptor.f = wrappedGetOwnPropertyDescriptor;
-    objectDefineProperty.f = wrappedDefineProperty;
-    addGetter(TypedArrayPrototype, 'buffer');
-    addGetter(TypedArrayPrototype, 'byteOffset');
-    addGetter(TypedArrayPrototype, 'byteLength');
-    addGetter(TypedArrayPrototype, 'length');
-  }
-
-  _export({ target: 'Object', stat: true, forced: !NATIVE_ARRAY_BUFFER_VIEWS }, {
-    getOwnPropertyDescriptor: wrappedGetOwnPropertyDescriptor,
-    defineProperty: wrappedDefineProperty
-  });
-
-  module.exports = function (TYPE, wrapper, CLAMPED) {
-    var BYTES = TYPE.match(/\d+$/)[0] / 8;
-    var CONSTRUCTOR_NAME = TYPE + (CLAMPED ? 'Clamped' : '') + 'Array';
-    var GETTER = 'get' + TYPE;
-    var SETTER = 'set' + TYPE;
-    var NativeTypedArrayConstructor = global_1[CONSTRUCTOR_NAME];
-    var TypedArrayConstructor = NativeTypedArrayConstructor;
-    var TypedArrayConstructorPrototype = TypedArrayConstructor && TypedArrayConstructor.prototype;
-    var exported = {};
-
-    var getter = function (that, index) {
-      var data = getInternalState(that);
-      return data.view[GETTER](index * BYTES + data.byteOffset, true);
-    };
-
-    var setter = function (that, index, value) {
-      var data = getInternalState(that);
-      if (CLAMPED) value = (value = round(value)) < 0 ? 0 : value > 0xFF ? 0xFF : value & 0xFF;
-      data.view[SETTER](index * BYTES + data.byteOffset, value, true);
-    };
-
-    var addElement = function (that, index) {
-      nativeDefineProperty(that, index, {
-        get: function () {
-          return getter(this, index);
-        },
-        set: function (value) {
-          return setter(this, index, value);
-        },
-        enumerable: true
-      });
-    };
-
-    if (!NATIVE_ARRAY_BUFFER_VIEWS) {
-      TypedArrayConstructor = wrapper(function (that, data, offset, $length) {
-        anInstance(that, TypedArrayConstructorPrototype);
-        var index = 0;
-        var byteOffset = 0;
-        var buffer, byteLength, length;
-        if (!isObject(data)) {
-          length = toIndex(data);
-          byteLength = length * BYTES;
-          buffer = new ArrayBuffer(byteLength);
-        } else if (isArrayBuffer(data)) {
-          buffer = data;
-          byteOffset = toOffset(offset, BYTES);
-          var $len = data.byteLength;
-          if ($length === undefined) {
-            if ($len % BYTES) throw RangeError(WRONG_LENGTH);
-            byteLength = $len - byteOffset;
-            if (byteLength < 0) throw RangeError(WRONG_LENGTH);
-          } else {
-            byteLength = toLength($length) * BYTES;
-            if (byteLength + byteOffset > $len) throw RangeError(WRONG_LENGTH);
-          }
-          length = byteLength / BYTES;
-        } else if (isTypedArray(data)) {
-          return fromList(TypedArrayConstructor, data);
-        } else {
-          return functionCall(typedArrayFrom, TypedArrayConstructor, data);
-        }
-        setInternalState(that, {
-          buffer: buffer,
-          byteOffset: byteOffset,
-          byteLength: byteLength,
-          length: length,
-          view: new DataView(buffer)
-        });
-        while (index < length) addElement(that, index++);
-      });
-
-      if (objectSetPrototypeOf) objectSetPrototypeOf(TypedArrayConstructor, TypedArray);
-      TypedArrayConstructorPrototype = TypedArrayConstructor.prototype = objectCreate(TypedArrayPrototype);
-    } else if (typedArrayConstructorsRequireWrappers) {
-      TypedArrayConstructor = wrapper(function (dummy, data, typedArrayOffset, $length) {
-        anInstance(dummy, TypedArrayConstructorPrototype);
-        return inheritIfRequired(function () {
-          if (!isObject(data)) return new NativeTypedArrayConstructor(toIndex(data));
-          if (isArrayBuffer(data)) return $length !== undefined
-            ? new NativeTypedArrayConstructor(data, toOffset(typedArrayOffset, BYTES), $length)
-            : typedArrayOffset !== undefined
-              ? new NativeTypedArrayConstructor(data, toOffset(typedArrayOffset, BYTES))
-              : new NativeTypedArrayConstructor(data);
-          if (isTypedArray(data)) return fromList(TypedArrayConstructor, data);
-          return functionCall(typedArrayFrom, TypedArrayConstructor, data);
-        }(), dummy, TypedArrayConstructor);
-      });
-
-      if (objectSetPrototypeOf) objectSetPrototypeOf(TypedArrayConstructor, TypedArray);
-      forEach(getOwnPropertyNames(NativeTypedArrayConstructor), function (key) {
-        if (!(key in TypedArrayConstructor)) {
-          createNonEnumerableProperty(TypedArrayConstructor, key, NativeTypedArrayConstructor[key]);
-        }
-      });
-      TypedArrayConstructor.prototype = TypedArrayConstructorPrototype;
-    }
-
-    if (TypedArrayConstructorPrototype.constructor !== TypedArrayConstructor) {
-      createNonEnumerableProperty(TypedArrayConstructorPrototype, 'constructor', TypedArrayConstructor);
-    }
-
-    createNonEnumerableProperty(TypedArrayConstructorPrototype, TYPED_ARRAY_CONSTRUCTOR, TypedArrayConstructor);
-
-    if (TYPED_ARRAY_TAG) {
-      createNonEnumerableProperty(TypedArrayConstructorPrototype, TYPED_ARRAY_TAG, CONSTRUCTOR_NAME);
-    }
-
-    exported[CONSTRUCTOR_NAME] = TypedArrayConstructor;
-
-    _export({
-      global: true, forced: TypedArrayConstructor != NativeTypedArrayConstructor, sham: !NATIVE_ARRAY_BUFFER_VIEWS
-    }, exported);
-
-    if (!(BYTES_PER_ELEMENT in TypedArrayConstructor)) {
-      createNonEnumerableProperty(TypedArrayConstructor, BYTES_PER_ELEMENT, BYTES);
-    }
-
-    if (!(BYTES_PER_ELEMENT in TypedArrayConstructorPrototype)) {
-      createNonEnumerableProperty(TypedArrayConstructorPrototype, BYTES_PER_ELEMENT, BYTES);
-    }
-
-    setSpecies(CONSTRUCTOR_NAME);
-  };
-} else module.exports = function () { /* empty */ };
-});
-
-// `Int32Array` constructor
-// https://tc39.es/ecma262/#sec-typedarray-objects
-typedArrayConstructor('Int32', function (init) {
-  return function Int32Array(data, byteOffset, length) {
-    return init(this, data, byteOffset, length);
-  };
-});
-
-var aTypedArray$1 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$1 = arrayBufferViewCore.exportTypedArrayMethod;
-
-// `%TypedArray%.prototype.at` method
-// https://github.com/tc39/proposal-relative-indexing-method
-exportTypedArrayMethod$1('at', function at(index) {
-  var O = aTypedArray$1(this);
-  var len = lengthOfArrayLike(O);
-  var relativeIndex = toIntegerOrInfinity(index);
-  var k = relativeIndex >= 0 ? relativeIndex : len + relativeIndex;
-  return (k < 0 || k >= len) ? undefined : O[k];
-});
-
-var RangeError$4 = global_1.RangeError;
-var Int8Array$2 = global_1.Int8Array;
-var Int8ArrayPrototype$1 = Int8Array$2 && Int8Array$2.prototype;
-var $set = Int8ArrayPrototype$1 && Int8ArrayPrototype$1.set;
-var aTypedArray$2 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$2 = arrayBufferViewCore.exportTypedArrayMethod;
-
-var WORKS_WITH_OBJECTS_AND_GEERIC_ON_TYPED_ARRAYS = !fails(function () {
-  // eslint-disable-next-line es/no-typed-arrays -- required for testing
-  var array = new Uint8ClampedArray(2);
-  functionCall($set, array, { length: 1, 0: 3 }, 1);
-  return array[1] !== 3;
-});
-
-// https://bugs.chromium.org/p/v8/issues/detail?id=11294 and other
-var TO_OBJECT_BUG = WORKS_WITH_OBJECTS_AND_GEERIC_ON_TYPED_ARRAYS && arrayBufferViewCore.NATIVE_ARRAY_BUFFER_VIEWS && fails(function () {
-  var array = new Int8Array$2(2);
-  array.set(1);
-  array.set('2', 1);
-  return array[0] !== 0 || array[1] !== 2;
-});
-
-// `%TypedArray%.prototype.set` method
-// https://tc39.es/ecma262/#sec-%typedarray%.prototype.set
-exportTypedArrayMethod$2('set', function set(arrayLike /* , offset */) {
-  aTypedArray$2(this);
-  var offset = toOffset(arguments.length > 1 ? arguments[1] : undefined, 1);
-  var src = toObject(arrayLike);
-  if (WORKS_WITH_OBJECTS_AND_GEERIC_ON_TYPED_ARRAYS) return functionCall($set, this, src, offset);
-  var length = this.length;
-  var len = lengthOfArrayLike(src);
-  var index = 0;
-  if (len + offset > length) throw RangeError$4('Wrong length');
-  while (index < len) this[offset + index] = src[index++];
-}, !WORKS_WITH_OBJECTS_AND_GEERIC_ON_TYPED_ARRAYS || TO_OBJECT_BUG);
-
-var floor$3 = Math.floor;
-
-var mergeSort = function (array, comparefn) {
-  var length = array.length;
-  var middle = floor$3(length / 2);
-  return length < 8 ? insertionSort(array, comparefn) : merge(
-    array,
-    mergeSort(arraySliceSimple(array, 0, middle), comparefn),
-    mergeSort(arraySliceSimple(array, middle), comparefn),
-    comparefn
-  );
-};
-
-var insertionSort = function (array, comparefn) {
-  var length = array.length;
-  var i = 1;
-  var element, j;
-
-  while (i < length) {
-    j = i;
-    element = array[i];
-    while (j && comparefn(array[j - 1], element) > 0) {
-      array[j] = array[--j];
-    }
-    if (j !== i++) array[j] = element;
-  } return array;
-};
-
-var merge = function (array, left, right, comparefn) {
-  var llength = left.length;
-  var rlength = right.length;
-  var lindex = 0;
-  var rindex = 0;
-
-  while (lindex < llength || rindex < rlength) {
-    array[lindex + rindex] = (lindex < llength && rindex < rlength)
-      ? comparefn(left[lindex], right[rindex]) <= 0 ? left[lindex++] : right[rindex++]
-      : lindex < llength ? left[lindex++] : right[rindex++];
-  } return array;
-};
-
-var arraySort = mergeSort;
-
-var firefox = engineUserAgent.match(/firefox\/(\d+)/i);
-
-var engineFfVersion = !!firefox && +firefox[1];
-
-var engineIsIeOrEdge = /MSIE|Trident/.test(engineUserAgent);
-
-var webkit = engineUserAgent.match(/AppleWebKit\/(\d+)\./);
-
-var engineWebkitVersion = !!webkit && +webkit[1];
-
-var Array$5 = global_1.Array;
-var aTypedArray$3 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$3 = arrayBufferViewCore.exportTypedArrayMethod;
-var Uint16Array = global_1.Uint16Array;
-var un$Sort = Uint16Array && functionUncurryThis(Uint16Array.prototype.sort);
-
-// WebKit
-var ACCEPT_INCORRECT_ARGUMENTS = !!un$Sort && !(fails(function () {
-  un$Sort(new Uint16Array(2), null);
-}) && fails(function () {
-  un$Sort(new Uint16Array(2), {});
-}));
-
-var STABLE_SORT = !!un$Sort && !fails(function () {
-  // feature detection can be too slow, so check engines versions
-  if (engineV8Version) return engineV8Version < 74;
-  if (engineFfVersion) return engineFfVersion < 67;
-  if (engineIsIeOrEdge) return true;
-  if (engineWebkitVersion) return engineWebkitVersion < 602;
-
-  var array = new Uint16Array(516);
-  var expected = Array$5(516);
-  var index, mod;
-
-  for (index = 0; index < 516; index++) {
-    mod = index % 4;
-    array[index] = 515 - index;
-    expected[index] = index - 2 * mod + 3;
-  }
-
-  un$Sort(array, function (a, b) {
-    return (a / 4 | 0) - (b / 4 | 0);
-  });
-
-  for (index = 0; index < 516; index++) {
-    if (array[index] !== expected[index]) return true;
-  }
-});
-
-var getSortCompare = function (comparefn) {
-  return function (x, y) {
-    if (comparefn !== undefined) return +comparefn(x, y) || 0;
-    // eslint-disable-next-line no-self-compare -- NaN check
-    if (y !== y) return -1;
-    // eslint-disable-next-line no-self-compare -- NaN check
-    if (x !== x) return 1;
-    if (x === 0 && y === 0) return 1 / x > 0 && 1 / y < 0 ? 1 : -1;
-    return x > y;
-  };
-};
-
-// `%TypedArray%.prototype.sort` method
-// https://tc39.es/ecma262/#sec-%typedarray%.prototype.sort
-exportTypedArrayMethod$3('sort', function sort(comparefn) {
-  if (comparefn !== undefined) aCallable(comparefn);
-  if (STABLE_SORT) return un$Sort(this, comparefn);
-
-  return arraySort(aTypedArray$3(this), getSortCompare(comparefn));
-}, !STABLE_SORT || ACCEPT_INCORRECT_ARGUMENTS);
-
-var arrayFromConstructorAndList = function (Constructor, list) {
-  var index = 0;
-  var length = lengthOfArrayLike(list);
-  var result = new Constructor(length);
-  while (length > index) result[index] = list[index++];
-  return result;
-};
-
-var TYPED_ARRAY_CONSTRUCTOR$1 = arrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
-var aTypedArrayConstructor$2 = arrayBufferViewCore.aTypedArrayConstructor;
-
-// a part of `TypedArraySpeciesCreate` abstract operation
-// https://tc39.es/ecma262/#typedarray-species-create
-var typedArraySpeciesConstructor = function (originalArray) {
-  return aTypedArrayConstructor$2(speciesConstructor(originalArray, originalArray[TYPED_ARRAY_CONSTRUCTOR$1]));
-};
-
-var typedArrayFromSpeciesAndList = function (instance, list) {
-  return arrayFromConstructorAndList(typedArraySpeciesConstructor(instance), list);
-};
-
-var $filterReject = arrayIteration.filterReject;
-
-
-var aTypedArray$4 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$4 = arrayBufferViewCore.exportTypedArrayMethod;
-
-// `%TypedArray%.prototype.filterReject` method
-// https://github.com/tc39/proposal-array-filtering
-exportTypedArrayMethod$4('filterReject', function filterReject(callbackfn /* , thisArg */) {
-  var list = $filterReject(aTypedArray$4(this), callbackfn, arguments.length > 1 ? arguments[1] : undefined);
-  return typedArrayFromSpeciesAndList(this, list);
-}, true);
-
-// `Array.prototype.{ findLast, findLastIndex }` methods implementation
-var createMethod$2 = function (TYPE) {
-  var IS_FIND_LAST_INDEX = TYPE == 1;
-  return function ($this, callbackfn, that) {
-    var O = toObject($this);
-    var self = indexedObject(O);
-    var boundFunction = functionBindContext(callbackfn, that);
-    var index = lengthOfArrayLike(self);
-    var value, result;
-    while (index-- > 0) {
-      value = self[index];
-      result = boundFunction(value, index, O);
-      if (result) switch (TYPE) {
-        case 0: return value; // findLast
-        case 1: return index; // findLastIndex
-      }
-    }
-    return IS_FIND_LAST_INDEX ? -1 : undefined;
-  };
-};
-
-var arrayIterationFromLast = {
-  // `Array.prototype.findLast` method
-  // https://github.com/tc39/proposal-array-find-from-last
-  findLast: createMethod$2(0),
-  // `Array.prototype.findLastIndex` method
-  // https://github.com/tc39/proposal-array-find-from-last
-  findLastIndex: createMethod$2(1)
-};
-
-var $findLast = arrayIterationFromLast.findLast;
-
-var aTypedArray$5 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$5 = arrayBufferViewCore.exportTypedArrayMethod;
-
-// `%TypedArray%.prototype.findLast` method
-// https://github.com/tc39/proposal-array-find-from-last
-exportTypedArrayMethod$5('findLast', function findLast(predicate /* , thisArg */) {
-  return $findLast(aTypedArray$5(this), predicate, arguments.length > 1 ? arguments[1] : undefined);
-});
-
-var $findLastIndex = arrayIterationFromLast.findLastIndex;
-
-var aTypedArray$6 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$6 = arrayBufferViewCore.exportTypedArrayMethod;
-
-// `%TypedArray%.prototype.findLastIndex` method
-// https://github.com/tc39/proposal-array-find-from-last
-exportTypedArrayMethod$6('findLastIndex', function findLastIndex(predicate /* , thisArg */) {
-  return $findLastIndex(aTypedArray$6(this), predicate, arguments.length > 1 ? arguments[1] : undefined);
-});
-
-var Array$6 = global_1.Array;
-var push$3 = functionUncurryThis([].push);
-
-var arrayGroupBy = function ($this, callbackfn, that, specificConstructor) {
-  var O = toObject($this);
-  var self = indexedObject(O);
-  var boundFunction = functionBindContext(callbackfn, that);
-  var target = objectCreate(null);
-  var length = lengthOfArrayLike(self);
-  var index = 0;
-  var Constructor, key, value;
-  for (;length > index; index++) {
-    value = self[index];
-    key = toPropertyKey(boundFunction(value, index, O));
-    // in some IE10 builds, `hasOwnProperty` returns incorrect result on integer keys
-    // but since it's a `null` prototype object, we can safely use `in`
-    if (key in target) push$3(target[key], value);
-    else target[key] = [value];
-  }
-  // TODO: Remove this block from `core-js@4`
-  if (specificConstructor) {
-    Constructor = specificConstructor(O);
-    if (Constructor !== Array$6) {
-      for (key in target) target[key] = arrayFromConstructorAndList(Constructor, target[key]);
-    }
-  } return target;
-};
-
-// TODO: Remove from `core-js@4`
-
-
-
-
-var aTypedArray$7 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$7 = arrayBufferViewCore.exportTypedArrayMethod;
-
-// `%TypedArray%.prototype.groupBy` method
-// https://github.com/tc39/proposal-array-grouping
-exportTypedArrayMethod$7('groupBy', function groupBy(callbackfn /* , thisArg */) {
-  var thisArg = arguments.length > 1 ? arguments[1] : undefined;
-  return arrayGroupBy(aTypedArray$7(this), callbackfn, thisArg, typedArraySpeciesConstructor);
-}, true);
-
-// https://tc39.es/proposal-change-array-by-copy/#sec-array.prototype.toReversed
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.toReversed
-var arrayToReversed = function (O, C) {
-  var len = lengthOfArrayLike(O);
-  var A = new C(len);
-  var k = 0;
-  for (; k < len; k++) A[k] = O[len - k - 1];
-  return A;
-};
-
-var aTypedArray$8 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$8 = arrayBufferViewCore.exportTypedArrayMethod;
-var TYPED_ARRAY_CONSTRUCTOR$2 = arrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
-
-// `%TypedArray%.prototype.toReversed` method
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.toReversed
-exportTypedArrayMethod$8('toReversed', function toReversed() {
-  return arrayToReversed(aTypedArray$8(this), this[TYPED_ARRAY_CONSTRUCTOR$2]);
-}, true);
-
-var aTypedArray$9 = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$9 = arrayBufferViewCore.exportTypedArrayMethod;
-var TYPED_ARRAY_CONSTRUCTOR$3 = arrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
-var sort = functionUncurryThis(arrayBufferViewCore.TypedArrayPrototype.sort);
-
-// `%TypedArray%.prototype.toSorted` method
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.toSorted
-exportTypedArrayMethod$9('toSorted', function toSorted(compareFn) {
-  if (compareFn !== undefined) aCallable(compareFn);
-  var O = aTypedArray$9(this);
-  var A = arrayFromConstructorAndList(O[TYPED_ARRAY_CONSTRUCTOR$3], O);
-  return sort(A, compareFn);
-}, true);
-
-var arraySlice = functionUncurryThis([].slice);
-
-var max$2 = Math.max;
-var min$2 = Math.min;
-
-// https://tc39.es/proposal-change-array-by-copy/#sec-array.prototype.toSpliced
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.toSpliced
-var arrayToSpliced = function (O, C, args) {
-  var start = args[0];
-  var deleteCount = args[1];
-  var len = lengthOfArrayLike(O);
-  var actualStart = toAbsoluteIndex(start, len);
-  var argumentsLength = args.length;
-  var k = 0;
-  var insertCount, actualDeleteCount, newLen, A;
-  if (argumentsLength === 0) {
-    insertCount = actualDeleteCount = 0;
-  } else if (argumentsLength === 1) {
-    insertCount = 0;
-    actualDeleteCount = len - actualStart;
-  } else {
-    insertCount = argumentsLength - 2;
-    actualDeleteCount = min$2(max$2(toIntegerOrInfinity(deleteCount), 0), len - actualStart);
-  }
-  newLen = len + insertCount - actualDeleteCount;
-  A = new C(newLen);
-
-  for (; k < actualStart; k++) A[k] = O[k];
-  for (; k < actualStart + insertCount; k++) A[k] = args[k - actualStart + 2];
-  for (; k < newLen; k++) A[k] = O[k + actualDeleteCount - insertCount];
-
-  return A;
-};
-
-var aTypedArray$a = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$a = arrayBufferViewCore.exportTypedArrayMethod;
-var TYPED_ARRAY_CONSTRUCTOR$4 = arrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
-
-// `%TypedArray%.prototype.toSpliced` method
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.toSpliced
-// eslint-disable-next-line no-unused-vars -- required for .length
-exportTypedArrayMethod$a('toSpliced', function toSpliced(start, deleteCount /* , ...items */) {
-  return arrayToSpliced(aTypedArray$a(this), this[TYPED_ARRAY_CONSTRUCTOR$4], arraySlice(arguments));
-}, true);
-
-var Map$1 = getBuiltIn('Map');
-var MapPrototype = Map$1.prototype;
-var mapForEach = functionUncurryThis(MapPrototype.forEach);
-var mapHas = functionUncurryThis(MapPrototype.has);
-var mapSet = functionUncurryThis(MapPrototype.set);
-var push$4 = functionUncurryThis([].push);
-
-// `Array.prototype.uniqueBy` method
-// https://github.com/tc39/proposal-array-unique
-var arrayUniqueBy = function uniqueBy(resolver) {
-  var that = toObject(this);
-  var length = lengthOfArrayLike(that);
-  var result = arraySpeciesCreate(that, 0);
-  var map = new Map$1();
-  var resolverFunction = resolver != null ? aCallable(resolver) : function (value) {
-    return value;
-  };
-  var index, item, key;
-  for (index = 0; index < length; index++) {
-    item = that[index];
-    key = resolverFunction(item);
-    if (!mapHas(map, key)) mapSet(map, key, item);
-  }
-  mapForEach(map, function (value) {
-    push$4(result, value);
-  });
-  return result;
-};
-
-var aTypedArray$b = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$b = arrayBufferViewCore.exportTypedArrayMethod;
-var arrayUniqueBy$1 = functionUncurryThis(arrayUniqueBy);
-
-// `%TypedArray%.prototype.uniqueBy` method
-// https://github.com/tc39/proposal-array-unique
-exportTypedArrayMethod$b('uniqueBy', function uniqueBy(resolver) {
-  return typedArrayFromSpeciesAndList(this, arrayUniqueBy$1(aTypedArray$b(this), resolver));
-}, true);
-
-var RangeError$5 = global_1.RangeError;
-
-// https://tc39.es/proposal-change-array-by-copy/#sec-array.prototype.with
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.with
-var arrayWith = function (O, C, index, value) {
-  var len = lengthOfArrayLike(O);
-  var relativeIndex = toIntegerOrInfinity(index);
-  var actualIndex = relativeIndex < 0 ? len + relativeIndex : relativeIndex;
-  if (actualIndex >= len || actualIndex < 0) throw RangeError$5('Incorrect index');
-  var A = new C(len);
-  var k = 0;
-  for (; k < len; k++) A[k] = k === actualIndex ? value : O[k];
-  return A;
-};
-
-var aTypedArray$c = arrayBufferViewCore.aTypedArray;
-var exportTypedArrayMethod$c = arrayBufferViewCore.exportTypedArrayMethod;
-var TYPED_ARRAY_CONSTRUCTOR$5 = arrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
-
-// `%TypedArray%.prototype.with` method
-// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.with
-exportTypedArrayMethod$c('with', { 'with': function (index, value) {
-  return arrayWith(aTypedArray$c(this), this[TYPED_ARRAY_CONSTRUCTOR$5], index, value);
-} }['with'], true);
-
-const OBJP = Object.getPrototypeOf({});
-const FN = "function";
-const STR = "string";
-const equiv = (a, b) => {
-  let proto;
-
-  if (a === b) {
-    return true;
-  }
-
-  if (a != null) {
-    if (typeof a.equiv === FN) {
-      return a.equiv(b);
-    }
-  } else {
-    return a == b;
-  }
-
-  if (b != null) {
-    if (typeof b.equiv === FN) {
-      return b.equiv(a);
-    }
-  } else {
-    return a == b;
-  }
-
-  if (typeof a === STR || typeof b === STR) {
-    return false;
-  }
-
-  if ((proto = Object.getPrototypeOf(a), proto == null || proto === OBJP) && (proto = Object.getPrototypeOf(b), proto == null || proto === OBJP)) {
-    return equivObject(a, b);
-  }
-
-  if (typeof a !== FN && a.length !== undefined && typeof b !== FN && b.length !== undefined) {
-    return equivArrayLike(a, b);
-  }
-
-  if (a instanceof Set && b instanceof Set) {
-    return equivSet(a, b);
-  }
-
-  if (a instanceof Map && b instanceof Map) {
-    return equivMap(a, b);
-  }
-
-  if (a instanceof Date && b instanceof Date) {
-    return a.getTime() === b.getTime();
-  }
-
-  if (a instanceof RegExp && b instanceof RegExp) {
-    return a.toString() === b.toString();
-  } // NaN
-
-
-  return a !== a && b !== b;
-};
-const equivArrayLike = (a, b, _equiv = equiv) => {
-  let l = a.length;
-
-  if (l === b.length) {
-    while (l-- > 0 && _equiv(a[l], b[l]));
-  }
-
-  return l < 0;
-};
-const equivSet = (a, b, _equiv = equiv) => a.size === b.size && _equiv([...a.keys()].sort(), [...b.keys()].sort());
-const equivMap = (a, b, _equiv = equiv) => a.size === b.size && _equiv([...a].sort(), [...b].sort());
-const equivObject = (a, b, _equiv = equiv) => {
-  if (Object.keys(a).length !== Object.keys(b).length) {
-    return false;
-  }
-
-  for (let k in a) {
-    if (!b.hasOwnProperty(k) || !_equiv(a[k], b[k])) {
-      return false;
-    }
-  }
-
-  return true;
-};
-
-let _cachedFP;
-
-let _cachedPath;
-
-let _cachedEPC = [];
-let _cachedPathPos = [];
-
-const cachedFP = size => _cachedFP && _cachedFP.length >= size ? _cachedFP : _cachedFP = new Int32Array(size);
-
-const cachedPath = size => _cachedPath && _cachedPath.length >= size ? _cachedPath : _cachedPath = new Int32Array(size);
-
-const simpleDiff = (state, src, key, logDir, mode) => {
-  const n = src.length;
-  const linear = state.linear;
-  state.distance = n;
-
-  if (mode !== "only-distance") {
-    for (let i = 0, j = 0; i < n; i++, j += 3) {
-      linear[j] = logDir;
-      linear[j + 1] = i;
-      linear[j + 2] = src[i];
-    }
-
-    if (mode === "full") {
-      const _state = state[key];
-
-      for (let i = 0; i < n; i++) {
-        _state[i] = src[i];
-      }
-    }
-  }
-
-  return state;
-};
-/**
- * Based on "An O(NP) Sequence Comparison Algorithm"" by Wu, Manber,
- * Myers and Miller.
- *
- * @remarks
- * Various optimizations, fixes & refactorings. By default uses
- * {@link @thi.ng/equiv#equiv} for equality checks.
- *
- * - {@link https://publications.mpi-cbg.de/Wu_1990_6334.pdf}
- * - {@link http://www.itu.dk/stud/speciale/bepjea/xwebtex/litt/an-onp-sequence-comparison-algorithm.pdf}
- * - {@link https://github.com/cubicdaiya/onp}
- *
- * @param a - "old" array
- * @param b - "new" array
- * @param mode - result mode
- * @param equiv - equality predicate function
- */
-
-
-const diffArray = (a, b, mode = "full", equiv$1 = equiv) => {
-  const state = {
-    distance: 0,
-    adds: {},
-    dels: {},
-    const: {},
-    linear: []
-  };
-
-  if (a === b || a == null && b == null) {
-    return state;
-  } else if (a == null || a.length === 0) {
-    return simpleDiff(state, b, "adds", 1, mode);
-  } else if (b == null || b.length === 0) {
-    return simpleDiff(state, a, "dels", -1, mode);
-  }
-
-  const reverse = a.length >= b.length;
-
-  let _a, _b, na, nb;
-
-  if (reverse) {
-    _a = b;
-    _b = a;
-  } else {
-    _a = a;
-    _b = b;
-  }
-
-  na = _a.length;
-  nb = _b.length;
-  const offset = na + 1;
-  const delta = nb - na;
-  const doff = delta + offset;
-  const size = na + nb + 3;
-  const path = cachedPath(size).fill(-1, 0, size);
-  const fp = cachedFP(size).fill(-1, 0, size);
-  const epc = _cachedEPC;
-  const pathPos = _cachedPathPos;
-
-  const snake = (k, p, pp) => {
-    let r, y;
-
-    if (p > pp) {
-      r = path[k - 1];
-      y = p;
-    } else {
-      r = path[k + 1];
-      y = pp;
-    }
-
-    let x = y - (k - offset);
-
-    while (x < na && y < nb && equiv$1(_a[x], _b[y])) {
-      x++;
-      y++;
-    }
-
-    path[k] = pathPos.length / 3;
-    pathPos.push(x, y, r);
-    fp[k] = y;
-  };
-
-  let p = -1;
-  let k;
-
-  do {
-    p++;
-
-    for (k = -p + offset; k < doff; k++) {
-      snake(k, fp[k - 1] + 1, fp[k + 1]);
-    }
-
-    for (k = doff + p; k > doff; k--) {
-      snake(k, fp[k - 1] + 1, fp[k + 1]);
-    }
-
-    snake(doff, fp[doff - 1] + 1, fp[doff + 1]);
-  } while (fp[doff] !== nb);
-
-  state.distance = delta + 2 * p;
-
-  if (mode !== "only-distance") {
-    p = path[doff] * 3;
-
-    while (p >= 0) {
-      epc.push(p);
-      p = pathPos[p + 2] * 3;
-    }
-
-    if (mode === "full") {
-      buildFullLog(epc, pathPos, state, _a, _b, reverse);
-    } else {
-      buildLinearLog(epc, pathPos, state, _a, _b, reverse, mode === "only-distance-linear");
-    }
-  }
-
-  epc.length = 0;
-  pathPos.length = 0;
-  return state;
-};
-
-const buildFullLog = (epc, pathPos, state, a, b, reverse) => {
-  const linear = state.linear;
-  const _const = state.const;
-  let i = epc.length;
-  let px = 0;
-  let py = 0;
-  let adds;
-  let dels;
-  let aID;
-  let dID;
-
-  if (reverse) {
-    adds = state.dels;
-    dels = state.adds;
-    aID = -1;
-    dID = 1;
-  } else {
-    adds = state.adds;
-    dels = state.dels;
-    aID = 1;
-    dID = -1;
-  }
-
-  for (; i-- > 0;) {
-    const e = epc[i];
-    const ppx = pathPos[e];
-    const ppy = pathPos[e + 1];
-    const d = ppy - ppx;
-
-    while (px < ppx || py < ppy) {
-      const dp = py - px;
-
-      if (d > dp) {
-        linear.push(aID, py, adds[py] = b[py]);
-        py++;
-      } else if (d < dp) {
-        linear.push(dID, px, dels[px] = a[px]);
-        px++;
-      } else {
-        linear.push(0, px, _const[px] = a[px]);
-        px++;
-        py++;
-      }
-    }
-  }
-};
-
-const buildLinearLog = (epc, pathPos, state, a, b, reverse, inclConst) => {
-  const linear = state.linear;
-  const aID = reverse ? -1 : 1;
-  const dID = reverse ? 1 : -1;
-  let i = epc.length,
-      px = 0,
-      py = 0;
-
-  for (; i-- > 0;) {
-    const e = epc[i];
-    const ppx = pathPos[e];
-    const ppy = pathPos[e + 1];
-    const d = ppy - ppx;
-
-    while (px < ppx || py < ppy) {
-      const dp = py - px;
-
-      if (d > dp) {
-        linear.push(aID, py, b[py]);
-        py++;
-      } else if (d < dp) {
-        linear.push(dID, px, a[px]);
-        px++;
-      } else {
-        inclConst && linear.push(0, px, a[px]);
-        px++;
-        py++;
-      }
-    }
-  }
-};
-
-const diffObject = (a, b, mode = "full", _equiv = equiv) => a === b ? {
-  distance: 0
-} : mode === "only-distance" ? diffObjectDist(a, b, _equiv) : diffObjectFull(a, b, _equiv);
-
-const diffObjectDist = (a, b, _equiv) => {
-  if (!a) a = {};
-  if (!b) b = {};
-  let d = 0;
-
-  for (let k in a) {
-    const vb = b[k];
-    (vb === undefined || !_equiv(a[k], vb)) && d++;
-  }
-
-  for (let k in b) {
-    !(k in a) && d++;
-  }
-
-  return {
-    distance: d
-  };
-};
-
-const diffObjectFull = (a, b, _equiv) => {
-  if (!a) a = {};
-  if (!b) b = {};
-  let d = 0;
-  const adds = [];
-  const dels = [];
-  const edits = [];
-
-  for (let k in a) {
-    const vb = b[k];
-
-    if (vb === undefined) {
-      dels.push(k);
-      d++;
-    } else if (!_equiv(a[k], vb)) {
-      edits.push(k, vb);
-      d++;
-    }
-  }
-
-  for (let k in b) {
-    if (!(k in a)) {
-      adds.push(k);
-      d++;
-    }
-  }
-
-  return {
-    distance: d,
-    adds,
-    dels,
-    edits
-  };
-};
-
-const isArray$1 = Array.isArray;
-const max$3 = Math.max;
-const OBJP$1 = Object.getPrototypeOf({});
-const FN$1 = "function";
-const STR$1 = "string"; // child index tracking template buffer
-
-const INDEX = (() => {
-  const res = new Array(2048);
-
-  for (let i = 2, n = res.length; i < n; i++) {
-    res[i] = i - 2;
-  }
-
-  return res;
-})();
-
-const buildIndex = n => {
-  if (n <= INDEX.length) {
-    return INDEX.slice(0, n);
-  }
-
-  const res = new Array(n);
-
-  while (n-- > 2) {
-    res[n] = n - 2;
-  }
-
-  return res;
-};
-/**
- * See {@link HDOMImplementation} interface for further details.
- *
- * @param opts - hdom config options
- * @param impl - hdom implementation
- * @param parent - parent element (DOM node)
- * @param prev - previous tree
- * @param curr - current tree
- * @param child - child index
- */
-
-
-const diffTree = (opts, impl, parent, prev, curr, child = 0) => {
-  const attribs = curr[1];
-
-  if (attribs.__skip) {
-    return;
-  } // always replace element if __diff = false
-
-
-  if (attribs.__diff === false) {
-    releaseTree(prev);
-    impl.replaceChild(opts, parent, child, curr);
-    return;
-  }
-
-  const pattribs = prev[1];
-
-  if (pattribs && pattribs.__skip) {
-    impl.replaceChild(opts, parent, child, curr, false);
-    return;
-  } // delegate to branch-local implementation
-
-
-  let _impl = attribs.__impl;
-
-  if (_impl && _impl !== impl) {
-    return _impl.diffTree(opts, _impl, parent, prev, curr, child);
-  }
-
-  const delta = diffArray(prev, curr, "only-distance-linear", equiv$1);
-
-  if (delta.distance === 0) {
-    return;
-  }
-
-  const edits = delta.linear;
-  const el = impl.getChild(parent, child);
-  let ii;
-  let status;
-  let val;
-
-  if (edits[0] !== 0 || prev[1].key !== attribs.key) {
-    // LOGGER.fine("replace:", prev, curr);
-    releaseTree(prev);
-    impl.replaceChild(opts, parent, child, curr);
-    return;
-  }
-
-  if ((val = prev.__release) && val !== curr.__release) {
-    releaseTree(prev);
-  }
-
-  if (edits[3] !== 0) {
-    diffAttributes(impl, el, prev[1], curr[1]); // if attribs changed & distance == 2 then we're done here...
-
-    if (delta.distance === 2) {
-      return;
-    }
-  }
-
-  const numEdits = edits.length;
-  const prevLength = prev.length - 1;
-  const equivKeys = extractEquivElements(edits);
-  const offsets = buildIndex(prevLength + 1);
-
-  for (ii = 6; ii < numEdits; ii += 3) {
-    status = edits[ii];
-    if (!status) continue;
-
-    if (status === -1) {
-      diffDeleted(opts, impl, el, prev, curr, edits, ii, equivKeys, offsets, prevLength);
-    } else {
-      diffAdded(opts, impl, el, edits, ii, equivKeys, offsets, prevLength);
-    }
-  } // call __init after all children have been added/updated
-
-
-  if ((val = curr.__init) && val != prev.__init) {
-    val.apply(curr, [el, ...curr.__args]);
-  }
-};
-
-const diffDeleted = (opts, impl, el, prev, curr, edits, ii, equivKeys, offsets, prevLength) => {
-  const val = edits[ii + 2];
-
-  if (isArray$1(val)) {
-    let k = val[1].key;
-
-    if (k !== undefined && equivKeys[k][2] !== undefined) {
-      const eq = equivKeys[k];
-      k = eq[0]; // LOGGER.fine(`diff equiv key @ ${k}:`, prev[k], curr[eq[2]]);
-
-      diffTree(opts, impl, el, prev[k], curr[eq[2]], offsets[k]);
-    } else {
-      const idx = edits[ii + 1]; // LOGGER.fine("remove @", offsets[idx], val);
-
-      releaseTree(val);
-      impl.removeChild(el, offsets[idx]);
-      incOffsets(offsets, prevLength, idx);
-    }
-  } else if (typeof val === STR$1) {
-    impl.setContent(el, "");
-  }
-};
-
-const diffAdded = (opts, impl, el, edits, ii, equivKeys, offsets, prevLength) => {
-  const val = edits[ii + 2];
-
-  if (typeof val === STR$1) {
-    impl.setContent(el, val);
-  } else if (isArray$1(val)) {
-    const k = val[1].key;
-
-    if (k === undefined || equivKeys[k][0] === undefined) {
-      const idx = edits[ii + 1]; // LOGGER.fine("insert @", offsets[idx], val);
-
-      impl.createTree(opts, el, val, offsets[idx]);
-      decOffsets(offsets, prevLength, idx);
-    }
-  }
-};
-
-const incOffsets = (offsets, j, idx) => {
-  for (; j > idx; j--) {
-    offsets[j] = max$3(offsets[j] - 1, 0);
-  }
-};
-
-const decOffsets = (offsets, j, idx) => {
-  for (; j >= idx; j--) {
-    offsets[j]++;
-  }
-};
-/**
- * Helper function for {@link diffTree} to compute & apply the difference
- * between a node's `prev` and `curr` attributes.
- *
- * @param impl - hdom implementation
- * @param el - DOM element
- * @param prev - previous attributes
- * @param curr - current attributes
- *
- * @internal
- */
-
-
-const diffAttributes = (impl, el, prev, curr) => {
-  const delta = diffObject(prev, curr, "full", equiv);
-  impl.removeAttribs(el, delta.dels, prev);
-  let val = SEMAPHORE;
-  let i, e, edits;
-
-  for (edits = delta.edits, i = edits.length; (i -= 2) >= 0;) {
-    e = edits[i];
-    e[0] === "o" && e[1] === "n" && impl.removeAttribs(el, [e], prev);
-    e !== "value" ? impl.setAttrib(el, e, edits[i + 1], curr) : val = edits[i + 1];
-  }
-
-  for (edits = delta.adds, i = edits.length; i-- > 0;) {
-    e = edits[i];
-    e !== "value" ? impl.setAttrib(el, e, curr[e], curr) : val = curr[e];
-  }
-
-  val !== SEMAPHORE && impl.setAttrib(el, "value", val, curr);
-};
-/**
- * Recursively attempts to call the {@link ILifecycle.release} lifecycle
- * method on every element in given tree (branch), using depth-first
- * descent. Each element is checked for the presence of the `__release`
- * control attribute. If (and only if) it is set to `false`, further
- * descent into that element's branch is skipped.
- *
- * @param tree - hdom sub-tree
- *
- * @internal
- */
-
-const releaseTree = tree => {
-  if (isArray$1(tree)) {
-    let x;
-
-    if ((x = tree[1]) && x.__release === false) {
-      return;
-    }
-
-    if (tree.__release) {
-      // LOGGER.fine("call __release", tag);
-      tree.__release.apply(tree.__this, tree.__args);
-
-      delete tree.__release;
-    }
-
-    for (x = tree.length; x-- > 2;) {
-      releaseTree(tree[x]);
-    }
-  }
-};
-
-const extractEquivElements = edits => {
-  let k;
-  let val;
-  let ek;
-  const equiv = {};
-
-  for (let i = edits.length; (i -= 3) >= 0;) {
-    val = edits[i + 2];
-
-    if (isArray$1(val) && (k = val[1].key) !== undefined) {
-      ek = equiv[k];
-      !ek && (equiv[k] = ek = [,,]);
-      ek[edits[i] + 1] = edits[i + 1];
-    }
-  }
-
-  return equiv;
-};
-/**
- * Customized version {@link @thi.ng/equiv#equiv} which takes `__diff`
- * attributes into account (at any nesting level). If an hdom element's
- * attribute object contains `__diff: false`, the object will ALWAYS be
- * considered unequal, even if all other attributes in the object are
- * equivalent.
- *
- * @param a -
- * @param b -
- *
- * @internal
- */
-
-
-const equiv$1 = (a, b) => {
-  let proto;
-
-  if (a === b) {
-    return true;
-  }
-
-  if (a != null) {
-    if (typeof a.equiv === FN$1) {
-      return a.equiv(b);
-    }
-  } else {
-    return a == b;
-  }
-
-  if (b != null) {
-    if (typeof b.equiv === FN$1) {
-      return b.equiv(a);
-    }
-  } else {
-    return a == b;
-  }
-
-  if (typeof a === STR$1 || typeof b === STR$1) {
-    return false;
-  }
-
-  if ((proto = Object.getPrototypeOf(a), proto == null || proto === OBJP$1) && (proto = Object.getPrototypeOf(b), proto == null || proto === OBJP$1)) {
-    return !(a.__diff === false || b.__diff === false) && equivObject(a, b, equiv$1);
-  }
-
-  if (typeof a !== FN$1 && a.length !== undefined && typeof b !== FN$1 && b.length !== undefined) {
-    return equivArrayLike(a, b, equiv$1);
-  }
-
-  if (a instanceof Set && b instanceof Set) {
-    return equivSet(a, b, equiv$1);
-  }
-
-  if (a instanceof Map && b instanceof Map) {
-    return equivMap(a, b, equiv$1);
-  }
-
-  if (a instanceof Date && b instanceof Date) {
-    return a.getTime() === b.getTime();
-  }
-
-  if (a instanceof RegExp && b instanceof RegExp) {
-    return a.toString() === b.toString();
-  } // NaN
-
-
-  return a !== a && b !== b;
-};
-
-const isArray$2 = Array.isArray;
-
-const isNotStringAndIterable = x => x != null && typeof x !== "string" && typeof x[Symbol.iterator] === "function";
-
-const isString = x => typeof x === "string";
-
-// https://github.com/tc39/proposal-iterator-helpers
-
-
-
-
-
-
-
-var Promise = getBuiltIn('Promise');
-var TypeError$h = global_1.TypeError;
-
-_export({ target: 'AsyncIterator', proto: true, real: true, forced: true }, {
-  reduce: function reduce(reducer /* , initialValue */) {
-    var iterator = anObject(this);
-    var next = aCallable(iterator.next);
-    var noInitial = arguments.length < 2;
-    var accumulator = noInitial ? undefined : arguments[1];
-    aCallable(reducer);
-
-    return new Promise(function (resolve, reject) {
-      var loop = function () {
-        try {
-          Promise.resolve(anObject(functionCall(next, iterator))).then(function (step) {
-            try {
-              if (anObject(step).done) {
-                noInitial ? reject(TypeError$h('Reduce of empty iterator with no initial value')) : resolve(accumulator);
-              } else {
-                var value = step.value;
-                if (noInitial) {
-                  noInitial = false;
-                  accumulator = value;
-                  loop();
-                } else {
-                  Promise.resolve(reducer(accumulator, value)).then(function (result) {
-                    accumulator = result;
-                    loop();
-                  }, reject);
-                }
-              }
-            } catch (err) { reject(err); }
-          }, reject);
-        } catch (error) { reject(error); }
-      };
-
-      loop();
-    });
-  }
-});
-
-// https://github.com/tc39/proposal-iterator-helpers
-
-
-
-
-
-
-
-
-var IteratorPrototype$3 = iteratorsCore.IteratorPrototype;
-
-
-var TO_STRING_TAG$5 = wellKnownSymbol('toStringTag');
-
-var NativeIterator = global_1.Iterator;
-
-// FF56- have non-standard global helper `Iterator`
-var FORCED =  !isCallable(NativeIterator)
-  || NativeIterator.prototype !== IteratorPrototype$3
-  // FF44- non-standard `Iterator` passes previous tests
-  || !fails(function () { NativeIterator({}); });
-
-var IteratorConstructor = function Iterator() {
-  anInstance(this, IteratorPrototype$3);
-};
-
-if (!hasOwnProperty_1(IteratorPrototype$3, TO_STRING_TAG$5)) {
-  createNonEnumerableProperty(IteratorPrototype$3, TO_STRING_TAG$5, 'Iterator');
-}
-
-if (FORCED || !hasOwnProperty_1(IteratorPrototype$3, 'constructor') || IteratorPrototype$3.constructor === Object) {
-  createNonEnumerableProperty(IteratorPrototype$3, 'constructor', IteratorConstructor);
-}
-
-IteratorConstructor.prototype = IteratorPrototype$3;
-
-_export({ global: true, forced: FORCED }, {
-  Iterator: IteratorConstructor
-});
-
-// https://github.com/tc39/proposal-iterator-helpers
-
-
-
-
-
-
-var TypeError$i = global_1.TypeError;
-
-_export({ target: 'Iterator', proto: true, real: true, forced: true }, {
-  reduce: function reduce(reducer /* , initialValue */) {
-    anObject(this);
-    aCallable(reducer);
-    var noInitial = arguments.length < 2;
-    var accumulator = noInitial ? undefined : arguments[1];
-    iterate(this, function (value) {
-      if (noInitial) {
-        noInitial = false;
-        accumulator = value;
-      } else {
-        accumulator = reducer(accumulator, value);
-      }
-    }, { IS_ITERATOR: true });
-    if (noInitial) throw TypeError$i('Reduce of empty iterator with no initial value');
-    return accumulator;
-  }
-});
-
-/** @internal */
-
-const RE_TAG = /^([^\s\.#]+)(?:#([^\s\.#]+))?(?:\.([^\s#]+))?$/;
-/** @internal */
-
-const NO_SPANS = {
-  button: 1,
-  option: 1,
-  script: 1,
-  style: 1,
-  text: 1,
-  textarea: 1,
-  title: 1
-};
-
-const tagMap = tags => tags.split(" ").reduce((acc, x) => (acc[x] = true, acc), {});
-/** @internal */
-// tslint:disable-next-line
-
-
-const SVG_TAGS = tagMap("animate animateColor animateMotion animateTransform circle clipPath color-profile defs desc discard ellipse feBlend feColorMatrix feComponentTransfer feComposite feConvolveMatrix feDiffuseLighting feDisplacementMap feDistantLight feDropShadow feFlood feFuncA feFuncB feFuncG feFuncR feGaussianBlur feImage feMerge feMergeNode feMorphology feOffset fePointLight feSpecularLighting feSpotLight feTile feTurbulence filter font foreignObject g image line linearGradient marker mask metadata mpath path pattern polygon polyline radialGradient rect set stop style svg switch symbol text textPath title tref tspan use view");
-/** @internal */
-// tslint:disable-next-line
-
-const VOID_TAGS = tagMap("area base br col command embed hr img input keygen link meta param source stop track use wbr ?xml");
-/** @internal */
-// tslint:disable-next-line
-
-const NO_CLOSE_EMPTY = tagMap("animate circle ellipse line path polygon polyline rect");
-
-/**
- * Returns true iff `x` implements {@link IDeref}.
- *
- * @param x -
- */
-const isDeref = x => x != null && typeof x["deref"] === "function";
-/**
- * If `x` implements {@link IDeref}, returns its wrapped value, else
- * returns `x` itself.
- *
- * @param x -
- */
-
-const deref = x => isDeref(x) ? x.deref() : x;
-
-const isFunction = x => typeof x === "function";
-
-const css = rules => {
-  let css = "";
-  let v;
-
-  for (let r in rules) {
-    v = deref(rules[r]);
-    isFunction(v) && (v = v(rules));
-    v != null && (css += `${r}:${v};`);
-  }
-
-  return css;
-};
-
-/**
- * Takes an object of RDF/XML prefixes and returns formatted string for
- * the RDFa `prefix` attribute.
- *
- * @example
- * ```ts
- * import { foaf, xsd } from "@thi.ng/prefixes";
- *
- * formatPrefixes({ foaf, xsd })
- * // "foaf: http://xmlns.com/foaf/0.1/ rdf: http://www.w3.org/2001/XMLSchema#"
- * ```
- *
- * @param prefixes -
- */
-const formatPrefixes = prefixes => Object.keys(prefixes).reduce((acc, k) => (acc.push(`${k}: ${prefixes[k]}`), acc), []).join(" ");
-
-const XML_SVG = "http://www.w3.org/2000/svg";
-
-const isArray$3 = isArray$2;
-const isNotStringAndIterable$1 = isNotStringAndIterable;
-const isString$1 = isString;
-
-const maybeInitElement = (el, tree) => tree.__init && tree.__init.apply(tree.__this, [el, ...tree.__args]);
-/**
- * See {@link HDOMImplementation} interface for further details.
- *
- * @param opts - hdom config options
- * @param parent - DOM element
- * @param tree - component tree
- * @param insert - child index
- */
-
-
-const createTree = (opts, impl, parent, tree, insert, init = true) => {
-  if (isArray$3(tree)) {
-    const tag = tree[0];
-
-    if (typeof tag === "function") {
-      return createTree(opts, impl, parent, tag.apply(null, [opts.ctx, ...tree.slice(1)]), insert);
-    }
-
-    const attribs = tree[1];
-
-    if (attribs.__impl) {
-      return attribs.__impl.createTree(opts, parent, tree, insert, init);
-    }
-
-    const el = impl.createElement(parent, tag, attribs, insert);
-
-    if (tree.length > 2) {
-      const n = tree.length;
-
-      for (let i = 2; i < n; i++) {
-        createTree(opts, impl, el, tree[i], undefined, init);
-      }
-    }
-
-    init && maybeInitElement(el, tree);
-    return el;
-  }
-
-  if (isNotStringAndIterable$1(tree)) {
-    const res = [];
-
-    for (let t of tree) {
-      res.push(createTree(opts, impl, parent, t, insert, init));
-    }
-
-    return res;
-  }
-
-  if (tree == null) {
-    return parent;
-  }
-
-  return impl.createTextElement(parent, tree);
-};
-/**
- * See {@link HDOMImplementation} interface for further details.
- *
- * @param opts - hdom config options
- * @param parent - DOM element
- * @param tree - component tree
- * @param index - child index
- */
-
-const hydrateTree = (opts, impl, parent, tree, index = 0) => {
-  if (isArray$3(tree)) {
-    const el = impl.getChild(parent, index);
-
-    if (typeof tree[0] === "function") {
-      hydrateTree(opts, impl, parent, tree[0].apply(null, [opts.ctx, ...tree.slice(1)]), index);
-    }
-
-    const attribs = tree[1];
-
-    if (attribs.__impl) {
-      return attribs.__impl.hydrateTree(opts, parent, tree, index);
-    }
-
-    maybeInitElement(el, tree);
-
-    for (let a in attribs) {
-      a[0] === "o" && a[1] === "n" && impl.setAttrib(el, a, attribs[a]);
-    }
-
-    for (let n = tree.length, i = 2; i < n; i++) {
-      hydrateTree(opts, impl, el, tree[i], i - 2);
-    }
-  } else if (isNotStringAndIterable$1(tree)) {
-    for (let t of tree) {
-      hydrateTree(opts, impl, parent, t, index);
-      index++;
-    }
-  }
-};
-/**
- * Creates a new DOM element of type `tag` with optional `attribs`. If
- * `parent` is not `null`, the new element will be inserted as child at
- * given `insert` index. If `insert` is missing, the element will be
- * appended to the `parent`'s list of children. Returns new DOM node.
- *
- * If `tag` is a known SVG element name, the new element will be created
- * with the proper SVG XML namespace.
- *
- * @param parent - DOM element
- * @param tag - component tree
- * @param attribs - attributes
- * @param insert - child index
- */
-
-const createElement = (parent, tag, attribs, insert) => {
-  const el = SVG_TAGS[tag] ? document.createElementNS(XML_SVG, tag) : document.createElement(tag);
-  attribs && setAttribs(el, attribs);
-  return addChild(parent, el, insert);
-};
-const createTextElement = (parent, content, insert) => addChild(parent, document.createTextNode(content), insert);
-const addChild = (parent, child, insert) => parent ? insert === undefined ? parent.appendChild(child) : parent.insertBefore(child, parent.children[insert]) : child;
-const getChild = (parent, child) => parent.children[child];
-const replaceChild = (opts, impl, parent, child, tree, init = true) => (impl.removeChild(parent, child), impl.createTree(opts, parent, tree, child, init));
-const cloneWithNewAttribs = (el, attribs) => {
-  const res = el.cloneNode(true);
-  setAttribs(res, attribs);
-  el.parentNode.replaceChild(res, el);
-  return res;
-};
-const setContent = (el, body) => el.textContent = body;
-const setAttribs = (el, attribs) => {
-  for (let k in attribs) {
-    setAttrib(el, k, attribs[k], attribs);
-  }
-
-  return el;
-};
-/**
- * Sets a single attribute on given element. If attrib name is NOT an
- * event name (prefix: "on") and its value is a function, it is called
- * with given `attribs` object (usually the full attrib object passed to
- * {@link setAttribs}) and the function's return value is used as the actual
- * attrib value.
- *
- * Special rules apply for certain attributes:
- *
- * - "style": delegated to {@link setStyle}
- * - "value": delegated to {@link updateValueAttrib}
- * - attrib IDs starting with "on" are treated as event listeners
- *
- * If the given (or computed) attrib value is `false` or `undefined` the
- * attrib is removed from the element.
- *
- * @param el - DOM element
- * @param id - attribute name
- * @param val - attribute value
- * @param attribs - object of all attribs
- */
-
-const setAttrib = (el, id, val, attribs) => {
-  if (id.startsWith("__")) return;
-  const isListener = id[0] === "o" && id[1] === "n";
-
-  if (!isListener && typeof val === "function") {
-    val = val(attribs);
-  }
-
-  if (val !== undefined && val !== false) {
-    switch (id) {
-      case "style":
-        setStyle(el, val);
-        break;
-
-      case "value":
-        updateValueAttrib(el, val);
-        break;
-
-      case "prefix":
-        el.setAttribute(id, isString$1(val) ? val : formatPrefixes(val));
-        break;
-
-      case "accesskey":
-        el.accessKey = val;
-        break;
-
-      case "contenteditable":
-        el.contentEditable = val;
-        break;
-
-      case "tabindex":
-        el.tabIndex = val;
-        break;
-
-      case "align":
-      case "autocapitalize":
-      case "checked":
-      case "dir":
-      case "draggable":
-      case "hidden":
-      case "id":
-      case "lang":
-      case "namespaceURI":
-      case "scrollTop":
-      case "scrollLeft":
-      case "title":
-        // TODO add more properties / enumerated attribs?
-        el[id] = val;
-        break;
-
-      default:
-        isListener ? setListener(el, id.substr(2), val) : el.setAttribute(id, val === true ? "" : val);
-    }
-  } else {
-    el[id] != null ? el[id] = null : el.removeAttribute(id);
-  }
-
-  return el;
-};
-/**
- * Updates an element's `value` property. For form elements it too
- * ensures the edit cursor retains its position.
- *
- * @param el - DOM element
- * @param value - value
- */
-
-const updateValueAttrib = (el, value) => {
-  let ev;
-
-  switch (el.type) {
-    case "text":
-    case "textarea":
-    case "password":
-    case "search":
-    case "number":
-    case "email":
-    case "url":
-    case "tel":
-    case "date":
-    case "datetime-local":
-    case "time":
-    case "week":
-    case "month":
-      if ((ev = el.value) !== undefined && typeof value === "string") {
-        const off = value.length - (ev.length - (el.selectionStart || 0));
-        el.value = value;
-        el.selectionStart = el.selectionEnd = off;
-        break;
-      }
-
-    default:
-      el.value = value;
-  }
-};
-const removeAttribs = (el, attribs, prev) => {
-  for (let i = attribs.length; i-- > 0;) {
-    const a = attribs[i];
-
-    if (a[0] === "o" && a[1] === "n") {
-      removeListener(el, a.substr(2), prev[a]);
-    } else {
-      el.hasAttribute(a) ? el.removeAttribute(a) : el[a] = null;
-    }
-  }
-};
-const setStyle = (el, styles) => (el.setAttribute("style", css(styles)), el);
-/**
- * Adds event listener (possibly with options).
- *
- * @param el - DOM element
- * @param id - event name (w/o `on` prefix)
- * @param listener -
- */
-
-const setListener = (el, id, listener) => isString$1(listener) ? el.setAttribute("on" + id, listener) : isArray$3(listener) ? el.addEventListener(id, ...listener) : el.addEventListener(id, listener);
-/**
- * Removes event listener (possibly with options).
- *
- * @param el - DOM element
- * @param id - event name (w/o `on` prefix)
- * @param listener -
- */
-
-const removeListener = (el, id, listener) => isArray$3(listener) ? el.removeEventListener(id, ...listener) : el.removeEventListener(id, listener);
-const clearDOM = el => el.innerHTML = "";
-const removeChild = (parent, childIdx) => {
-  const n = parent.children[childIdx];
-  n !== undefined && parent.removeChild(n);
-};
-
-const OBJP$2 = Object.getPrototypeOf;
-/**
- * Similar to {@link isObject}, but also checks if prototype is that of
- * `Object` (or `null`).
- *
- * @param x -
- */
-
-const isPlainObject = x => {
-  let p;
-  return x != null && typeof x === "object" && ((p = OBJP$2(x)) === null || OBJP$2(p) === null);
-};
-
-var FunctionPrototype$2 = Function.prototype;
-var apply = FunctionPrototype$2.apply;
-var call$2 = FunctionPrototype$2.call;
-
-// eslint-disable-next-line es/no-reflect -- safe
-var functionApply = typeof Reflect == 'object' && Reflect.apply || (functionBindNative ? call$2.bind(apply) : function () {
-  return call$2.apply(apply, arguments);
-});
-
-var normalizeStringArgument = function (argument, $default) {
-  return argument === undefined ? arguments.length < 2 ? '' : $default : toString_1(argument);
-};
-
-// `InstallErrorCause` abstract operation
-// https://tc39.es/proposal-error-cause/#sec-errorobjects-install-error-cause
-var installErrorCause = function (O, options) {
-  if (isObject(options) && 'cause' in options) {
-    createNonEnumerableProperty(O, 'cause', options.cause);
-  }
-};
-
-var replace = functionUncurryThis(''.replace);
-
-var TEST = (function (arg) { return String(Error(arg).stack); })('zxcasd');
-var V8_OR_CHAKRA_STACK_ENTRY = /\n\s*at [^:]*:[^\n]*/;
-var IS_V8_OR_CHAKRA_STACK = V8_OR_CHAKRA_STACK_ENTRY.test(TEST);
-
-var clearErrorStack = function (stack, dropEntries) {
-  if (IS_V8_OR_CHAKRA_STACK && typeof stack == 'string') {
-    while (dropEntries--) stack = replace(stack, V8_OR_CHAKRA_STACK_ENTRY, '');
-  } return stack;
-};
-
-var errorStackInstallable = !fails(function () {
-  var error = Error('a');
-  if (!('stack' in error)) return true;
-  // eslint-disable-next-line es/no-object-defineproperty -- safe
-  Object.defineProperty(error, 'stack', createPropertyDescriptor(1, 7));
-  return error.stack !== 7;
-});
-
-var wrapErrorConstructorWithCause = function (FULL_NAME, wrapper, FORCED, IS_AGGREGATE_ERROR) {
-  var OPTIONS_POSITION = IS_AGGREGATE_ERROR ? 2 : 1;
-  var path = FULL_NAME.split('.');
-  var ERROR_NAME = path[path.length - 1];
-  var OriginalError = getBuiltIn.apply(null, path);
-
-  if (!OriginalError) return;
-
-  var OriginalErrorPrototype = OriginalError.prototype;
-
-  // V8 9.3- bug https://bugs.chromium.org/p/v8/issues/detail?id=12006
-  if ( hasOwnProperty_1(OriginalErrorPrototype, 'cause')) delete OriginalErrorPrototype.cause;
-
-  if (!FORCED) return OriginalError;
-
-  var BaseError = getBuiltIn('Error');
-
-  var WrappedError = wrapper(function (a, b) {
-    var message = normalizeStringArgument(IS_AGGREGATE_ERROR ? b : a, undefined);
-    var result = IS_AGGREGATE_ERROR ? new OriginalError(a) : new OriginalError();
-    if (message !== undefined) createNonEnumerableProperty(result, 'message', message);
-    if (errorStackInstallable) createNonEnumerableProperty(result, 'stack', clearErrorStack(result.stack, 2));
-    if (this && objectIsPrototypeOf(OriginalErrorPrototype, this)) inheritIfRequired(result, this, WrappedError);
-    if (arguments.length > OPTIONS_POSITION) installErrorCause(result, arguments[OPTIONS_POSITION]);
-    return result;
-  });
-
-  WrappedError.prototype = OriginalErrorPrototype;
-
-  if (ERROR_NAME !== 'Error') {
-    if (objectSetPrototypeOf) objectSetPrototypeOf(WrappedError, BaseError);
-    else copyConstructorProperties(WrappedError, BaseError, { name: true });
-  }
-
-  copyConstructorProperties(WrappedError, OriginalError);
-
-  try {
-    // Safari 13- bug: WebAssembly errors does not have a proper `.name`
-    if (OriginalErrorPrototype.name !== ERROR_NAME) {
-      createNonEnumerableProperty(OriginalErrorPrototype, 'name', ERROR_NAME);
-    }
-    OriginalErrorPrototype.constructor = WrappedError;
-  } catch (error) { /* empty */ }
-
-  return WrappedError;
-};
-
-/* eslint-disable no-unused-vars -- required for functions `.length` */
-
-
-
-
-
-var WEB_ASSEMBLY = 'WebAssembly';
-var WebAssembly = global_1[WEB_ASSEMBLY];
-
-var FORCED$1 = Error('e', { cause: 7 }).cause !== 7;
-
-var exportGlobalErrorCauseWrapper = function (ERROR_NAME, wrapper) {
-  var O = {};
-  O[ERROR_NAME] = wrapErrorConstructorWithCause(ERROR_NAME, wrapper, FORCED$1);
-  _export({ global: true, forced: FORCED$1 }, O);
-};
-
-var exportWebAssemblyErrorCauseWrapper = function (ERROR_NAME, wrapper) {
-  if (WebAssembly && WebAssembly[ERROR_NAME]) {
-    var O = {};
-    O[ERROR_NAME] = wrapErrorConstructorWithCause(WEB_ASSEMBLY + '.' + ERROR_NAME, wrapper, FORCED$1);
-    _export({ target: WEB_ASSEMBLY, stat: true, forced: FORCED$1 }, O);
-  }
-};
-
-// https://github.com/tc39/proposal-error-cause
-exportGlobalErrorCauseWrapper('Error', function (init) {
-  return function Error(message) { return functionApply(init, this, arguments); };
-});
-exportGlobalErrorCauseWrapper('EvalError', function (init) {
-  return function EvalError(message) { return functionApply(init, this, arguments); };
-});
-exportGlobalErrorCauseWrapper('RangeError', function (init) {
-  return function RangeError(message) { return functionApply(init, this, arguments); };
-});
-exportGlobalErrorCauseWrapper('ReferenceError', function (init) {
-  return function ReferenceError(message) { return functionApply(init, this, arguments); };
-});
-exportGlobalErrorCauseWrapper('SyntaxError', function (init) {
-  return function SyntaxError(message) { return functionApply(init, this, arguments); };
-});
-exportGlobalErrorCauseWrapper('TypeError', function (init) {
-  return function TypeError(message) { return functionApply(init, this, arguments); };
-});
-exportGlobalErrorCauseWrapper('URIError', function (init) {
-  return function URIError(message) { return functionApply(init, this, arguments); };
-});
-exportWebAssemblyErrorCauseWrapper('CompileError', function (init) {
-  return function CompileError(message) { return functionApply(init, this, arguments); };
-});
-exportWebAssemblyErrorCauseWrapper('LinkError', function (init) {
-  return function LinkError(message) { return functionApply(init, this, arguments); };
-});
-exportWebAssemblyErrorCauseWrapper('RuntimeError', function (init) {
-  return function RuntimeError(message) { return functionApply(init, this, arguments); };
-});
-
-const defError = (prefix, suffix = msg => msg !== undefined ? ": " + msg : "") => class extends Error {
-  constructor(msg) {
-    super(prefix(msg) + suffix(msg));
-  }
-
-};
-
-const IllegalArgumentError = defError(() => "illegal argument(s)");
-const illegalArgs = msg => {
-  throw new IllegalArgumentError(msg);
-};
-
-// `RegExp.prototype.flags` getter implementation
-// https://tc39.es/ecma262/#sec-get-regexp.prototype.flags
-var regexpFlags = function () {
-  var that = anObject(this);
-  var result = '';
-  if (that.global) result += 'g';
-  if (that.ignoreCase) result += 'i';
-  if (that.multiline) result += 'm';
-  if (that.dotAll) result += 's';
-  if (that.unicode) result += 'u';
-  if (that.sticky) result += 'y';
-  return result;
-};
-
-// babel-minify and Closure Compiler transpiles RegExp('a', 'y') -> /a/y and it causes SyntaxError
-var $RegExp = global_1.RegExp;
-
-var UNSUPPORTED_Y = fails(function () {
-  var re = $RegExp('a', 'y');
-  re.lastIndex = 2;
-  return re.exec('abcd') != null;
-});
-
-// UC Browser bug
-// https://github.com/zloirock/core-js/issues/1008
-var MISSED_STICKY = UNSUPPORTED_Y || fails(function () {
-  return !$RegExp('a', 'y').sticky;
-});
-
-var BROKEN_CARET = UNSUPPORTED_Y || fails(function () {
-  // https://bugzilla.mozilla.org/show_bug.cgi?id=773687
-  var re = $RegExp('^r', 'gy');
-  re.lastIndex = 2;
-  return re.exec('str') != null;
-});
-
-var regexpStickyHelpers = {
-  BROKEN_CARET: BROKEN_CARET,
-  MISSED_STICKY: MISSED_STICKY,
-  UNSUPPORTED_Y: UNSUPPORTED_Y
-};
-
-// babel-minify and Closure Compiler transpiles RegExp('.', 's') -> /./s and it causes SyntaxError
-var $RegExp$1 = global_1.RegExp;
-
-var regexpUnsupportedDotAll = fails(function () {
-  var re = $RegExp$1('.', 's');
-  return !(re.dotAll && re.exec('\n') && re.flags === 's');
-});
-
-// babel-minify and Closure Compiler transpiles RegExp('(?<a>b)', 'g') -> /(?<a>b)/g and it causes SyntaxError
-var $RegExp$2 = global_1.RegExp;
-
-var regexpUnsupportedNcg = fails(function () {
-  var re = $RegExp$2('(?<a>b)', 'g');
-  return re.exec('b').groups.a !== 'b' ||
-    'b'.replace(re, '$<a>c') !== 'bc';
-});
-
-/* eslint-disable regexp/no-empty-capturing-group, regexp/no-empty-group, regexp/no-lazy-ends -- testing */
-/* eslint-disable regexp/no-useless-quantifier -- testing */
-
-
-
-
-
-
-
-var getInternalState$2 = internalState.get;
-
-
-
-var nativeReplace = shared('native-string-replace', String.prototype.replace);
-var nativeExec = RegExp.prototype.exec;
-var patchedExec = nativeExec;
-var charAt = functionUncurryThis(''.charAt);
-var indexOf$1 = functionUncurryThis(''.indexOf);
-var replace$1 = functionUncurryThis(''.replace);
-var stringSlice$1 = functionUncurryThis(''.slice);
-
-var UPDATES_LAST_INDEX_WRONG = (function () {
-  var re1 = /a/;
-  var re2 = /b*/g;
-  functionCall(nativeExec, re1, 'a');
-  functionCall(nativeExec, re2, 'a');
-  return re1.lastIndex !== 0 || re2.lastIndex !== 0;
-})();
-
-var UNSUPPORTED_Y$1 = regexpStickyHelpers.BROKEN_CARET;
-
-// nonparticipating capturing group, copied from es5-shim's String#split patch.
-var NPCG_INCLUDED = /()??/.exec('')[1] !== undefined;
-
-var PATCH = UPDATES_LAST_INDEX_WRONG || NPCG_INCLUDED || UNSUPPORTED_Y$1 || regexpUnsupportedDotAll || regexpUnsupportedNcg;
-
-if (PATCH) {
-  patchedExec = function exec(string) {
-    var re = this;
-    var state = getInternalState$2(re);
-    var str = toString_1(string);
-    var raw = state.raw;
-    var result, reCopy, lastIndex, match, i, object, group;
-
-    if (raw) {
-      raw.lastIndex = re.lastIndex;
-      result = functionCall(patchedExec, raw, str);
-      re.lastIndex = raw.lastIndex;
-      return result;
-    }
-
-    var groups = state.groups;
-    var sticky = UNSUPPORTED_Y$1 && re.sticky;
-    var flags = functionCall(regexpFlags, re);
-    var source = re.source;
-    var charsAdded = 0;
-    var strCopy = str;
-
-    if (sticky) {
-      flags = replace$1(flags, 'y', '');
-      if (indexOf$1(flags, 'g') === -1) {
-        flags += 'g';
-      }
-
-      strCopy = stringSlice$1(str, re.lastIndex);
-      // Support anchored sticky behavior.
-      if (re.lastIndex > 0 && (!re.multiline || re.multiline && charAt(str, re.lastIndex - 1) !== '\n')) {
-        source = '(?: ' + source + ')';
-        strCopy = ' ' + strCopy;
-        charsAdded++;
-      }
-      // ^(? + rx + ) is needed, in combination with some str slicing, to
-      // simulate the 'y' flag.
-      reCopy = new RegExp('^(?:' + source + ')', flags);
-    }
-
-    if (NPCG_INCLUDED) {
-      reCopy = new RegExp('^' + source + '$(?!\\s)', flags);
-    }
-    if (UPDATES_LAST_INDEX_WRONG) lastIndex = re.lastIndex;
-
-    match = functionCall(nativeExec, sticky ? reCopy : re, strCopy);
-
-    if (sticky) {
-      if (match) {
-        match.input = stringSlice$1(match.input, charsAdded);
-        match[0] = stringSlice$1(match[0], charsAdded);
-        match.index = re.lastIndex;
-        re.lastIndex += match[0].length;
-      } else re.lastIndex = 0;
-    } else if (UPDATES_LAST_INDEX_WRONG && match) {
-      re.lastIndex = re.global ? match.index + match[0].length : lastIndex;
-    }
-    if (NPCG_INCLUDED && match && match.length > 1) {
-      // Fix browsers whose `exec` methods don't consistently return `undefined`
-      // for NPCG, like IE8. NOTE: This doesn' work for /(.?)?/
-      functionCall(nativeReplace, match[0], reCopy, function () {
-        for (i = 1; i < arguments.length - 2; i++) {
-          if (arguments[i] === undefined) match[i] = undefined;
-        }
-      });
-    }
-
-    if (match && groups) {
-      match.groups = object = objectCreate(null);
-      for (i = 0; i < groups.length; i++) {
-        group = groups[i];
-        object[group[0]] = match[group[1]];
-      }
-    }
-
-    return match;
-  };
-}
-
-var regexpExec = patchedExec;
-
-// `RegExp.prototype.exec` method
-// https://tc39.es/ecma262/#sec-regexp.prototype.exec
-_export({ target: 'RegExp', proto: true, forced: /./.exec !== regexpExec }, {
-  exec: regexpExec
-});
-
-// TODO: Remove from `core-js@4` since it's moved to entry points
-
-
-
-
-
-
-
-
-var SPECIES$3 = wellKnownSymbol('species');
-var RegExpPrototype = RegExp.prototype;
-
-var fixRegexpWellKnownSymbolLogic = function (KEY, exec, FORCED, SHAM) {
-  var SYMBOL = wellKnownSymbol(KEY);
-
-  var DELEGATES_TO_SYMBOL = !fails(function () {
-    // String methods call symbol-named RegEp methods
-    var O = {};
-    O[SYMBOL] = function () { return 7; };
-    return ''[KEY](O) != 7;
-  });
-
-  var DELEGATES_TO_EXEC = DELEGATES_TO_SYMBOL && !fails(function () {
-    // Symbol-named RegExp methods call .exec
-    var execCalled = false;
-    var re = /a/;
-
-    if (KEY === 'split') {
-      // We can't use real regex here since it causes deoptimization
-      // and serious performance degradation in V8
-      // https://github.com/zloirock/core-js/issues/306
-      re = {};
-      // RegExp[@@split] doesn't call the regex's exec method, but first creates
-      // a new one. We need to return the patched regex when creating the new one.
-      re.constructor = {};
-      re.constructor[SPECIES$3] = function () { return re; };
-      re.flags = '';
-      re[SYMBOL] = /./[SYMBOL];
-    }
-
-    re.exec = function () { execCalled = true; return null; };
-
-    re[SYMBOL]('');
-    return !execCalled;
-  });
-
-  if (
-    !DELEGATES_TO_SYMBOL ||
-    !DELEGATES_TO_EXEC ||
-    FORCED
-  ) {
-    var uncurriedNativeRegExpMethod = functionUncurryThis(/./[SYMBOL]);
-    var methods = exec(SYMBOL, ''[KEY], function (nativeMethod, regexp, str, arg2, forceStringMethod) {
-      var uncurriedNativeMethod = functionUncurryThis(nativeMethod);
-      var $exec = regexp.exec;
-      if ($exec === regexpExec || $exec === RegExpPrototype.exec) {
-        if (DELEGATES_TO_SYMBOL && !forceStringMethod) {
-          // The native String method already delegates to @@method (this
-          // polyfilled function), leasing to infinite recursion.
-          // We avoid it by directly calling the native @@method method.
-          return { done: true, value: uncurriedNativeRegExpMethod(regexp, str, arg2) };
-        }
-        return { done: true, value: uncurriedNativeMethod(str, regexp, arg2) };
-      }
-      return { done: false };
-    });
-
-    redefine(String.prototype, KEY, methods[0]);
-    redefine(RegExpPrototype, SYMBOL, methods[1]);
-  }
-
-  if (SHAM) createNonEnumerableProperty(RegExpPrototype[SYMBOL], 'sham', true);
-};
-
-var charAt$1 = functionUncurryThis(''.charAt);
-var charCodeAt = functionUncurryThis(''.charCodeAt);
-var stringSlice$2 = functionUncurryThis(''.slice);
-
-var createMethod$3 = function (CONVERT_TO_STRING) {
-  return function ($this, pos) {
-    var S = toString_1(requireObjectCoercible($this));
-    var position = toIntegerOrInfinity(pos);
-    var size = S.length;
-    var first, second;
-    if (position < 0 || position >= size) return CONVERT_TO_STRING ? '' : undefined;
-    first = charCodeAt(S, position);
-    return first < 0xD800 || first > 0xDBFF || position + 1 === size
-      || (second = charCodeAt(S, position + 1)) < 0xDC00 || second > 0xDFFF
-        ? CONVERT_TO_STRING
-          ? charAt$1(S, position)
-          : first
-        : CONVERT_TO_STRING
-          ? stringSlice$2(S, position, position + 2)
-          : (first - 0xD800 << 10) + (second - 0xDC00) + 0x10000;
-  };
-};
-
-var stringMultibyte = {
-  // `String.prototype.codePointAt` method
-  // https://tc39.es/ecma262/#sec-string.prototype.codepointat
-  codeAt: createMethod$3(false),
-  // `String.prototype.at` method
-  // https://github.com/mathiasbynens/String.prototype.at
-  charAt: createMethod$3(true)
-};
-
-var charAt$2 = stringMultibyte.charAt;
-
-// `AdvanceStringIndex` abstract operation
-// https://tc39.es/ecma262/#sec-advancestringindex
-var advanceStringIndex = function (S, index, unicode) {
-  return index + (unicode ? charAt$2(S, index).length : 1);
-};
-
-var floor$4 = Math.floor;
-var charAt$3 = functionUncurryThis(''.charAt);
-var replace$2 = functionUncurryThis(''.replace);
-var stringSlice$3 = functionUncurryThis(''.slice);
-var SUBSTITUTION_SYMBOLS = /\$([$&'`]|\d{1,2}|<[^>]*>)/g;
-var SUBSTITUTION_SYMBOLS_NO_NAMED = /\$([$&'`]|\d{1,2})/g;
-
-// `GetSubstitution` abstract operation
-// https://tc39.es/ecma262/#sec-getsubstitution
-var getSubstitution = function (matched, str, position, captures, namedCaptures, replacement) {
-  var tailPos = position + matched.length;
-  var m = captures.length;
-  var symbols = SUBSTITUTION_SYMBOLS_NO_NAMED;
-  if (namedCaptures !== undefined) {
-    namedCaptures = toObject(namedCaptures);
-    symbols = SUBSTITUTION_SYMBOLS;
-  }
-  return replace$2(replacement, symbols, function (match, ch) {
-    var capture;
-    switch (charAt$3(ch, 0)) {
-      case '$': return '$';
-      case '&': return matched;
-      case '`': return stringSlice$3(str, 0, position);
-      case "'": return stringSlice$3(str, tailPos);
-      case '<':
-        capture = namedCaptures[stringSlice$3(ch, 1, -1)];
-        break;
-      default: // \d\d?
-        var n = +ch;
-        if (n === 0) return match;
-        if (n > m) {
-          var f = floor$4(n / 10);
-          if (f === 0) return match;
-          if (f <= m) return captures[f - 1] === undefined ? charAt$3(ch, 1) : captures[f - 1] + charAt$3(ch, 1);
-          return match;
-        }
-        capture = captures[n - 1];
-    }
-    return capture === undefined ? '' : capture;
-  });
-};
-
-var TypeError$j = global_1.TypeError;
-
-// `RegExpExec` abstract operation
-// https://tc39.es/ecma262/#sec-regexpexec
-var regexpExecAbstract = function (R, S) {
-  var exec = R.exec;
-  if (isCallable(exec)) {
-    var result = functionCall(exec, R, S);
-    if (result !== null) anObject(result);
-    return result;
-  }
-  if (classofRaw(R) === 'RegExp') return functionCall(regexpExec, R, S);
-  throw TypeError$j('RegExp#exec called on incompatible receiver');
-};
-
-var REPLACE = wellKnownSymbol('replace');
-var max$4 = Math.max;
-var min$3 = Math.min;
-var concat$1 = functionUncurryThis([].concat);
-var push$5 = functionUncurryThis([].push);
-var stringIndexOf = functionUncurryThis(''.indexOf);
-var stringSlice$4 = functionUncurryThis(''.slice);
-
-var maybeToString = function (it) {
-  return it === undefined ? it : String(it);
-};
-
-// IE <= 11 replaces $0 with the whole match, as if it was $&
-// https://stackoverflow.com/questions/6024666/getting-ie-to-replace-a-regex-with-the-literal-string-0
-var REPLACE_KEEPS_$0 = (function () {
-  // eslint-disable-next-line regexp/prefer-escape-replacement-dollar-char -- required for testing
-  return 'a'.replace(/./, '$0') === '$0';
-})();
-
-// Safari <= 13.0.3(?) substitutes nth capture where n>m with an empty string
-var REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE = (function () {
-  if (/./[REPLACE]) {
-    return /./[REPLACE]('a', '$0') === '';
-  }
-  return false;
-})();
-
-var REPLACE_SUPPORTS_NAMED_GROUPS = !fails(function () {
-  var re = /./;
-  re.exec = function () {
-    var result = [];
-    result.groups = { a: '7' };
-    return result;
-  };
-  // eslint-disable-next-line regexp/no-useless-dollar-replacements -- false positive
-  return ''.replace(re, '$<a>') !== '7';
-});
-
-// @@replace logic
-fixRegexpWellKnownSymbolLogic('replace', function (_, nativeReplace, maybeCallNative) {
-  var UNSAFE_SUBSTITUTE = REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE ? '$' : '$0';
-
-  return [
-    // `String.prototype.replace` method
-    // https://tc39.es/ecma262/#sec-string.prototype.replace
-    function replace(searchValue, replaceValue) {
-      var O = requireObjectCoercible(this);
-      var replacer = searchValue == undefined ? undefined : getMethod(searchValue, REPLACE);
-      return replacer
-        ? functionCall(replacer, searchValue, O, replaceValue)
-        : functionCall(nativeReplace, toString_1(O), searchValue, replaceValue);
-    },
-    // `RegExp.prototype[@@replace]` method
-    // https://tc39.es/ecma262/#sec-regexp.prototype-@@replace
-    function (string, replaceValue) {
-      var rx = anObject(this);
-      var S = toString_1(string);
-
-      if (
-        typeof replaceValue == 'string' &&
-        stringIndexOf(replaceValue, UNSAFE_SUBSTITUTE) === -1 &&
-        stringIndexOf(replaceValue, '$<') === -1
-      ) {
-        var res = maybeCallNative(nativeReplace, rx, S, replaceValue);
-        if (res.done) return res.value;
-      }
-
-      var functionalReplace = isCallable(replaceValue);
-      if (!functionalReplace) replaceValue = toString_1(replaceValue);
-
-      var global = rx.global;
-      if (global) {
-        var fullUnicode = rx.unicode;
-        rx.lastIndex = 0;
-      }
-      var results = [];
-      while (true) {
-        var result = regexpExecAbstract(rx, S);
-        if (result === null) break;
-
-        push$5(results, result);
-        if (!global) break;
-
-        var matchStr = toString_1(result[0]);
-        if (matchStr === '') rx.lastIndex = advanceStringIndex(S, toLength(rx.lastIndex), fullUnicode);
-      }
-
-      var accumulatedResult = '';
-      var nextSourcePosition = 0;
-      for (var i = 0; i < results.length; i++) {
-        result = results[i];
-
-        var matched = toString_1(result[0]);
-        var position = max$4(min$3(toIntegerOrInfinity(result.index), S.length), 0);
-        var captures = [];
-        // NOTE: This is equivalent to
-        //   captures = result.slice(1).map(maybeToString)
-        // but for some reason `nativeSlice.call(result, 1, result.length)` (called in
-        // the slice polyfill when slicing native arrays) "doesn't work" in safari 9 and
-        // causes a crash (https://pastebin.com/N21QzeQA) when trying to debug it.
-        for (var j = 1; j < result.length; j++) push$5(captures, maybeToString(result[j]));
-        var namedCaptures = result.groups;
-        if (functionalReplace) {
-          var replacerArgs = concat$1([matched], captures, position, S);
-          if (namedCaptures !== undefined) push$5(replacerArgs, namedCaptures);
-          var replacement = toString_1(functionApply(replaceValue, undefined, replacerArgs));
-        } else {
-          replacement = getSubstitution(matched, S, position, captures, namedCaptures, replaceValue);
-        }
-        if (position >= nextSourcePosition) {
-          accumulatedResult += stringSlice$4(S, nextSourcePosition, position) + replacement;
-          nextSourcePosition = position + matched.length;
-        }
-      }
-      return accumulatedResult + stringSlice$4(S, nextSourcePosition);
-    }
-  ];
-}, !REPLACE_SUPPORTS_NAMED_GROUPS || !REPLACE_KEEPS_$0 || REGEXP_REPLACE_SUBSTITUTES_UNDEFINED_CAPTURE);
-
-/**
- * Takes a space separated string of existing CSS class names and merges
- * it with `val`, which is either another string of class names, an
- * object of booleans or an `IDeref` evaluating to either. Returns
- * updated class string.
- *
- * @remarks
- * If `val` evaluates to a string, it will be appended to `existing`.
- *
- * If `val` is an object, its keys are used as class names and their
- * values indicate if the class should be added or removed from the
- * existing set.
- *
- * @example
- * ```ts
- * mergeClasses("foo bar", { foo: false, baz: true })
- * // "bar baz"
- *
- * mergeClasses("foo bar", "baz");
- * // "baz"
- * ```
- *
- * @param existing -
- * @param val -
- */
-
-const mergeClasses = (existing, val) => {
-  val = deref(val);
-  if (val == null) return existing;
-  if (isString(val)) return existing + " " + val;
-  const classes = new Set(existing.split(" "));
-
-  for (let id in val) {
-    deref(val[id]) ? classes.add(id) : classes.delete(id);
-  }
-
-  return [...classes].join(" ");
-};
-/**
- * Takes an attrib object and optional element ID and CSS class names from Emmet-style
- * hiccup tag, then transforms and merges definitions, returns attribs.
- *
- * @param attribs -
- * @param id -
- * @param classes -
- */
-
-const mergeEmmetAttribs = (attribs, id, classes) => {
-  id && (attribs.id = id);
-  let aclass = deref(attribs.class);
-
-  if (classes) {
-    classes = classes.replace(/\./g, " ");
-    attribs.class = aclass ? mergeClasses(classes, aclass) : classes;
-  } else if (aclass) {
-    attribs.class = isString(aclass) ? aclass : mergeClasses("", aclass);
-  }
-
-  return attribs;
-};
-
-const isArray$4 = isArray$2;
-const isNotStringAndIterable$2 = isNotStringAndIterable;
-const isPlainObject$1 = isPlainObject;
-/**
- * Expands single hiccup element/component into its canonical form:
- *
- * ```
- * [tagname, {attribs}, ...children]
- * ```
- *
- * Emmet-style ID and class names in the original tagname are moved into
- * the attribs object, e.g.:
- *
- * ```
- * ["div#foo.bar.baz"] => ["div", {id: "foo", class: "bar baz"}]
- * ```
- *
- * If both Emmet-style classes AND a `class` attrib exists, the former
- * are appended to the latter:
- *
- * ```
- * ["div.bar.baz", {class: "foo"}] => ["div", {class: "foo bar baz"}]
- * ```
- *
- * Elements with `__skip` attrib enabled and no children, will have an
- * empty text child element injected.
- *
- * @param spec - single hdom component
- * @param keys -
- *
- * @internal
- */
-
-const normalizeElement = (spec, keys) => {
-  let tag = spec[0];
-  let hasAttribs = isPlainObject$1(spec[1]);
-  let match;
-  let name;
-  let attribs;
-
-  if (typeof tag !== "string" || !(match = RE_TAG.exec(tag))) {
-    illegalArgs(`${tag} is not a valid tag name`);
-  }
-
-  name = match[1]; // return orig if already normalized and satisfies key requirement
-
-  if (tag === name && hasAttribs && (!keys || spec[1].key)) {
-    return spec;
-  }
-
-  attribs = mergeEmmetAttribs(hasAttribs ? { ...spec[1]
-  } : {}, match[2], match[3]);
-  return attribs.__skip && spec.length < 3 ? [name, attribs] : [name, attribs, ...spec.slice(hasAttribs ? 2 : 1)];
-};
-/**
- * See {@link HDOMImplementation} interface for further details.
- *
- * @param opts - hdom config options
- * @param tree - component tree
- */
-
-const normalizeTree = (opts, tree) => _normalizeTree(tree, opts, opts.ctx, [0], opts.keys !== false, opts.span !== false);
-
-const _normalizeTree = (tree, opts, ctx, path, keys, span) => {
-  if (tree == null) {
-    return;
-  }
-
-  if (isArray$4(tree)) {
-    if (tree.length === 0) {
-      return;
-    }
-
-    let norm,
-        nattribs = tree[1],
-        impl; // if available, use branch-local normalize implementation
-
-    if (nattribs && (impl = nattribs.__impl) && (impl = impl.normalizeTree)) {
-      return impl(opts, tree);
-    }
-
-    const tag = tree[0]; // use result of function call
-    // pass ctx as first arg and remaining array elements as rest args
-
-    if (typeof tag === "function") {
-      return _normalizeTree(tag.apply(null, [ctx, ...tree.slice(1)]), opts, ctx, path, keys, span);
-    } // component object w/ life cycle methods
-    // (render() is the only required hook)
-
-
-    if (typeof tag.render === "function") {
-      const args = [ctx, ...tree.slice(1)];
-      norm = _normalizeTree(tag.render.apply(tag, args), opts, ctx, path, keys, span);
-
-      if (isArray$4(norm)) {
-        norm.__this = tag;
-        norm.__init = tag.init;
-        norm.__release = tag.release;
-        norm.__args = args;
-      }
-
-      return norm;
-    }
-
-    norm = normalizeElement(tree, keys);
-    nattribs = norm[1];
-
-    if (nattribs.__normalize === false) {
-      return norm;
-    }
-
-    if (keys && nattribs.key === undefined) {
-      nattribs.key = path.join("-");
-    }
-
-    return norm.length > 2 ? normalizeChildren(norm, nattribs, opts, ctx, path, keys, span) : norm;
-  }
-
-  return typeof tree === "function" ? _normalizeTree(tree(ctx), opts, ctx, path, keys, span) : typeof tree.toHiccup === "function" ? _normalizeTree(tree.toHiccup(opts.ctx), opts, ctx, path, keys, span) : typeof tree.deref === "function" ? _normalizeTree(tree.deref(), opts, ctx, path, keys, span) : span ? ["span", keys ? {
-    key: path.join("-")
-  } : {}, tree.toString()] : tree.toString();
-};
-
-const normalizeChildren = (norm, nattribs, opts, ctx, path, keys, span) => {
-  const tag = norm[0];
-  const res = [tag, nattribs];
-  span = span && !NO_SPANS[tag];
-
-  for (let i = 2, j = 2, k = 0, n = norm.length; i < n; i++) {
-    let el = norm[i];
-
-    if (el != null) {
-      const isarray = isArray$4(el);
-
-      if (isarray && isArray$4(el[0]) || !isarray && isNotStringAndIterable$2(el)) {
-        for (let c of el) {
-          c = _normalizeTree(c, opts, ctx, path.concat(k), keys, span);
-
-          if (c !== undefined) {
-            res[j++] = c;
-          }
-
-          k++;
-        }
-      } else {
-        el = _normalizeTree(el, opts, ctx, path.concat(k), keys, span);
-
-        if (el !== undefined) {
-          res[j++] = el;
-        }
-
-        k++;
-      }
-    }
-  }
-
-  return res;
-};
-
-/**
- * Default target implementation to manipulate browser DOM.
- */
-
-const DEFAULT_IMPL = {
-  createTree(opts, parent, tree, child, init) {
-    return createTree(opts, this, parent, tree, child, init);
-  },
-
-  hydrateTree(opts, parent, tree, child) {
-    return hydrateTree(opts, this, parent, tree, child);
-  },
-
-  diffTree(opts, parent, prev, curr, child) {
-    diffTree(opts, this, parent, prev, curr, child);
-  },
-
-  normalizeTree,
-
-  getElementById(id) {
-    return document.getElementById(id);
-  },
-
-  getChild,
-  createElement,
-  createTextElement,
-
-  replaceChild(opts, parent, child, tree, init) {
-    replaceChild(opts, this, parent, child, tree, init);
-  },
-
-  removeChild,
-  setContent,
-  removeAttribs,
-  setAttrib
-};
-
-var LogLevel;
-
-(function (LogLevel) {
-  LogLevel[LogLevel["FINE"] = 0] = "FINE";
-  LogLevel[LogLevel["DEBUG"] = 1] = "DEBUG";
-  LogLevel[LogLevel["INFO"] = 2] = "INFO";
-  LogLevel[LogLevel["WARN"] = 3] = "WARN";
-  LogLevel[LogLevel["SEVERE"] = 4] = "SEVERE";
-  LogLevel[LogLevel["NONE"] = 5] = "NONE";
-})(LogLevel || (LogLevel = {}));
-
-/**
- * No-op {@link ILogger} implementation, used as default logger for most
- * packages (where used).
- */
-
-const NULL_LOGGER = Object.freeze({
-  level: LogLevel.NONE,
-
-  fine() {},
-
-  debug() {},
-
-  info() {},
-
-  warn() {},
-
-  severe() {}
-
-});
-
-let LOGGER = NULL_LOGGER;
-const setLogger = logger => LOGGER = logger;
-
-const implementsFunction = (x, fn) => x != null && typeof x[fn] === "function";
-
-/**
- * Takes an arbitrary `ctx` object and array of `keys`. Attempts to call
- * `.deref()` on all given keys' values and stores result values instead
- * of original. Returns updated copy of `ctx` or original if `ctx` is
- * `null` or no keys were given.
- *
- * @param ctx - user context object
- * @param keys - keys to deref
- *
- * @internal
- */
-
-const derefContext = (ctx, keys) => {
-  if (ctx == null || !keys || !keys.length) return ctx;
-  const res = { ...ctx
-  };
-
-  for (let k of keys) {
-    const v = res[k];
-    implementsFunction(v, "deref") && (res[k] = v.deref());
-  }
-
-  return res;
-};
-
-const resolveRoot = (root, impl) => isString(root) ? impl.getElementById(root) : root;
-
-/**
- * One-off hdom tree conversion & target DOM application. Takes same
- * options as {@link start}, but performs no diffing and only creates or
- * hydrates target once. The given tree is first normalized and if
- * result is `null` or `undefined` no further action will be taken.
- *
- * @param tree - component tree
- * @param opts - hdom config options
- * @param impl - hdom implementation
- */
-
-const renderOnce = (tree, opts = {}, impl = DEFAULT_IMPL) => {
-  opts = {
-    root: "app",
-    ...opts
-  };
-  opts.ctx = derefContext(opts.ctx, opts.autoDerefKeys);
-  const root = resolveRoot(opts.root, impl);
-  tree = impl.normalizeTree(opts, tree);
-  if (!tree) return;
-  opts.hydrate ? impl.hydrateTree(opts, root, tree) : impl.createTree(opts, root, tree);
-};
-
-/**
- * Takes an hiccup tree (array, function or component object w/ life
- * cycle methods) and an optional object of DOM update options. Starts
- * RAF update loop, in each iteration first normalizing given tree, then
- * computing diff to previous frame's tree and applying any changes to
- * the real DOM. The `ctx` option can be used for passing arbitrary
- * config data or state down into the hiccup component tree. Any
- * embedded component function in the tree will receive this context
- * object (shallow copy) as first argument, as will life cycle methods
- * in component objects. If the `autoDerefKeys` option is given,
- * attempts to auto-expand/deref the given keys in the user supplied
- * context object (`ctx` option) prior to *each* tree normalization. All
- * of these values should implement the {@link @thi.ng/api#IDeref}
- * interface (e.g. atoms, cursors, views, rstreams etc.). This feature
- * can be used to define dynamic contexts linked to the main app state,
- * e.g. using derived views provided by {@link @thi.ng/atom# | @thi.ng/atom}.
- *
- * **Selective updates**: No updates will be applied if the given hiccup
- * tree is `undefined` or `null` or a root component function returns no
- * value. This way a given root function can do some state handling of
- * its own and implement fail-fast checks to determine no DOM updates
- * are necessary, save effort re-creating a new hiccup tree and request
- * skipping DOM updates via this function. In this case, the previous
- * DOM tree is kept around until the root function returns a tree again,
- * which then is diffed and applied against the previous tree kept as
- * usual. Any number of frames may be skipped this way.
- *
- * **Important:** Unless the `hydrate` option is enabled, the parent
- * element given is assumed to have NO children at the time when
- * `start()` is called. Since hdom does NOT track the real DOM, the
- * resulting changes will result in potentially undefined behavior if
- * the parent element wasn't empty. Likewise, if `hydrate` is enabled,
- * it is assumed that an equivalent DOM (minus listeners) already exists
- * (i.e. generated via SSR) when `start()` is called. Any other
- * discrepancies between the pre-existing DOM and the hdom trees will
- * cause undefined behavior.
- *
- * Returns a function, which when called, immediately cancels the update
- * loop.
- *
- * @param tree - hiccup DOM tree
- * @param opts - options
- * @param impl - hdom target implementation
- */
-
-const start = (tree, opts = {}, impl = DEFAULT_IMPL) => {
-  const _opts = {
-    root: "app",
-    ...opts
-  };
-  let prev = [];
-  let isActive = true;
-  const root = resolveRoot(_opts.root, impl);
-
-  const update = () => {
-    if (isActive) {
-      _opts.ctx = derefContext(opts.ctx, _opts.autoDerefKeys);
-      const curr = impl.normalizeTree(_opts, tree);
-
-      if (curr != null) {
-        if (_opts.hydrate) {
-          impl.hydrateTree(_opts, root, curr);
-          _opts.hydrate = false;
-        } else {
-          impl.diffTree(_opts, root, prev, curr);
-        }
-
-        prev = curr;
-      } // check again in case one of the components called cancel
-
-
-      isActive && requestAnimationFrame(update);
-    }
-  };
-
-  requestAnimationFrame(update);
-  return () => isActive = false;
-};
-
-export { DEFAULT_IMPL, LOGGER, addChild, clearDOM, cloneWithNewAttribs, createElement, createTextElement, createTree, diffAttributes, diffTree, equiv$1 as equiv, getChild, hydrateTree, normalizeElement, normalizeTree, releaseTree, removeAttribs, removeChild, removeListener, renderOnce, replaceChild, resolveRoot, setAttrib, setAttribs, setContent, setListener, setLogger, setStyle, start, updateValueAttrib };
